% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr}
\alias{alr}
\alias{alr.default}
\alias{alr.hyperSpec}
\title{[!] Additive logratio transformation}
\usage{
alr(x, ...)

\method{alr}{default}(x, ivar = ncol(x), ...)

\method{alr}{hyperSpec}(x, ivar = ncol(x$spc), ...)
}
\arguments{
\item{x}{A composition, not necessarily closed. (A numeric data frame or matrix.)}

\item{...}{further parameters to \code{\link[compositions]{alr}}.}

\item{ivar}{The number of column in \code{x} to be used as denominator variable.
By default the last column is taken.}
}
\value{
Returns the transformed data matrix with one dimension less than
     \code{x}. Thus for \code{hyperSpec} object the last value of
     x$.wavelength is removed.
}
\description{
Compute the additive log ratio transform of a (dataset of) composition(s).
 Default method is function
\code{\link[compositions]{alr}} from package \pkg{compositions}.
}
\examples{
data(Spectra2)
Spectra <- Spectra2[,,300~600]
alr(Spectra)

ind <- wl2i(Spectra, 550) # number of column at 550 nm.
alr(Spectra, ind)

plot(Spectra)
plot(alr(Spectra))
plot(alr(Spectra, ind))

}
\seealso{
\code{\link[compositions]{alr}}

Other row-wise transformations: \code{\link{apt}},
  \code{\link{clr}}, \code{\link{cpt}}, \code{\link{ilr}},
  \code{\link{ipt}}
}
\concept{row-wise transformations}
