% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.R
\name{apt}
\alias{apt}
\alias{apt.default}
\alias{apt.hyperSpec}
\title{[!]Additive planar transform}
\usage{
apt(x, ...)

\method{apt}{default}(x, ...)

\method{apt}{hyperSpec}(x, ...)
}
\arguments{
\item{x}{A numeric data frame or matrix.}

\item{...}{further parameters to \code{\link[compositions]{apt}}.}
}
\value{
Returns the transformed data matrix with the same dimension as
        \code{x}.
}
\description{
Compute the additive planar transform of a (dataset of) compositions.
Default method is function \code{\link[compositions]{apt}} from package \pkg{compositions}.
}
\examples{
apt(Spectra2)

plot(Spectra2)
plot(apt(Spectra2))
}
\seealso{
\code{\link[compositions]{apt}}

Other row-wise transformations: 
\code{\link{alr}()},
\code{\link{clr}()},
\code{\link{cpt}()},
\code{\link{ilr}()},
\code{\link{ipt}()}
}
\concept{row-wise transformations}
