% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\alias{binning.default}
\alias{binning.hyperSpec}
\title{[!] Signal binning}
\usage{
binning(X, bins, bin.size)

\method{binning}{default}(X, ...)

\method{binning}{hyperSpec}(X, ...)
}
\arguments{
\item{X}{a numeric \code{data.frame}, \code{matrix} or \code{vector} to process.}

\item{bins}{the number of bins.}

\item{bin.size}{the desired size of the bins.}
}
\value{
A \code{matrix}, \code{vector} or \code{hyperSpec} object with
        the filtered signal(s).
}
\description{
Compute average values of a signal in pre-determined bins (col-wise subsets).
 The bin size can be determined either directly or by specifying the number
 of bins. Sometimes called boxcar transformation in signal processing
}
\details{
Method \code{binning.default} is the function \code{\link[prospectr]{binning}}
from package \pkg{prospectr}. \cr

Method \code{binning.default} is the function \code{\link[prospectr]{binning}}
from package \pkg{prospectr}. \cr
}
\note{
Find examples in \code{\link[prospectr]{binning}}.
}
\seealso{
\code{\link[prospectr]{binning}}

Other methods for \code{\link[=hyperSpec-class]{hyperSpec}} in \pkg{spHelper}: 
\code{\link{gapDer}()}

Other \pkg{prospectr} function for \pkg{hyperSpec}: 
\code{\link{gapDer}()}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Antoine Stevens & Leonardo Ramirez-Lopez (function \code{prospectr::binning})\cr
Vilmantas Gegzna (method \code{binning.hyperSpec})
}
\concept{\pkg{prospectr} function for \pkg{hyperSpec}}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
\concept{methods for \code{\link[=hyperSpec-class]{hyperSpec}} in \pkg{spHelper}}
