% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci_mean.R
\name{boot_ci_mean}
\alias{boot_ci_mean}
\alias{boot_ci_fun}
\alias{boot_ci_corr}
\title{Bootstrapped mean and its confidence interval}
\usage{
boot_ci_mean(x, conf = 0.95, R = 1000, sim = "balanced", type = c("norm"))

boot_ci_fun(
  x,
  FUN,
  conf = 0.95,
  R = 1000,
  sim = "balanced",
  type = c("norm"),
  label = as.character(match.call()$FUN)
)

boot_ci_corr(
  x,
  y = NULL,
  method = c("spearman", "kendall", "pearson")[1],
  use = "everything",
  conf = 0.95,
  R = 1000,
  sim = "balanced",
  type = c("norm"),
  label = "corr_coef"
)
}
\arguments{
\item{x}{a vector.}

\item{conf}{
    A scalar or vector containing the confidence level(s) of the required
    interval(s).
  }

\item{R}{
    The number of bootstrap replicates.  Usually this will be a single
    positive integer.  For importance resampling, some resamples may use
    one set of weights and others use a different set of weights.  In
    this case \code{R} would be a vector of integers where each
    component gives the number of resamples from each of the rows of
    weights.
  }

\item{sim}{
    A character string indicating the type of simulation required.
    Possible values are \code{"ordinary"} (the default),
    \code{"parametric"}, \code{"balanced"}, \code{"permutation"}, or
    \code{"antithetic"}.  Importance resampling is specified by
    including importance weights; the type of importance resampling must
    still be specified but may only be \code{"ordinary"} or
    \code{"balanced"} in this case.
  }

\item{type}{
    A vector of character strings representing the type of intervals
    required. The value should be any subset of the values
    \code{c("norm","basic", "stud", "perc", "bca")} or simply \code{"all"}
    which will compute all five types of intervals.
  }

\item{FUN}{a function, that takes a vector returns one number, e.g. mean, median, etc.}

\item{label}{(string) a label for function to be used as column name.}

\item{y}{a vector.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}
}
\value{
A data frame with bootstrapped mean and its confidence interval.
}
\description{
Function calculates bootstrapped mean (or other function)
and its confidence interval for a vector \code{x}.
}
\details{
\code{boot_ci_corr} calculates confidence interval for correlation
coefficient between vectors \code{x} and \code{y}.
}
\examples{

set.seed(1)
x <- rnorm(1000, mean = .5, sd = .1)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
set.seed(1)
boot_ci_mean(x)

#       ci_lower  mean     ci_upper
#  1   0.4923028 0.4988352 0.5053676

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
set.seed(1)
boot_ci_fun(x, IQR)

#     ci_lower       IQR   ci_upper
# 1   0.1307229 0.1385801 0.1486593

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
set.seed(1)

MeDiAn <- median
boot_ci_fun(x, MeDiAn, label = "m")

#     ci_lower    median ci_upper
# 1 0.4900485 0.4964676 0.502184
set.seed(1)
x <- rnorm(30)
y <- x - rnorm(30) + runif(30,-2,2)
plot(x,y)

set.seed(1)
boot_ci_corr(x, y)

#    ci_lower   corr_coef  ci_upper
#   -0.1051065  0.243604   0.6067977

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


df <- data.frame(x, y)
set.seed(1)
boot_ci_corr(df)

}
