% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Various_functions.R
\name{bru}
\alias{bru}
\title{Print a line made of symbols.}
\usage{
bru(symbol = "=", len = 60, after = {     if (print)          1     else 0
  }, before = 0, print = TRUE)
}
\arguments{
\item{symbol}{- desired symbol or sequence of symbols. Default is "="}

\item{len}{- length of a line \cr default is 60.}

\item{after}{- number of new (empty) lines/rows to be added afterwards.
\code{0} means that following text continues in the same row.
Default is \cr\code{if (print==TRUE) 1 else 0}.}

\item{before}{- number of peceeding emptyrows. Default is 0.}

\item{print}{- if \code{TRUE} (defailt) - print,
if \code{FALSE} - return as a string}
}
\description{
Print a line made of symbols.
}
\examples{

bru
bru("-")
bru("= ")

}

