% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_class_perform.R
\name{calculate_performance}
\alias{calculate_performance}
\title{Calculate best berformance measure}
\usage{
calculate_performance(pred, measure)
}
\arguments{
\item{pred}{object of class "prediction" from package ROCR}

\item{measure}{one of the following performance measures
\code{c("auc","j","bac","kappa")}}
}
\value{
best value of selected performance measure
}
\description{
Function that calculates the best performance measure
}
