% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc__to_remove.R
\name{class_add}
\alias{class_add}
\title{Add an additional S3 class label to an object}
\usage{
class_add(x, new_class)
}
\arguments{
\item{x}{An object to modify.}

\item{new_class}{(\code{character})\cr A name of a new class.
May be a vector of names.}
}
\description{
Add an additional S3 class label to an object
}
\examples{
class_add(list("ok"), "ok_list")

}
