% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R
\name{clc}
\alias{clc}
\title{Clear console (Matlab style)}
\usage{
clc()
}
\description{
Clears console. The same as \code{CTRL + L}in RStudio.
}
\seealso{
Other family \pkg{spHelper} utilities: \code{\link{make.nameswin}},
  \code{\link{st01}}
}

