% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\alias{clr.default}
\alias{clr.hyperSpec}
\title{[!] Centered logratio transformation}
\usage{
clr(x, ...)

\method{clr}{default}(x, ...)

\method{clr}{hyperSpec}(x, ...)
}
\arguments{
\item{x}{A numeric data frame or matrix.}

\item{...}{further parameters to \code{\link[compositions]{clr}}.}
}
\value{
Returns the transformed data matrix with the same dimension as
        \code{x}.
}
\description{
A data transformation according to the centered logratio
transformation is done. default method is function
\code{\link[compositions]{clr}} from package \pkg{compositions}.
}
\examples{
clr(Spectra2)

plot(Spectra2)
plot(clr(Spectra2))
}
\seealso{
\code{\link[compositions]{clr}}

Other row-wise transformations: 
\code{\link{alr}()},
\code{\link{apt}()},
\code{\link{cpt}()},
\code{\link{ilr}()},
\code{\link{ipt}()}
}
\concept{row-wise transformations}
