% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRatios.R
\name{colRatios}
\alias{colRatios}
\title{[!] Calculate ratios between values of every column pair.}
\usage{
colRatios(x, col.names = c("full", "short", "rm.letters"))
}
\arguments{
\item{x}{A matrix-like data.}

\item{col.names}{One of the fpollowing: c("full", "short", "rm.letters").
\cr\cr [NOT UPDATED DESCRIPTION] If \code{FALSE}, whole names of compared columns
       will be used for column of ratios, e.g. \code{"Column 1/ Column 2"}. If
       \code{TRUE} (default), shorter version of column name will be used,
        where parts of name which are common for both columns will not be
        repeated, e.g. \code{"Column 1/2"}.}
}
\value{
A matrix with ratios between values of every column pair of the matrix \code{x}.
}
\description{
[!] Calculate ratios between values of every column pair.
}
\examples{
x <- Scores2[[]]
r1 <- colRatios(x)
r2 <- colRatios(x, col.names = "short")
r3 <- colRatios(x, col.names = "rm.letters")

head(x)
head(r1)

colnames(r1)
colnames(r2)
colnames(r3)

}
