% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_vec2mat.R
\name{corr_vec2mat}
\alias{corr_vec2mat}
\title{[+] Construct a correlation matrix from a coefficients' vector}
\usage{
corr_vec2mat(vec)
}
\arguments{
\item{vec}{A Vector with known correlation coefficients.}
}
\value{
A constructed correlation matrix.
}
\description{
Construct a correlation matrix from a vector vector of coefficients, that
 will be placed in the lower triangle of a matrix (as illustrated in example
 Example 1B) and then mirrored symetrically to the upper triangle.
}
\details{
@details The number of elements in the vector should be equal to the number
 of elements in lower triangle of a n*n matrix. If there's insufficient
 number of elements in vector, hen only first m melements will be used to
 construct(n-1)*(n-1) matrix.
}
\examples{

# ------------------------------------------------------------
# Example 1A: Vector into a matrix

vector = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6)
corr_vec2mat(vector)

# ------------------------------------------------------------

\donttest{
\dontrun{

 # Example 1B: Vector is transformed to a matrix by filling
 # it column-wise:

  ## vector = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6)
  ##
  ##  matrix =
  ##       [,1] [,2] [,3] [,4]
  ##  [1,]   .     .   .    .
  ##  [2,]  0.1    .   .    .
  ##  [3,]  0.2  0.4   .    .
  ##  [4,]  0.3  0.5  0.6   .
  ##
  ## In this example only the matrix elements of interest are shown.

 # -----------------------------------------------------------------
 # Example 2: inappropriate number of coefficients - warning appears

 corr_vec2mat(vector[1:5])
  ## Warning message:
  ## Only first 3 coefficient(s) out of 5 will be used to construct symmetric matrix with 3 rows.

 # ------------------------------------------------------------
 # Example 3: ERROR appears - values of coeffs must be between [-1;1]

 corr_vec2mat(1:5)
  ## Error in corr_vec2mat(1:5) :
  ## All values in input vector must be between [-1 and 1]
}}

}
\author{
Vilmantas Gegzna
}
\seealso{
Other matrix operations: \code{\link{ind.matrix}},
  \code{\link{nDiag2nTri}}, \code{\link{nTri2nDiag}},
  \code{\link{new.matrix}}, \code{\link{sortMaxOnDiag}},
  \code{\link{which.in}}, \code{\link{which.max.all}},
  \code{\link{which.max.perRow}}

Other simmulation functions: \code{\link{simCorrVars}},
  \code{\link{simSpectra}}
}

