% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrVec2Mat.R
\name{corrVec2Mat}
\alias{corrVec2Mat}
\title{[!!!] Construct a correlation matrix}
\usage{
corrVec2Mat(vec)
}
\arguments{
\item{vec}{A Vector with known correlation coefficients to construct
a correlation matrix.}
}
\value{
A constructed correlation matrix.
}
\description{
[!!!] Construct a correlation matrix
}
\examples{

# ------------------------------------------------------------
# Example 1A: Vector into a matrix

vector = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6)
corrVec2Mat(vector)

# ------------------------------------------------------------
# Example 1B: Vector is transformed to a matrix by filling
#  it column-wise:

 # vector = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6)
 #
 #  matrix =
 #       [,1] [,2] [,3] [,4]
 #  [1,]   .     .   .    .
 #  [2,]  0.1    .   .    .
 #  [3,]  0.2  0.4   .    .
 #  [4,]  0.3  0.5  0.6   .
 #
 # In this example only the matrix elements of interest are shown.

# ------------------------------------------------------------
# Example 2: inappropriate number of coefficients - warning appears

corrVec2Mat(vector[1:5])
 ## Warning message:
 ## Only first 3 coefficient(s) out of 5 will be used to construct symmetric matrix with 3 rows.

# ------------------------------------------------------------
# Example 3: ERROR appears - coeefs must be between [-1;1]

corrVec2Mat(1:5)
 ## Error in corrVec2Mat(1:5) :
 ## All values in input vector must be between [-1 and 1]

}
\seealso{
\link{nTri2nRow}, \link{nRow2nTri}
}

