% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_spectra.R
\name{count_spectra}
\alias{count_spectra}
\title{Summary statistics of factor variable in a hypecSpec object by ID}
\usage{
count_spectra(
  OBJ,
  Var = NULL,
  ID = "ID",
  decimals = 1,
  include_na = TRUE,
  na_level = "(Missing)",
  na.rm = TRUE,
  total = FALSE
)
}
\arguments{
\item{OBJ}{hyperSpec object}

\item{Var}{(string) variable name of factor variable}

\item{ID}{(string) variable name of ID variable}

\item{decimals}{(integer) number of decimals to round to calculating the percentages.
Default is 2.}

\item{include_na}{(logical) If \code{TRUE}, \code{NA} values are included.}

\item{na_level}{(character) Name for the level, which represents \code{NA} value.}

\item{na.rm}{(logical) used by function \code{sum()}. Default is \code{TRUE}.}

\item{FALSE}{(logical) Add row with "total".}
}
\value{
a data frame with counts and percentages by group and ID
}
\description{
Summary statistics of factor variable in a hypecSpec object by ID
}
\examples{

count_spectra(Spectra2, "class", ID = "gr")
count_spectra(OBJ = Spectra2, Var = "class", ID = "gr")

}
\seealso{
Other count spectra functions: 
\code{\link{has_enough_IDs}()},
\code{\link{has_too_few_IDs}()}
}
\concept{count spectra functions}
