% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt}
\alias{cpt}
\alias{cpt.default}
\alias{cpt.hyperSpec}
\title{[!] Centered planar transform}
\usage{
cpt(x, ...)

\method{cpt}{default}(x, ...)

\method{cpt}{hyperSpec}(x, ...)
}
\arguments{
\item{x}{A numeric data frame or matrix.}

\item{...}{further parameters to \code{\link[compositions]{cpt}}.}
}
\value{
Returns the transformed data matrix with the same dimension as
        \code{x}.
}
\description{
A data transformation according to the centered planar
transformation is done. Default method is function
\code{\link[compositions]{cpt}} from package \pkg{compositions}.
}
\examples{
cpt(Spectra2)

plot(Spectra2)
plot(cpt(Spectra2))
}
\seealso{
\code{\link[compositions]{cpt}}

Other row-wise transformations: \code{\link{alr}},
  \code{\link{apt}}, \code{\link{clr}}, \code{\link{ilr}},
  \code{\link{ipt}}
}
\concept{row-wise transformations}
