% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpFun.R
\name{createFolds_stratified}
\alias{createFolds_stratified}
\title{Wraper of the function \code{\link{createFolds_strat}}}
\usage{
createFolds_stratified(ID = NULL, groups = NULL, k = 5,
  returnTrain = TRUE)
}
\arguments{
\item{ID}{- vector}

\item{groups}{- vector}

\item{k}{- number of folds k = 5}

\item{returnTrain}{- returnTrain=TRUE}
}
\value{
Result is the same as produced by function \code{\link{createFolds}}
}
\description{
Wraper of the function \code{\link{createFolds_strat}}
}
\examples{
createFolds_stratified

}

