% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_DEPRECATED.R, R/5_DEPRECATED_createFoldsSB.R
\name{createFolds2}
\alias{createFolds2}
\alias{stratifiedFolds}
\alias{createFoldsBS}
\title{DEPRECATED!}
\usage{
createFolds2(data = NULL, gr = NULL, ID = NULL, k = 5, returnTrain = TRUE, ...)

stratifiedFolds(
  data = NULL,
  gr = NULL,
  ID = NULL,
  k = 5,
  returnTrain = TRUE,
  ...
)

createFoldsBS(
  data = NULL,
  stratify_by = NULL,
  block_by = NULL,
  k = 5,
  returnTrain = TRUE,
  times = 1,
  seeds = NULL,
  gr = stratify_by,
  ID = block_by
)
}
\arguments{
\item{data}{A data frame, that contains variables which names are denoted
by arguments \code{ID} and by \code{gr}.}

\item{k}{(integer) A number of folds, default \code{k = 5}.}

\item{returnTrain}{(logical) If \code{TRUE}, returns indices of variables in
training set. If \code{FALSE}, returns indices of
variables in test set.}

\item{stratify_by, gr}{A vector or a name of factor variable in \code{data}, which levels
will be used for \emph{stratification}. E.g., vector with
medical groups.}

\item{block_by, ID}{A vector or a name of variable in \code{data}, that contains
identification codes/numbers (ID). These codes will be used for blocking.}

\item{times}{(integer) number of repetitions for repeated cross-vatidtion.}

\item{seeds}{(vector of integers | \code{NULL}) Seeds for random number
            generator for each repetition.\cr
            If \code{seeds = NULL} random seeds are generated.\cr
            If number of repetitions is
            greater than number of provided seeds, random seeds are
            generated and added to the provided ones. The first seed will
            be used to ensure reproducibility of the randomly generated seeds.\cr

             (See \code{\link[base]{set.seed}} for more information about
             random number generation).}
}
\value{
A list of folds. In each fold there are indices observations.
        The structure of outpus is the similar to one created with
        \code{\link[caret]{createFolds}()}.
}
\description{
DEPRECATED!\cr
Create folds with blocking and stratification (BS)
  for (repeated) k-fold cross-validation
Function randomly divides observations into folds that are used for (repeated)
k-fold cross-validation. In these folds observations are:
\enumerate{
     \item \bold{blocked} by values in variable \code{ID} (i.e. observations
       with the same "ID" are treated as one unit (a block) and are always
        in the same fold);
     \item \bold{stratified} by levels of factor variable \code{gr} (the proportions of
     these grouped units of observations per each group (level) are kept aproximately
     constant throughout all folds).
 }
}
\note{
If \code{k} is such big, that some folds have no observations of a certain group
      (i.e. level in \code{gr}), an error is returned. In that case smaller value of
       \code{k} is recommended. \cr \cr
        \code{createFolds2}, \code{stratifiedFolds} is a wrapper of
        \code{createFoldsBS}.
}
\seealso{
\code{\link[caret]{createFolds}}\cr
Test if folds are blocked and stratified \code{\link{cvo_test_bs}}
}
\author{
Vilmantas Gegzna
}
