% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_DEPRECATED.R
\name{cvo_n_folds}
\alias{cvo_n_folds}
\title{DEPRECATED!}
\usage{
cvo_n_folds(cvo)
}
\arguments{
\item{cvo}{A cross-validation object}
}
\description{
DEPRECATED! Use \code{\link{cvo_count_folds}} instead.\cr

Function \code{cvo_n_folds}
returns total number of folds in a \code{cvo} object.
}
\seealso{
\code{\link{cvo_count_folds}}
}
