% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_datasets.R
\docType{data}
\name{Datasets-Loadings}
\alias{Datasets-Loadings}
\alias{Loadings}
\alias{Loadings3}
\alias{Loadings4}
\title{[!] DataSets: Loadings}
\format{An object of class \code{hyperSpec} of dimension 5 x 4 x 501.}
\usage{
Loadings

Loadings3

Loadings4
}
\description{
A dataset containing Loadings ...
}
\keyword{datasets}

