% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_datasets.R
\docType{data}
\name{Datasets-Scores}
\alias{Datasets-Scores}
\alias{Scores}
\alias{Scores3}
\alias{Scores4}
\title{[!] DataSets: Scores}
\format{An object of class \code{hyperSpec} of dimension 150 x 3 x 5.}
\usage{
Scores

Scores3

Scores4
}
\description{
A dataset containing Scores ...
}
\keyword{datasets}

