% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_datasets.R
\docType{data}
\name{Datasets-Spectra}
\alias{Datasets-Spectra}
\alias{Spectra}
\alias{Spectra3}
\alias{Spectra4}
\title{[!] DataSets: Spectra}
\format{An object of class \code{hyperSpec} of dimension 150 x 3 x 501.}
\usage{
Spectra

Spectra3

Spectra4
}
\description{
A dataset containing spevtroscopic data ...
}
\keyword{datasets}

