% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doALS2.R
\name{doALS2}
\alias{doALS2}
\title{[!] Wrapper function for alternating least squares multivariate curve resolution (MCR-ALS)}
\usage{
doALS2(Xl, PureS, maxiter = 100, normS = 0.5, uniC = FALSE,
  uniS = FALSE, nonnegS = TRUE, nonnegC = TRUE, optS1st = FALSE,
  baseline = FALSE, closureC = list(), ...)
}
\arguments{
\item{Xl}{a list of (numerical) data matrices. Missing values are not
    allowed.}

\item{PureS}{Initial estimates of pure spectral components.}

\item{maxiter}{The maximum number of iterations to perform (where an
    iteration is optimization of either \code{AList} and \code{C})}

\item{normS}{numeric indicating whether the spectra are normalized; if
    \code{normS>0}, the spectra are normalized.  If \code{normS==1} the
    maximum of the spectrum of each component is constrained to be equal
    to one; if \code{normS > 0 && normS!=1} then the norm of the
    spectrum of each component is constrained to be equal to one.  
    }

\item{uniC}{logical indicating whether unimodality constraints should be
    applied to the columns of \code{C}}

\item{uniS}{logical indicating whether unimodality constraints should be
    applied to the columns of \code{S}}

\item{nonnegS}{logical indicating whether the components (columns) of
    the matrix \code{S} should be constrained to non-negative values}

\item{nonnegC}{logical indicating whether the components (columns) of
    the matrix \code{C} should be constrained to non-negative values}

\item{optS1st}{logical indicating whether the first constrained least
    squares regression should estimate \code{S} or \code{CList}.}

\item{baseline}{logical indicating whether a baseline component is
    present; if \code{baseline=TRUE} then this component is exempt from 
    constraints unimodality or non-negativity}

\item{closureC}{list; if the length is zero, then
  no closure constraints are applied.  If the length is not zero, it
  should be equal to the number of datasets in the analysis, and contain
  numeric vectors consisting of the desired value of
  the sum of each row of the concentration matrix.
}

\item{...}{Other parametars to be passed to function \link[ALS]{als}.}
}
\value{
The same as in \link[alsace]{doALS}.
}
\description{
This function is the same as \link[alsace]{doALS} from package \pkg{doALS}
(version 1.6.0), just with more possible input parameters, which are passed to
\link[ALS]{als} from package \pkg{ALS}.
}
\examples{

# /NO examples YET/

}
\seealso{
\link[alsace]{doALS}, \link[ALS]{als}
}

