% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCol.R
\name{dropCol}
\alias{dropCol}
\title{[!] Remove colums of hyperSpec object with all NA values}
\usage{
dropCol(sp)
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}
}
\value{
Object without columns that contain all NA values.
}
\description{
[!] Remove colums of hyperSpec object with all NA values
}
\examples{

 dropCol(sp)

}

