% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmExpr.R
\name{expr2text}
\alias{expr2text}
\title{[.] Convert `expression` and `call` to text and remove quotes}
\usage{
expr2text(x)
}
\arguments{
\item{x}{`Call` object or `expression`}
}
\description{
[.] Convert `expression` and `call` to text and remove quotes
}
\seealso{
Other \pkg{spHelper} utilities: 
\code{\link{IQR_outliers}()},
\code{\link{mad_outliers}()},
\code{\link{rmExpr}()},
\code{\link{sd_outliers}()},
\code{\link{subt}()},
\code{\link{unGroup}()},
\code{\link{uncall}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} utilities}
