% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapDer.R
\name{gapDer}
\alias{gapDer}
\alias{gapDer.default}
\alias{gapDer.hyperSpec}
\title{[!] Gap-Segment Derivative}
\usage{
gapDer(X, m = 1, w = 1, s = 1, delta.wav)

\method{gapDer}{default}(X, m = 1, w = 1, s = 1, ...)

\method{gapDer}{hyperSpec}(X, m = 1, w = 1, s = 1, ...)
}
\arguments{
\item{X}{a numeric \code{matrix} , \code{data.frame} or \code{vector} to transform.}

\item{m}{the order of the derivative, between 1 and 4 (default = 1).}

\item{w}{the filter length (should be odd and >=1), i.e. the spacing between points over which the derivative is computed.}

\item{s}{sthe egment size, i.e. the range over which the points are averaged (default = 1, i.e. no smoothing corresponding to 'Norris' Gap Derivative).}

\item{delta.wav}{the sampling interval (or band spacing).}
}
\value{
A \code{matrix}, \code{vector} or \code{hyperSpec} object with
        the filtered signal(s).
}
\description{
Gap-Segment derivatives of a data \code{matrix}, \code{data.frame},
\code{vector} or \code{hyperSpec} object.
}
\details{
Method \code{gapDer.default} is the function \code{\link[prospectr]{gapDer}}
from package \pkg{prospectr}. \cr

Method \code{gapDer.default} is the function \code{\link[prospectr]{gapDer}}
from package \pkg{prospectr}. \cr

The sampling interval specified with the \code{delta.wav} argument is used
for scaling and get numerically correct derivatives. \cr

The convolution function is written in C++/Rcpp for faster computations.
}
\note{
Find examples in \code{\link[prospectr]{gapDer}}.
}
\references{
Hopkins (2002). NIR News 14(5), 10.
}
\seealso{
\code{\link[prospectr]{gapDer}}

Other methods for \code{\link[=hyperSpec-class]{hyperSpec}} in \pkg{spHelper}: 
\code{\link{binning}()}

Other \pkg{prospectr} function for \pkg{hyperSpec}: 
\code{\link{binning}()}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Antoine Stevens (function \code{prospectr::gapDer})\cr
Vilmantas Gegzna (method \code{gapDer.hyperSpec})
}
\concept{\pkg{prospectr} function for \pkg{hyperSpec}}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
\concept{methods for \code{\link[=hyperSpec-class]{hyperSpec}} in \pkg{spHelper}}
