% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpFun.R
\name{getScores}
\alias{getScores}
\title{Calculate component amplitudes (a.k.a scotes) by matrix multiplication}
\usage{
getScores(y, k_sp)
}
\arguments{
\item{y}{- matrix with experimental spectra}

\item{k_sp}{- matrix with components' spectra}
}
\value{
k_amp - amplitudes of the components
}
\description{
$k_amp = y * k_sp * inv(k_sp' * k_sp)$;
 formula is taken from [1] and adapted
 [1] M. Brydegaard et al. IEEE Photonics J 2011:3(3);406-21.
}
\examples{
# e.g.:
    y = Object
    k_sp = loadings

getScores(y, k_sp)

}

