% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarValues.R
\name{getVarValues}
\alias{getVarValues}
\title{[!] Parse a call of a function and return approriate values for a variable.}
\usage{
getVarValues(variable, data, CALL, env = parent.frame())
}
\arguments{
\item{variable}{A name of a variable (with or without quotes).}

\item{data}{A name of a data frame (with or without quotes).}

\item{CALL}{(\code{\link[=call-class]{Call}}) to be parsed.}

\item{env}{Envoronmen to search for "free" variables in.}
}
\value{
#' @export
}
\description{
Parse dataframe \code{data}, variable \code{variable} and a call
\code{CALL} of a function and return approriate values of the
\code{variable}. \cr If possible, return \code{data[,variable]} \cr
Otherwise return \code{variable}.
}
\examples{

# A Data frame
   df <- data.frame(A = "Values A (DATA.FRAME)",
                    E = "Values E (DATA.FRAME)")

# Vectors
   A <- "Values of the vector 'A'"
   B <- "Values of the vector 'B'"


# A Call object `CALL`:

fun  <- function(data, gr, ID) match.call()
CALL <- fun(df, A, B)
CALL
## fun(data = df, gr = A, ID = B)

# Possible outputs -----------------------------------------------------------------------

getVarValues(variable = gr, data = df, CALL = CALL)
## [1] Values A (DATA.FRAME)

getVarValues(A, df, CALL)
## "Values of the vector 'A'"

getVarValues(B, df, CALL)
##

getVarValues(gr, df, CALL)
## Values A (DATA.FRAME)

# UNEXPECTED results -----------------------------------------------------------------------

getVarValues(ID, df, CALL) # `ID` is neither in `df`, nor in environment `env`
## NULL

getVarValues(G, df, CALL) # ERROR, as variable G does not exist.
##  Error in eval(expr, envir, enclos) : object 'G' not found

getVarValues(F, df, CALL) # F is a special variable: F = FALSE
##  FALSE

getVarValues(c, df, CALL) # c() is a function.
## function (..., recursive = FALSE)  .Primitive("c")

}

