% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_crosstab2.R
\name{gg_crosstab2}
\alias{gg_crosstab2}
\title{[!] Plot crosstab for 2 variables in hyperSpec object}
\usage{
gg_crosstab2(sp, x_var, y_var, xLabel = labels(sp, x_var),
  yLabel = labels(sp, y_var), ID_var = "ID",
  rotate.x.tick.labels = NULL, useNA = "ifany")
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{x_var, }{y_var names of variables inside \code{sp}, that contain
information to be crosstabulated.}

\item{xLabel, }{yLabel labels for variables \code{x_var, y_var} respectively.}

\item{ID_var}{name of variable inside \code{sp}, that contain ID of sample.}

\item{rotate.x.tick.labels}{Determine if x axis tick labels should be rotated
in 45 degrees angle (to avoid overlapping). Possible valued \code{TRUE},
\code{FALSE} or \code{NULL} (default, rotate if maximum length of x tick
 label is more that 3 symbols).}

\item{useNA}{whether to include \code{NA} values in the table.
    See \sQuote{Details}.  Can be abbreviated.}
}
\value{
2 ploted crosstabs: first for number of medical samples,
 second for number of spectra.
}
\description{
[!] Plot crosstab for 2 variables in hyperSpec object
}
\examples{
library(spHelper)

gg_crosstab2(Spectra2,"gr","class", ID_var = NULL)


Spectra2$gr3<- c("a123", "b123")
gg_crosstab2(Spectra2,"gr3","class", ID_var = NULL)

}
