% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_spectra.R
\name{has_too_few_IDs}
\alias{has_too_few_IDs}
\title{Get names of groups that has to few unique cases.}
\usage{
has_too_few_IDs(OBJ, Var = colnames(OBJ)[1], ID = "ID", n_min = 5,
  na.rm = TRUE)
}
\arguments{
\item{OBJ}{hyperSpec object}

\item{Var}{(string) variable name of factor variable}

\item{ID}{(string) variable name of ID variable}

\item{n_min}{(integer) minimum number of unique IDs not to be treated as too few.}

\item{na.rm}{(logical) used by function \code{sum()}. Default is \code{TRUE}.}
}
\value{
Vector of strings.
}
\description{
Get names of groups that has to few unique cases.
}
\examples{
has_too_few_IDs(Spectra2, "class", ID = "gr")

has_too_few_IDs(Spectra2, "class", ID = "gr", n_min = 2)

}
\seealso{
Other count spectra functions: \code{\link{count_spectra}},
  \code{\link{has_enough_IDs}}
}
\concept{count spectra functions}
