% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_spc2df.R
\name{hy_spc2df}
\alias{hy_spc2df}
\title{Return transposed matrix of spectra in hyperSpec object}
\usage{
hy_spc2df(obj)
}
\arguments{
\item{obj}{hyperSpec object}
}
\value{
A data frame with wavelengths in the first column and each row of
hyperSpec object in a separate column.
}
\description{
Transpose matrix of spectra in hyperSpec object, add column with wavelengths
and return as a dataframe
}
\examples{
  library(spHelper)

  chondro[[]][1,1:6]
  hy_spc2df(chondro)[1:6,1:2]
}
