% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd.color.R
\name{hyAdd.color}
\alias{hyAdd.color}
\title{[+] Add a variable with color names to \code{hyperSpec} object}
\usage{
hyAdd.color(sp = NULL, by, palette = c("#377EB8", "#4DAF4A", "#984EA3",
  "#FF7F00", "#A65628", "#F781BF", "#999999"))
}
\arguments{
\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}

\item{by}{A factor variale which levels will correspond to colors in
\code{.color}.}

\item{palette}{A color palette (vector with colors for each level in
\code{by}). If this argument is not provided, default palette is used.}
}
\value{
\code{HyperSpec} object with added/replaced column \code{.color}.
             Lables of variable \code{.color} indicate unique colors used
             (illustration in section "Examples").
}
\description{
Add (or overwrite, if already exists) a column \code{.color} with color
names that correspond to levels of factor variable \code{by}.
}
\examples{

data(Spectra)
Spectra2 <- hyAdd.color(Spectra, "class")

colnames(Spectra)
   #> [1] "gr"    "class" "spc"

colnames(Spectra2)
   #> [1] "gr"     "class"  "spc"    ".color"

# Names of colors, used for each level of factor variable
labels(Spectra2,".color")
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

hyGet.palette(Spectra2)
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

# ATTENTION -------------------------------------------------

# Preserve labels/ color palette:
Spectra2[1,".color"] <- "red"
hyGet.palette(Spectra2)
    #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

# Overwrites labels/ color palette:
Spectra2$.color[1] <- "red"
hyGet.palette(Spectra2)
hyGet.palette(Spectra2)
    #>  Warning message:
    #>  In hyGet.palette(Spectra2) : Values of pallete do not exist.
    #>  Most probably they are overwriten by operation `$.color<-`.

labels(Spectra2,".color")
    #>  ".color"

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

