% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd.Label.wl.R
\name{hyAdd.Label.wl}
\alias{hyAdd.Label.wl}
\title{[!] Add a standardised label for x-axis in \code{hyperSpec} object}
\usage{
hyAdd.Label.wl(sp, modeX, label)
}
\arguments{
\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}

\item{modeX}{Mode of x axis. A string, that indicates either "Fluorescence",
"wavelength", "Raman" or "IR".}

\item{label}{A manually eneterd text or expression of the label.}
}
\value{
Object with added label.
}
\description{
[!] Add a standardised label for x-axis in \code{hyperSpec} object
}
\examples{

Sp2 <- hyAdd.Label.wl(Spectra, "IR")

labels(Spectra, ".wavelength")
#> expression(lambda/nm)

labels(Sp2, ".wavelength")
#> expression(list(Delta * tilde(nu), cm^-1))

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

