% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd.Labels_TD2009.R
\name{hyAdd.Labels_TD2009}
\alias{hyAdd.Labels_TD2009}
\title{[+] Add labels to "TD_2009" dataset}
\usage{
hyAdd.Labels_TD2009(sp, language = "EN")
}
\arguments{
\item{sp}{A \code{\link[=hyperSpec-class]{hyperSpec}} object of TD_2009 data,
created by function \code{\link{read3csv2hy}}.}

\item{language}{A string, indicating a language of labels. Possible entries are
\code{EN} - English and \code{LT} - Lithuanian. Default is \code{EN}.}
}
\value{
A labeled object with reduced number of data columns.
}
\description{
Function is designed to label data collected during investigation called "TD_2009".
Select data columns, that are not removed, and add labels to
 \code{\link[=hyperSpec-class]{hyperSpec}} object of "TD_2009" data.
}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD2014}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

