% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd.R
\name{hyAdd}
\alias{hyAdd}
\title{[!] Add a variable to \code{hyperSpec} object}
\usage{
hyAdd(sp, var.name, by, unique.values)
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{var.name}{The name of variable to be added.}

\item{by}{A factor variale which levels will correspond to values in
\code{unique.values}.}

\item{unique.values}{Unique values used to create the variavle.}
}
\value{
\code{HyperSpec} object with added/replaced variable.
             Lables of the variable indicate unique values used.
}
\description{
Add (or overwrite, if already exists) a variable which name is passed as
 \code{var.name} with values in \code{unique.values}
that correspond to levels of factor variable \code{by}.
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{IQR_outliers}},
  \code{\link{binning}}, \code{\link{file}},
  \code{\link{gapDer}}, \code{\link{hy2mat}},
  \code{\link{hyAdd_Label_wl}},
  \code{\link{hyAdd_Labels_PAP_PD_2014}},
  \code{\link{hyAdd_Labels_TD2009}},
  \code{\link{hyDrop_NA}}, \code{\link{hyGet_palette}},
  \code{\link{hyRm_palette}}, \code{\link{mad_outliers}},
  \code{\link{mean_Nsd}}, \code{\link{median_Nmad}},
  \code{\link{plot_hyPalette}},
  \code{\link{prepare_PAP_RK_2014__MATLAB_failui}},
  \code{\link{read.OOIBase32}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView}}, \code{\link{read.sp.csv2}},
  \code{\link{read3csv2hy}}, \code{\link{replace_spc}},
  \code{\link{sd_outliers}}, \code{\link{spStat}},
  \code{\link{sp_class_perform}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
