% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd_color.R
\name{hyRm_palette}
\alias{hyRm_palette}
\alias{hyAdd_palette}
\alias{hyAdd_color}
\title{[+!] Add a variable with color names to \code{hyperSpec} object}
\usage{
hyRm_palette(sp = NULL)

hyAdd_palette(sp = NULL, by, palette = c(RColorBrewer::brewer.pal(9,
  "Set1")), label = NULL, ...)

hyAdd_color(sp = NULL, by, palette = c(RColorBrewer::brewer.pal(9,
  "Set1")), label = NULL, ...)
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{by}{A factor variale which levels will correspond to colors in
\code{.color}.}

\item{palette}{A color palette (vector with colors for each level in
\code{by}). If this argument is not provided, default palette is used.}

\item{label}{Label for column \code{.color}. If \code{label = NULL}, default
label is added.}
}
\value{
\code{HyperSpec} object with added/replaced column \code{.color}.
             Lables of variable \code{.color} indicate unique colors used
             (illustration in section "Examples").
}
\description{
Add (or overwrite, if already exists) a column \code{.color} with color
names that correspond to levels of factor variable \code{by}.
Function \code{hyRm_palette} removes this palette.
}
\examples{

data(Spectra2)
Spectra1 <- hyAdd_color(Spectra2, "class")

colnames(Spectra2)
   #> [1] "gr"    "class" "spc"

colnames(Spectra1)
   #> [1] "gr"     "class"  "spc"    ".color"

# Default labels
labels(Spectra1,".color")
   #>  [1] "Colors for \\"class\\""

hyGet_palette(Spectra1)
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

# ATTENTION -------------------------------------------------

# Preserve labels:
Spectra1[1,".color"] <- "red"
hyGet_palette(Spectra1)

labels(Spectra1,".color")

# Overwrites labels:
Spectra1$.color[1] <- "red"

hyGet_palette(Spectra1)

labels(Spectra1,".color")
    #>  ".color"

}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{IQR_outliers}},
  \code{\link{binning}}, \code{\link{file}},
  \code{\link{gapDer}}, \code{\link{hy2mat}},
  \code{\link{hyAdd_Label_wl}},
  \code{\link{hyAdd_Labels_PAP_PD_2014}},
  \code{\link{hyAdd_Labels_TD2009}}, \code{\link{hyAdd}},
  \code{\link{hyDrop_NA}}, \code{\link{hyGet_palette}},
  \code{\link{mad_outliers}}, \code{\link{mean_Nsd}},
  \code{\link{median_Nmad}}, \code{\link{plot_hyPalette}},
  \code{\link{prepare_PAP_RK_2014__MATLAB_failui}},
  \code{\link{read.OOIBase32}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView}}, \code{\link{read.sp.csv2}},
  \code{\link{read3csv2hy}}, \code{\link{replace_spc}},
  \code{\link{sd_outliers}}, \code{\link{spStat}},
  \code{\link{sp_class_perform}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
