% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd_Label_wl.R, R/label_wl.R
\name{hyAdd_Label_wl}
\alias{hyAdd_Label_wl}
\alias{label_wl}
\title{[!] Add a standardised label for x-axis in \code{hyperSpec} object}
\usage{
hyAdd_Label_wl(sp, modeX = NULL, label = NULL)

label_wl(modeX = NULL, label = NULL)
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{modeX}{Mode of x axis. A string, that indicates either "Fluorescence",
"wavelength", "Raman" or "IR".}

\item{label}{A manually eneterd text or expression of the label.}
}
\value{
Object with added label.
}
\description{
[!] Add a standardised label for x-axis in \code{hyperSpec} object
}
\examples{

Sp2 <- hyAdd_Label_wl(Spectra2, "IR")

labels(Spectra2, ".wavelength")
#> expression(lambda/nm)

labels(Sp2, ".wavelength")
#> expression(list(Delta * tilde(nu), cm^-1))

}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
