% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd_Labels_TD2009.R
\name{hyAdd_Labels_TD2009}
\alias{hyAdd_Labels_TD2009}
\title{[+] Add labels to "TD_2009" dataset}
\usage{
hyAdd_Labels_TD2009(sp, language = "EN")
}
\arguments{
\item{sp}{A \code{\link[=hyperSpec-class]{hyperSpec}} object of TD_2009 data,
created by function \code{\link{read3csv2hy}}.}

\item{language}{A string, indicating a language of labels. Possible entries are
\code{EN} - English and \code{LT} - Lithuanian. Default is \code{EN}.}
}
\value{
A labeled object with reduced number of data columns.
}
\description{
Function is designed to label data collected during investigation called "TD_2009".
Select data columns, that are not removed, and add labels to
 \code{\link[=hyperSpec-class]{hyperSpec}} object of "TD_2009" data.
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
