% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyDrop.NA.R
\name{hyDrop.NA}
\alias{hyDrop.NA}
\title{[+] Remove variables of \code{hyperSpec} object that contain only \code{NA} values}
\usage{
hyDrop.NA(sp)
}
\arguments{
\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}
}
\value{
Object without columns with all \code{NA} values.
}
\description{
Remove variables (columns) of \code{\link[=hyperSpec-class]{hyperSpec}}
object that contain only \code{NA} values.
}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

