% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyDrop_NA.R
\name{hyDrop_NA}
\alias{hyDrop_NA}
\title{[+] Remove variables of \code{hyperSpec} object that contain only \code{NA} values}
\usage{
hyDrop_NA(sp)
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}
}
\value{
Object without columns with all \code{NA} values.
}
\description{
Remove variables (columns) of \code{\link[=hyperSpec-class]{hyperSpec}}
object that contain only \code{NA} values.
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
