% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyGet.palette.R
\name{hyGet.palette}
\alias{hyGet.palette}
\title{[+] Get color palette used to create variable'.color'}
\usage{
hyGet.palette(sp)
}
\arguments{
\item{sp}{A \code{hyperSpec} object, that contais variable \code{.color}
added with function \code{\link{hyAdd.color}}}
}
\value{
Color palette used to create variable \code{.color}.
}
\description{
[+] Get color palette used to create variable'.color'
}
\examples{
data(Spectra)
Spectra2 <- hyAdd.color(Spectra, "class")

# Names of colors, used for each level of factor variable

hyGet.palette(Spectra2)
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

labels(Spectra2,".color")
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

}

