% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyGet_palette.R
\name{hyGet_palette}
\alias{hyGet_palette}
\alias{hyGet_palette0}
\title{[+] Get color palette used to create variable'.color'}
\usage{
hyGet_palette(sp)

hyGet_palette0(sp)
}
\arguments{
\item{sp}{A \code{hyperSpec} object, that contais variable \code{.color}
added with function \code{\link{hyAdd_color}}}
}
\value{
Color palette used to create variable \code{.color}.
}
\description{
[+] Get color palette used to create variable'.color'
}
\note{
If class of \code{sp} is not \code{hyperSpec}, function returns
      \code{NULL}.
}
\examples{

data(Spectra2)
Spectra1 <- hyAdd_color(Spectra2, "class")

colnames(Spectra2)
   #> [1] "gr"    "class" "spc"

colnames(Spectra1)
   #> [1] "gr"     "class"  "spc"    ".color"

# Default label
labels(Spectra1,".color")
   #>  [1] "Colors for \"class\""

# Names of attributes of variable ".color":
attributes(Spectra1$.color)  \%>\% names
  #> [1] "palette"      "variableName"

# Values of attributes:
attributes(Spectra1$.color)
  #> $palette
  #> color label
  #> 1 #E41A1C     K
  #> 2 #377EB8     l
  #> 3 #4DAF4A     N
  #> 4 #984EA3    S1
  #>
  #> $variableName
  #> [1] "class"

# Extract color palette:
hyGet_palette(Spectra1)
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

   #>      colors   labels
   #>  1 #E41A1C     K
   #>  2 #377EB8     l
   #>  3 #4DAF4A     N
   #>  4 #984EA3    S1

# Extract palette (just colors) as vector:
hyGet_palette0(Spectra1)
   #>  [1] #E41A1C #377EB8 #4DAF4A #984EA3
   #>  Levels: #377EB8 #4DAF4A #984EA3 #E41A1C

# ATTENTION -------------------------------------------------

# Preserve labels/ color palette (operator ` ]<-`):
Spectra1[1,".color"] <- "red"
hyGet_palette(Spectra1)
    #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

# Overwrite labels/ color palette (operator `$<-`):
# Spectra2a$.color[1] <- "red"
# hyGet_palette(Spectra2a)
    #>  Warning message:
    #>  In hyGet_palette(Spectra2a) : Values of pallete do not exist.
    #>  Most probably they are overwriten by operation `$.color<-`.

labels(Spectra1,".color")
    #>  ".color"

}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
