% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperSpec.R
\name{hyperSpec}
\alias{hyperSpec}
\title{Initialize new hyperSpec object}
\usage{
hyperSpec(spc = NULL, data = NULL, wavelength = NULL,
  labels = NULL)
}
\arguments{
\item{spc}{the spectra matrix.

\code{spc} does not need to be a matrix, it is converted explicitly by
  \code{I (as.matrix (spc))}.}

\item{data}{\code{data.frame}, possibly with the spectra in
  \code{data$spc}, and further variates in more columns.  A matrix can be
  entered as \emph{one} column of a data frame by: \code{data.frame (spc =
  I (as.matrix (spc)))}.

However, it will usually be more convenient if the spectra are given in
  \code{spc}}

\item{wavelength}{The wavelengths corresponding to the columns of
\code{data}. If no wavelengths are given, an appropriate vector is
derived from the column names of \code{data$spc}. If this is not
possible, \code{1 : ncol (data$spc)} is used instead.}

\item{labels}{A \code{list} containing the labels for the columns of the
  \code{data} slot of the \code{hyperSpec} object and for the wavelength
  (in \code{label$.wavelength}). The labels should be given in a form ready
  for the text-drawing functions (see \code{\link[grDevices]{plotmath}}).

If \code{label} is not given, a list containing \code{NULL} for each of the
  columns of\code{data} and \code{wavelength} is used.}
}
\value{
A \code{hyperSpec} object.
}
\description{
A wrapper function to initialize a new \code{hyperSpec} object. More
 information in \pkg{hyperSpec} \code{\link[hyperSpec]{initialize}}.
}
