% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr.R
\name{ilr}
\alias{ilr}
\alias{ilr.default}
\alias{ilr.hyperSpec}
\title{[!] Isometric logratio transformation}
\usage{
ilr(x, V = ilrBase(x), ...)

\method{ilr}{default}(x, V = ilrBase(x), ...)

\method{ilr}{hyperSpec}(x, V = ilrBase(x), ...)
}
\arguments{
\item{x}{A composition, not necessarily closed. (A numeric data frame or matrix.)}

\item{...}{further parameters to \code{\link[compositions]{ilr}}.}
}
\value{
Returns the transformed data matrix with one dimension less than
     \code{x}. Thus for \code{hyperSpec} object the last value of
     \code{x$.wavelength} is removed.
}
\description{
Compute the isometric log ratio transform of a (dataset of) composition(s).
 Default method is function
\code{\link[compositions]{ilr}} from package \pkg{compositions}.
}
\examples{
ilr(Spectra)

plot(Spectra)
plot(ilr(Spectra))

}
\seealso{
\code{\link[compositions]{ilr}}, \code{\link[compositions]{ilrBase}}

Other row-wise transformations: 
\code{\link{alr}()},
\code{\link{apt}()},
\code{\link{clr}()},
\code{\link{cpt}()},
\code{\link{ipt}()}
}
\concept{row-wise transformations}
