% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.matrix.R
\name{ind.matrix}
\alias{ind.matrix}
\title{Create a matrix filled with its element indices}
\usage{
ind.matrix(x)
}
\arguments{
\item{x}{A matrix.}
}
\description{
Genereta a matrix of the same size as provided matrix \code{x}, where each
element represents its index.
}
\examples{

 ind.matrix(matrix(NA, 2, 5))
 ind.matrix(matrix(NA, 5, 2))

}
\author{
Vilmantas Gegzna
}
\seealso{
\link[base]{row},\link[base]{col}

Other matrix operations: \code{\link{corr_vec2mat}},
  \code{\link{nDiag2nTri}}, \code{\link{nTri2nDiag}},
  \code{\link{new.matrix}}, \code{\link{sortMaxOnDiag}},
  \code{\link{which.in}}, \code{\link{which.max.all}},
  \code{\link{which.max.perRow}}
}

