% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpFun.R
\name{info_dim}
\alias{info_dim}
\title{Calculate information dimension of a matix.}
\usage{
info_dim(y, makePlot = FALSE)
}
\arguments{
\item{y}{- data matrix (rows = observations, columns = variables)}
}
\value{
\itemize{
 \item{"$"}{Rounded towards positive infinitive}
 \item{"parameter 2"}{Stuff}
}
     info - struktÅ«ra su laukais:
             dim_exact: informacijos dimensijos reikÅ<U+FFFD>mÄ— (nesuapvalinta).
                eigval: tikrinÄ—s reikÅ<U+FFFD>mÄ—s
}
\description{
The function calculates a measure, called "information dimension"
}
\note{
PrieÅ<U+FFFD> pradedant vykdyti operacijÄ…, svarbus Å¾ingsnis pasirinkti tinkamÄ…
normavimo buda. To nepadarius gausime klaidingÄ… atsakymÄ….
y=sp_normuok(y,x,'1',495);

Taip pat labai svarbus ir triukÅ<U+FFFD>mo lygis. DidÄ—jant triukÅ<U+FFFD>mui atitinkamai
padidinamas maksimalus dimensijÅ³ skaiÄius.

eigenvalues - Tikrines reiksmes / Singular values
pk - tikimybines dimensiju vertes, skirtos entropijos ivertinimui.
explain = pk;
}
\references{
[1]	R. Cangelosi and A. Goriely, Component retention in principal
      component analysis with application to cDNA microarray data.
      Biol Direct, 2, 2 (2007), \url{http://dx.doi.org/10.1186/1745-6150-2-2}
}

