% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Addins.R
\name{insertIn_Addin}
\alias{insertIn_Addin}
\title{Insert \%in\%.}
\usage{
insertIn_Addin()
}
\description{
Call this function as an addin to insert \code{ \%in\% } at the cursor position.
}
\seealso{
Other \pkg{spHelper} addins: \code{\link{insertMatMuliplication_Addin}},
  \code{\link{insertPipeline_Addin}}
}

