% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Addins.R
\name{insertMatMuliplication_Addin}
\alias{insertMatMuliplication_Addin}
\title{Insert \%*\%.}
\usage{
insertMatMuliplication_Addin()
}
\description{
Call this function as an addin to insert \code{ \%*\% } at the cursor position.
}
\seealso{
Other \pkg{spHelper} addins: \code{\link{insertIn_Addin}},
  \code{\link{insertPipeline_Addin}}
}

