% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipt.R
\name{ipt}
\alias{ipt}
\alias{ipt.default}
\alias{ipt.hyperSpec}
\title{[!] Isometric planar transformation}
\usage{
ipt(x, V = ilrBase(x), ...)

\method{ipt}{default}(x, V = ilrBase(x), ...)

\method{ipt}{hyperSpec}(x, V = ilrBase(x), ...)
}
\arguments{
\item{x}{A composition, not necessarily closed. (A numeric data frame or matrix.)}

\item{...}{further parameters to \code{\link[compositions]{ipt}}.}
}
\value{
Returns the transformed data matrix with one dimension less than
     \code{x}. Thus for \code{hyperSpec} object the last value of
     \code{x$.wavelength} is removed.
}
\description{
Compute the isometric planar transform of a (dataset of) composition(s).
 Default method is function
\code{\link[compositions]{ipt}} from package \pkg{compositions}.
}
\examples{
ipt(Spectra)

plot(Spectra)
plot(ipt(Spectra))

}
\seealso{
\code{\link[compositions]{ipt}}, \code{\link[compositions]{ilrBase}}

Other row-wise transformations: 
\code{\link{alr}()},
\code{\link{apt}()},
\code{\link{clr}()},
\code{\link{cpt}()},
\code{\link{ilr}()}
}
\concept{row-wise transformations}
