% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR_outliers.R
\name{IQR_outliers}
\alias{IQR_outliers}
\title{Outlier detection with IQR based distances (limits)}
\usage{
IQR_outliers(x, k_out = 3, k_inn = 1.5)
}
\arguments{
\item{x}{Matrix-like data or \code{hyperSpec} object.}

\item{k_out}{Number of IQR distances from Q1 and Q3 for \code{extreme}
outlier identification. Default value is 3.}

\item{k_inn}{Number of IQR diustances from Q1 and Q3 for \code{probable}
outlier idientification. Default value is 1.5.}
}
\value{
A list with these variables
\itemize{
   \item{\emph{outer} / \emph{inner} - matrices with values of outer / inner limits respectively}
   \item{Logical matrices indicating \code{TRUE} for points that are:
         \itemize{
             \item{\emph{is_out} - (extreme) outliers (excluding probable outliers);}
             \item{\emph{is_pout} - probable outliers (excluding extreme outliers);}
             \item{\emph{is_bothOut} - either probable or extreme outliers;}
          }
     }

     \item{Logical vectors indicating spectra / rows with at least one:
          \itemize{
             \item{\emph{row_out} -  (extreme) outlier point;}
             \item{\emph{row_pout} - probable outlier point (excluding extreme outliers);}
             \item{\emph{row_bothOut} - either probable or extreme outlier point.}
          }
     }
}
If x is a \code{hyperSpec} object, then the listed variables are returned as
\code{hyperSpec} objects. \cr

\bold{Note} that probable outliers are points between corresponding lower as
well as corresponding upper \code{inner} and \code{outer} limits.
Extreme outliers are points outside lower and upper \code{outer} limits.
}
\description{
Outlier detection with interquartile range (IQR) based distances, defined as:
\bold{lower} limit \eqn{Q1 - k * IQR} and
\bold{upper} limit \eqn{Q3 + k * IQR}. Here Q1 and Q3 is the first and the
third quantile respectively.
\bold{Inner} lower/upper limit for probable outlier detection has \code{k = k_inn} and
\bold{outer} lower/upper limit for (extreme) outlier detection has \code{k = k_out},
(\code{k_out > k_inn}).
}
\examples{

outl <- IQR_outliers(Spectra2)

plot(outl$outer, col = NA)
plot(Spectra2, add = TRUE, spc.nmax = nrow(Spectra2))
plot(outl$outer, add = TRUE, col = "red",    lines.args = list(lwd = 3))
plot(outl$inner,  add = TRUE, col = "orange", lines.args = list(lwd = 3))
legend("topright",legend=c("Outer limit", "Inner limit", "Spectra"),
 col = c("red", "orange", "black"), lwd = c(3,3,1), lty = 1)
title("IQR based distances")

}
\seealso{
Read about boxplots and outliers
(\href{http://www.physics.csbsju.edu/stats/box2.html}{Box Plot: Display of Distribution})

Other \pkg{spHelper} utilities: 
\code{\link{expr2text}()},
\code{\link{mad_outliers}()},
\code{\link{rmExpr}()},
\code{\link{sd_outliers}()},
\code{\link{subt}()},
\code{\link{unGroup}()},
\code{\link{uncall}()}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}

Other outlier detection functions in \pkg{spHelper}: 
\code{\link{mad_outliers}()},
\code{\link{sd_outliers}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
\concept{\pkg{spHelper} utilities}
\concept{outlier detection functions in \pkg{spHelper}}
