% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_internal_functions.R
\name{ldf}
\alias{ldf}
\alias{as.ldf}
\title{[!] As long data frame}
\usage{
ldf(x, ...)

as.ldf(x, ...)
}
\arguments{
\item{x}{a \code{hyperSpec} object}

\item{...}{ignored}
}
\description{
Wrapper for \code{\link[hyperSpec]{as.long.df}}:\cr
\code{as.long.df(x,  rownames = TRUE, na.rm = TRUE,...)}
}
