% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.functions.R
\name{list.functions}
\alias{list.functions}
\title{[+] List all functions in a package}
\usage{
list.functions(Package = "spHelper")
}
\arguments{
\item{Package}{A name of a package. Default \code{Package = "spHelper"}.}
}
\value{
A data frame with function names in a package.
}
\description{
List all functions in a package and return it as a data frame.
}
\examples{

list.functions()
list.functions(Package = "tidyr")

}
\author{
Vilmantas Gegzna
}

