% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAddRm.R
\name{listAddRm}
\alias{listAddRm}
\title{[.] Find and display added an removed names}
\usage{
listAddRm(initial, final)
}
\arguments{
\item{initial}{A charecter vector with initial names.}

\item{final}{A charecter vector with final names.}
}
\value{
A list with added and removed names
}
\description{
[.] Find and display added an removed names
}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{clear}},
  \code{\link{expr2text}}, \code{\link{make.filenames}},
  \code{\link{make.firstCapitals}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{st01}},
  \code{\link{subt}}, \code{\link{uncall}}
}

