% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFunctions.R
\name{listFunctions}
\alias{listFunctions}
\title{[+] List all functions in a package}
\usage{
listFunctions(Package = "spHelper")
}
\arguments{
\item{Package}{A name of a package. Default \code{Package = "spHelper"}.}
}
\value{
A data frame with function names in a package.
}
\description{
List all functions in a package and return it as a data frame.
}
\examples{

listFunctions()
listFunctions(Package = "tidyr")

}
\author{
Vilmantas Gegzna
}

