% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.firstCapitals.R
\name{make.firstCapitals}
\alias{fCap}
\alias{make.firstCapitals}
\title{[+] Convert first letters of all words to capitals}
\usage{
make.firstCapitals(x)

fCap(x)
}
\arguments{
\item{x}{Either a string or a vector of strings.}
}
\value{
The same string as input with all words starting in capital letters.
}
\description{
Convert first letters of all words in a string to capitals.
}
\details{
\code{fCap} is a wrapper of \code{make.firstCapitals}.
}
\examples{

make.firstCapitals('laa laa laa')
##[1] "Laa Laa Laa"

fCap('laa laa laa')
##[1] "Laa Laa Laa"

}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clear}}, \code{\link{expr2text}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{subt}},
  \code{\link{uncall}}
}

