% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.filenames.R
\name{make.nameswin}
\alias{make.nameswin}
\title{[+] Convert string into Windows compatible filename}
\usage{
make.nameswin(s)
}
\arguments{
\item{s}{A string.}
}
\value{
A string without spacial symbols.
}
\description{
Function replaces symbols \ / : * ? " < > | in a string \code{s}
 with underscore. After replacement string can be used as a file name
in Windows.
}
\examples{
 s <- '\\\\ / : * ? " < > |'
 make.nameswin(s)
 #> "_ _ _ _ _ _ _ _ _"

 s2 <- "Hello?"
 make.nameswin(s2)
 #> "Hello_"


}
\seealso{
Other family \pkg{spHelper} utilities: \code{\link{clc}},
  \code{\link{st01}}
}

