% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_kappa_tmp.R
\name{make_table}
\alias{make_table}
\title{[Internal] Make table (from prediction object)}
\usage{
make_table(pred, index)
}
\arguments{
\item{pred}{object of class "\code{prediction}"
from package \pkg{ROCR}}

\item{index}{index of value in the object}
}
\value{
Clasification table with TP, FP, FN, TN values
}
\description{
[Internal] Make table (from prediction object)
}
