% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFirstCapital.R
\name{makeFirstCapital}
\alias{fCap}
\alias{makeFirstCapital}
\title{[!] Convert the first letter to capital.}
\usage{
makeFirstCapital(x)

fCap(x)
}
\arguments{
\item{x}{- a string or vector of strings}
}
\value{
The same string with all words starting with capital letter.
}
\description{
[!] Convert the first letter to capital.
}
\details{
\code{fCap} is a wrapper of \code{makeFirstCapital}
}
\examples{

makeFirstCapital('laa laa laa')
##[1] "Laa Laa Laa"

fCap('laa laa laa')
##[1] "Laa Laa Laa"

}

