% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_.R
\name{measure_bac}
\alias{measure_bac}
\title{Balanced accuracy (for 2 or more groups)}
\usage{
measure_bac(truth, response)
}
\arguments{
\item{truth}{a vector with true (reference) values}

\item{response}{a vector with response (predicted) values}
}
\description{
Balanced accuracy (for 2 or more groups)
}
\seealso{
Other measures_: 
\code{\link{measure_bac_2gr}()},
\code{\link{measure_tn}()},
\code{\link{measure_tpr}()},
\code{\link{measure_tp}()}
}
\concept{measures_}
