% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_.R
\name{measure_bac_2gr}
\alias{measure_bac_2gr}
\title{Balanced accuracy (for binary classification)}
\usage{
measure_bac_2gr(truth, response, negative, positive)
}
\arguments{
\item{truth}{a vector with true (reference) values}

\item{response}{a vector with response (predicted) values}
}
\description{
# This function is based on function `measureBAC` from `mlr` package
}
\seealso{
Other measures_: 
\code{\link{measure_bac}()},
\code{\link{measure_tn}()},
\code{\link{measure_tpr}()},
\code{\link{measure_tp}()}
}
\concept{measures_}
