% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_kappa_tmp.R
\name{measure_kappa_tmp}
\alias{measure_kappa_tmp}
\title{[Internal] measure_kappa_tmp}
\usage{
measure_kappa_tmp(pred)
}
\arguments{
\item{pred}{object of class "prediction" from package ROCR}
}
\description{
[Internal] measure_kappa_tmp
}
