% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_.R
\name{measure_tn}
\alias{measure_tn}
\title{Calculate number of true negatives}
\usage{
measure_tn(truth, response, negative)
}
\description{
Calculate number of true negatives
}
\seealso{
Other measures_: 
\code{\link{measure_bac_2gr}()},
\code{\link{measure_bac}()},
\code{\link{measure_tpr}()},
\code{\link{measure_tp}()}
}
\concept{measures_}
