% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_.R
\name{measure_tp}
\alias{measure_tp}
\title{Calculate number of true positives}
\usage{
measure_tp(truth, response, positive)
}
\description{
Calculate number of true positives
}
\seealso{
Other measures_: \code{\link{measure_bac_2gr}},
  \code{\link{measure_bac}}, \code{\link{measure_tn}},
  \code{\link{measure_tpr}}
}
\concept{measures_}
