% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_.R
\name{measure_tpr}
\alias{measure_tpr}
\title{Rate of correctly identified values in a certain group}
\usage{
measure_tpr(truth, response, level)
}
\arguments{
\item{truth}{a vector with true (reference) values}

\item{response}{a vector with response (predicted) values}

\item{level}{(character) a name of level (group) of interest}
}
\description{
Rate of correctly identified values in a certain group
}
\seealso{
Other measures_: \code{\link{measure_bac_2gr}},
  \code{\link{measure_bac}}, \code{\link{measure_tn}},
  \code{\link{measure_tp}}
}
\concept{measures_}
