% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{median_CI}
\alias{median_CI}
\title{Confidence interval for median}
\usage{
median_CI(x = NULL, MEDIAN = median(x), n = length(x), IQR = IQR(x))
}
\arguments{
\item{x}{data}
}
\value{
List with results... ???
}
\description{
CI for case where variance of population is unknown.
}
\examples{

library(spHelper)

x <- c(3,0,2,1,0,4,3,2,1,2)
# median_CI(x)

}
\seealso{
Other statistical functions: 
\code{\link{normal_var_CI}()},
\code{\link{poisson_lambda_CI}()},
\code{\link{quartiles}()}
}
\concept{statistical functions}
