% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_Nmad.R
\name{median_Nmad}
\alias{median_Nmad}
\title{[!] Find spectra that represent median +/- n median absolute deviations (MAD)}
\usage{
median_Nmad(
  sp,
  n = 2,
  plus.minus = TRUE,
  var_fun = mad,
  center_fun = median,
  center_name = as.character(match.call()$center_fun),
  var_name = as.character(match.call()$var_fun)
)
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{n}{Number of MADs.}

\item{plus.minus}{Logical. If \code{TRUE}, two spectra representing \code{±n}
are calculated. Otherwise only one spectrum representing \code{n} is calculated.}

\item{var_fun}{Function that calculates variability measure. Default is
\code{var_fun = mad}.}

\item{center_fun}{Function that calculates center tendency. Default is
\code{center_fun = median}.}

\item{center_name}{A string with the name for measure of center tendency
(will be used to create a row name in \code{.name}).}

\item{var_name}{A string with the name for measure of variability
(will be used to create a row name in \code{.name}).}
}
\value{
A \code{hyperSpec} object with spectra at \code{median(sp) ± n*MAD(sp)}.
}
\description{
Function finds spectra that represent mean +/- n MADs
and adds column \code{.name}.
}
\examples{

data(Spectra2)
Margins  <- median_Nmad(Spectra2)
Margins3 <- median_Nmad(Spectra2, n = 3)

plotspc(Margins)


# Plot data and margins =======================================

Marg2 <- as.long.df(Margins,  rownames = TRUE, na.rm = FALSE)
Marg3 <- as.long.df(Margins3, rownames = TRUE, na.rm = FALSE)

Spectra2$.name <- factor("Spectra")

qplotspc(Spectra2, spc.nmax = nrow(Spectra2),
         mapping = aes(x = .wavelength, y = spc, group = .rownames, color = .name),
         alpha = .25) + theme_light() +
    geom_line(data = Marg2, aes(color = .name), lwd = 1) +
    geom_line(data = Marg3, aes(color = .name), lwd = 1) +
    scale_color_manual(" ", values = c("darkgreen","blue", "red")) +
    scale_alpha_manual(" ", values = ".2", guide = FALSE)

}
\seealso{
\link[hyperSpec]{scale},  \link{outside_mean_pm_Nsd},
  \link[hyperSpec]{mean_sd}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
