% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3__Mode.R
\name{Mode}
\alias{Mode}
\title{[.] Compute a mode (most frequent value)}
\source{
\href{http://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode}{stackoverflow.com - Mode}
}
\usage{
Mode(x)
}
\arguments{
\item{x}{A matrix-like data.}
}
\description{
Function, that defines, what is a Mode (a statistic measure)
}

