% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.matrix.R
\name{new.matrix}
\alias{new.matrix}
\title{Put new values to the old matrix}
\usage{
new.matrix(x, values = NA, byrow = FALSE, dimnames = NULL)
}
\arguments{
\item{x}{An old matrix.}

\item{values}{The new values for matrix}

\item{byrow}{logical. If \code{FALSE} (the default) the matrix is
    filled by columns, otherwise the matrix is filled by rows.}

\item{dimnames}{A \code{\link{dimnames}} attribute for the matrix:
    \code{NULL} or a \code{list} of length 2 giving the row and column
    names respectively.  An empty list is treated as \code{NULL}, and a
    list of length one as row names.  The list can be named, and the
    list names will be used as names for the dimensions.}
}
\description{
A convenience function to create a matrix of the same size as the provided one,
just with new values.
}
\examples{

 x <- matrix(NA, 2, 5)
 new.matrix(x, 1)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other matrix operations in \pkg{spHelper}: \code{\link{corr_vec2mat}},
  \code{\link{ind.matrix}}, \code{\link{nDiag2nTri}},
  \code{\link{nTri2nDiag}}, \code{\link{sortDescOnDiag}},
  \code{\link{which.in}}, \code{\link{which.max.all}},
  \code{\link{which.max.perRow}}
}

