% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nID_nSp.R
\name{nID_nSp}
\alias{nID_nSp}
\alias{nID_nObs}
\title{[+!] Calculate number and percentage of unique IDs and observations}
\usage{
nID_nSp(
  data = NULL,
  ID,
  gr,
  ID_text = "medical specimens",
  observation_text = "spectra",
  decimals = 1
)

nID_nObs(
  data = NULL,
  ID,
  gr,
  ID_text = "unique IDs",
  observation_text = "observations"
)
}
\arguments{
\item{data}{A data frame with variables, that names are denoted by \code{ID} and \code{gr}}

\item{ID}{Vector with speciment IDs (Either a vector or a variable name in \code{data}).}

\item{gr}{Vector with speciment groups (Either a vector or a variable name in \code{data}).}

\item{ID_text}{A title for unique IDs, that apear in the table.}

\item{observation_text}{A title for observations, that apear in the table.}

\item{decimals}{A number of maximum decimal places in percentages.
This parameter is passed to function \code{\link[base]{round}}. Default is 0.}
}
\value{
Table (data frame) with count and percentages.
}
\description{
Calculate number and percentage of unique IDs and observations.\cr

number(IDs) >= number(observations)

# DEFAULT BEHAVIOUR OF FUNCTION CHANGED
}
\examples{
data(DataSet1)

nID_nSp(DataSet1, "ID", "gr")
nID_nSp(ID = DataSet1$ID, gr = DataSet1$gr)
pander::pander(nID_nSp(DataSet1, "ID", "gr"))

nID_nObs(DataSet1, "ID", "gr")
pander::pander(nID_nObs(DataSet1, "ID", "gr"))

# For hyperSpec object
nID_nSp(Spectra2$.., ID = "class", "gr")

}
\author{
Vilmantas Gegzna
}
