% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{normal_var_CI}
\alias{normal_var_CI}
\title{Confidence interval for variance (parameter sigma squared) of Noral distribution}
\usage{
normal_var_CI(x = NULL, var = var(x), n = length(x),
  conf.level = 0.95)
}
\arguments{
\item{x}{data}

\item{var}{variance}

\item{n}{number of samples}

\item{conf.level}{confidence level, number between 0 and 1, usually 0.95.}
}
\value{
List with results... ???
}
\description{
CI for case where variance of population is unknown.
}
\examples{

library(spHelper)

x <- c(3,0,2,1,0,4,3,2,1,2)
normal_var_CI(x)

normal_var_CI(var = 10, n = 20)

}
\seealso{
Other statistical functions: \code{\link{median_CI}},
  \code{\link{poisson_lambda_CI}}, \code{\link{quartiles}}
}
\concept{statistical functions}
