% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nRow2nTri.R
\name{nRow2nTri}
\alias{nRow2nTri}
\title{Matrix: calculate a number of elements in a lower/upper triangle of a symmetric matrix of size n*n}
\usage{
nRow2nTri(n)
}
\arguments{
\item{n}{A number of rows/columns in a symmetric matrix.}
}
\value{
A number of elements in a lower/upper triangle of a matrix (excluding the diagonal).
}
\description{
Number of elements in a lower/upper triangle of a symmetric matrix of size n*n
}
\details{
nCoeff = (n^2-n)/2, thus
n = (1+sqrt(1+8*nCoeff))/2
}
\examples{

nRow2nTri(5)
nRow2nTri(7)

}
\seealso{
\link{nTri2nRow}, \link{simCorrVars}

Other Matrix operations: \code{\link{corr_vec2mat}},
  \code{\link{nTri2nRow}}, \code{\link{which.in}}
}

