% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nTri2nDiag.R
\name{nTri2nDiag}
\alias{nTri2nDiag}
\title{[+] Matrix: length of diagonal when a number of elements in a lower/upper triangle is given}
\usage{
nTri2nDiag(nTri)
}
\arguments{
\item{nTri}{A number of elements in a lower/upper triangle of a matrix (excluding the diagonal).}
}
\value{
A number of rows/columns in a symmetric matrix. May be a fractional number.
}
\description{
Calculate a number of diagonal elements of a symmetric matrix when a number
elements in a lower/upper triangle is given.
}
\details{
Dedived from equation nCoeff = (n^2-n)/2
}
\examples{
nTri2nDiag(10)
nTri2nDiag(6)

}
\seealso{
Other matrix operations: \code{\link{corr_vec2mat}},
  \code{\link{nDiag2nTri}}, \code{\link{which.in}}
}

