% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nTri2nRow.R
\name{nTri2nRow}
\alias{nTri2nRow}
\title{Matrix: calculate a dimension of a symmetric matrix by a number elements in a lower/upper triangle}
\usage{
nTri2nRow(nTri)
}
\arguments{
\item{nTri}{A number of elements in a lower/upper triangle of a matrix (excluding the diagonal).}
}
\value{
A number of rows/columns in a symmetric matrix. May be a fractional number.
}
\description{
Matrix: calculate a dimension of a symmetric matrix by a number elements in a lower/upper triangle
}
\details{
Dedived from equation nCoeff = (n^2-n)/2
}
\examples{
nTri2nRow(10)
nTri2nRow(6)

}
\seealso{
\link{nRow2nTri}, \link{simCorrMat}
}

