% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_internal_functions.R
\name{parse_string}
\alias{parse_string}
\title{Parse a string and extract information to a dataframe.}
\usage{
parse_string(x, pattern, ignore.case = FALSE, ...)
}
\arguments{
\item{x}{A string to be parsed}

\item{pattern}{Either a string with regular expression or a function.}

\item{ignore.case}{A logical parameter to be passed to \code{\link{regexp2df}}.}

\item{...}{further parameters to be passed to \code{\link{regexp2df}}.}
}
\value{
A data frame with parsed information.
}
\description{
This function similar to \code{\link{regexp2df}}, but accepts both
strings and functions as values of argument \code{pattern}.
}
