% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_1.R
\name{parser_1}
\alias{parser_1}
\title{[!] Parse a filename}
\usage{
parser_1(x)
}
\arguments{
\item{x}{A sting.}
}
\value{
A data frame.
}
\description{
Parse a string with filename and return a dataframe with
extracted information.
}
\details{
Is \code{x} is a filename with a path, only the filename without path is parsed.
}
\examples{
string <- "aliejus;silicis;0.5mW;1;NA;_RamanShift_14-51-36-968.txt"
parser_1(string)

string3 <- c(string,string,string)
parser_1(string3)

}
