% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_2015.R
\name{parser_TD2015}
\alias{parser_TD2015}
\title{[!] Parse a filename}
\usage{
parser_TD2015(x)
}
\arguments{
\item{x}{A sting.}
}
\value{
A data frame.
}
\description{
Parse a string with filename and return a dataframe with
extracted information.
}
\details{
Is \code{x} is a filename with a path, only the filename without path is parsed.
}
\examples{
stringA <- "Duomenu_pvz/Fluorescencija/2010-03-16/ID_001/2.Scope"
parser_TD2015(stringA)

stringB <- "Duomenu_pvz/Fluorescencija/2014-08-13/ID_451/56.txt"
parser_TD2015(stringB)

stringC <- "Duomenu_pvz/Fluorescencija/2009-07-13/ID_051/66.Master.Scope"
parser_TD2015(stringC)

x <- c(stringA,stringB,stringC)
parser_TD2015(x)

}
