% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_internal_functions.R
\name{percent2probs}
\alias{percent2probs}
\title{Converts percentage to perobabilities}
\usage{
percent2probs(percent)
}
\arguments{
\item{percent}{percentage (number grom 1 to 100)}
}
\value{
Range of probabilities (two numbers) that are nearest to 0.5.
}
\description{
Converts percentage to perobabilities
}
\examples{
percent2probs(100)
percent2probs(50)

}
