% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCR_funs.R
\name{performance0}
\alias{performance0}
\title{[.]  Cusstomizable function `performance0`: function to create performance objects}
\usage{
performance0(prediction.obj, measure, x.measure = "cutoff", ...)
}
\arguments{
\item{prediction.obj}{An object of class \code{prediction}.}

\item{measure}{Performance measure to use for the evaluation. A
    complete list of the performance measures that are available for \code{measure}
    and \code{x.measure} is given in the 'Details' section.}

\item{x.measure}{A second performance measure. If different from the
    default, a two-dimensional curve, with \code{x.measure} taken to be the
    unit in direction of the x axis, and \code{measure} to be the unit in
    direction of the y axis, is created. This curve is parametrized with
    the cutoff.}

\item{...}{Optional arguments (specific to individual performance
    measures).}
}
\value{
An S4 object of class performance.
}
\description{
A more flexible version of \code{\link[ROCR]{performance}}.
}
\examples{

 ## computing a simple ROC curve (x-axis: fpr, y-axis: tpr)
library(ROCR)
data(ROCR.simple)
pred <- prediction( ROCR.simple$predictions, ROCR.simple$labels)
perf <- performance0(pred,"tpr","fpr")
plot(perf)


}
\seealso{
\code{\link[ROCR]{performance}}
}
