% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_performance.R
\name{performance_measures}
\alias{performance_measures}
\alias{get_performance}
\alias{get_max}
\alias{AUC}
\alias{tp}
\alias{fp}
\alias{fn}
\alias{tn}
\alias{cutoff}
\alias{Se}
\alias{Sp}
\alias{J}
\alias{Bac}
\alias{Acc}
\alias{Kappa}
\alias{Wkappa}
\alias{kappa_helper}
\alias{make_conf_matrix}
\title{Performance measures for ROCR::prediction object}
\usage{
get_performance(pred, measure)

get_max(pred, FUN)

AUC(pred)

tp(pred)

fp(pred)

fn(pred)

tn(pred)

cutoff(pred)

Se(pred)

Sp(pred)

J(pred)

Bac(pred)

Acc(pred)

Kappa(pred)

Wkappa(pred)

kappa_helper(pred, FUN_)

make_conf_matrix(TP, FN, FP, TN)
}
\arguments{
\item{pred}{An object of class \code{prediction} from package \pkg{ROCR}.}

\item{measure}{(\code{string(1)})\cr A string with the name of
                classification performance measure to use. Currently
                available options:
\itemize{
      \item \code{"bac"} - for balanced accuracy (mean of sensitivity and specificity);\cr
      \item \code{"kappa"} - for Cohens kappa;\cr
      \item \code{"wkappa"} - for weighted Cohens kappa;\cr
      \item \code{"j"} - for Youden's index;\cr
      \item \code{"auc"} - for area under the ROC curve;\cr
      \item \code{"acc"} - for accuracy (total proportion of correctly identified cases).
}}

\item{FUN}{Function to apply (one of
\code{\link{Acc}},
\code{\link{Bac}},
\code{\link{J}},
\code{\link{Kappa}},
\code{\link{Wkappa}},
 etc.)}

\item{FUN_}{Function to apply (either \code{\link{measure_kappa}} or
\code{\link{measure_wkappa}})}

\item{TP}{Number of true positives.}

\item{FN}{Number of false negatives.}

\item{FP}{Number of false positives.}

\item{TN}{Number of true negatives.}
}
\value{
Function \code{get_performance} returns a numeric vector with 2 elements:
\itemize{
    \item the first element is the \emph{highest} value of selected performance measure;
    \item the second element is either corresponding \bold{cut-off value}, or \code{NA}
if the measure is \code{"auc"}.
}
}
\description{
Function \code{get_performance} calculates the best performance measure when
         \code{prediction} object is given. \cr
}
