% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colors.R
\name{plot_colors}
\alias{plot_colors}
\title{[!] Vizualize colors}
\usage{
plot_colors(col, text = as.character(col), Title = "Colors",
  cex = 1.2, as.legend = FALSE, ...)
}
\arguments{
\item{col}{A list of either color names or color codes.}

\item{text}{Strings that describe each color.}

\item{cex}{character expansion factor \bold{relative} to current
    \code{par("cex")}.  Used for text, and provides the default for
    \code{pt.cex}.}
}
\value{
A plot made with R package \pkg{graphics}.
}
\description{
A convenience function to Vizualize colors by either color name or
color code.
}
\examples{

plot_colors("#ee0000")
plot_colors("#ee0000", "RED color")
plot_colors(c("red","red4","orange2","green3","skyblue"))

}
\seealso{
Other \pkg{spHelper} plots: \code{\link{check_palette}},
  \code{\link{layer_spRangeMean}},
  \code{\link{plot_hyPalette}},
  \code{\link{plot_spCompare}}, \code{\link{plot_spDiff}},
  \code{\link{plot_spDistribution}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{qplot_spRangeCenter}},
  \code{\link{qplot_spRangeMedian}},
  \code{\link{qplot_spStat}}, \code{\link{qplot_spc}},
  \code{\link{rmExpr}}, \code{\link{rm_stripes}},
  \code{\link{stat_chull}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} plots}
