% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confusion.R
\name{plot_confusion}
\alias{plot_confusion}
\alias{plot_confusion2}
\title{[!] Visualize a confusion matrix (classification table)}
\usage{
plot_confusion(conf, Title = "Classification table", xLabel = NULL,
  yLabel = NULL, subTitle = NULL, shades = c("prop", "max", "const",
  "none"), guide = FALSE, text.size = 4, decimals = 2)

plot_confusion2(Prediction, Reference, ...)
}
\arguments{
\item{conf}{A confusion matrix / A classificattion table (either a table
or a square matrix).}

\item{Title}{The main title for the plot.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{subTitle}{Subtitle - second line of a title.}

\item{shades}{A way, how cells are shaded.
\describe{
    \item{prop}{(default) A shade (intensity) of a color is proportianal
          to a value of a cell. Distribution of colors is balanced
          according to number of classes, but not balanced according to
          number of observations per class.}
    \item{max}{Cell with absolute maximum value is represented by the
    most intensive color. Other cells are represented by colors which
    are diluted proportionaly to their value in accordance with the
    maximum value.}
    \item{const}{Constant red and green colors.}
    \item{none}{All cells are grey.}
}}

\item{guide}{Logical. If \code{TRUE}, a legend is plotted.}

\item{text.size}{The size of text inside cells}

\item{decimals}{The number of decimal positions in rounding. Default is 2
(i.e., precission is 0.01).}

\item{Prediction}{A factor variable with \bold{predicted} groups.}

\item{Reference}{A factor variable  with \bold{reference} groups.}

\item{...}{Parameters used in \code{plot_confusion}.}
}
\value{
A plot of confusion matrix and additional statistics (`ggplot` object).
}
\description{
Plot a confusion matrix (classification table) and
 additional statistics.
}
\examples{

# Generate data: Random guess  ============================
 N <- 1000 # number of observations

Prediction <- sample(c("A","B","C","D"), N, replace = TRUE)
Reference  <- sample(c("A", "B","C","D"),N, replace = TRUE)

# This function:
plot_confusion2(Prediction,Reference)

# does the same as:
conf <- table(Prediction,Reference)
plot_confusion(conf)

# At least 50\% of the cases agree =========================
ind <- sample(1:N,round(0.50*N))
Reference[ind] <- Prediction[ind]
conf2 <- table(Prediction,Reference)

plot_confusion(conf2)

# Most of the cases agree =================================
ind <- sample(1:N,round(N*.80))
Reference[ind] <- Prediction[ind]
conf3 <- table(Prediction,Reference)

plot_confusion(conf3)

# Proportions =============================================

plot_confusion(conf3)

# Shades: proportional =====================================

plot_confusion(conf,shades = "prop",  subTitle = "shades: 'prop', correct by chance")
plot_confusion(conf,shades = "max",   subTitle = "shades: 'max', correct by chance")

plot_confusion(conf2,shades = "prop", subTitle = "shades: 'prop', correct >50\%")
plot_confusion(conf2,shades = "max",  subTitle = "shades: 'max', correct >50\%")

plot_confusion(conf3,shades = "prop", subTitle = "shades: 'prop', correct >80\%")
plot_confusion(conf3,shades = "max",  subTitle = "shades: 'max', correct >80\%")

# Shades: constant and none ================================

plot_confusion(conf3,shades = "const",subTitle = "shades: constant")
plot_confusion(conf3,shades = "none", subTitle = "shades: none")


}
\seealso{
Other spHelper plots: \code{\link{InfoDim_plot}},
  \code{\link{plot_SpDiff}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kAmp}}, \code{\link{plot_kSp}},
  \code{\link{plot_stat}}
}

