% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hy.palette.R
\name{plot_hy.palette}
\alias{plot_hy.palette}
\title{[!] Plot the color palette used in a hyperSpec object}
\usage{
plot_hy.palette(sp = NULL, by = ".color", palette = hyGet.palette(sp),
  cex = if (as.legend == FALSE) 1.2 else 1, Title = if (!is.null(sp))
  "Group colors" else "Colors", as.legend = FALSE, xpd = NA,
  x = "topright", y = NULL, bty = "n", ...)
}
\arguments{
\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}

\item{by}{\bold{Either} a name of factor variale in \code{sp} which levels
correspond to colors in \code{palette} \bold{or}, if \code{sp} is
not present, a charter vector of names to be ploted as a text.}

\item{palette}{A list of color names or color codes.}

\item{cex}{character expansion factor \bold{relative} to current
    \code{par("cex")}.  Used for text, and provides the default for
    \code{pt.cex} and \code{title.cex}.}

\item{Title}{The title.}

\item{as.legend}{Logical. If \code{TRUE}, the result is used as legend for
existing R base graphics plot.  If \code{FALSE} (default), a
separate plot is drawn.}

\item{xpd}{A logical value or \code{NA}. If \code{FALSE}, all plotting is
clipped to the plot region, if \code{TRUE}, all plotting is
clipped to the figure region, and if \code{NA}, all plotting is
clipped to the device region.  See also \code{\link[graphics]{clip}}.}

\item{x, y}{the x and y co-ordinates to be used to position the legend.
They can be specified by keyword or in any way which is accepted
by \link[grDevices]{xy.coords}: See 'Details' in \code{\link[graphics]{legend}}.}

\item{bty}{The type of box to be drawn around the legend. The allowed values
are "o" and "n" (the default).}

\item{...}{Other arguments to be passed to \code{\link[graphics]{legend}}.
(except \code{legend}, \code{title} and \code{fill})}
}
\value{
A plot made with R package \pkg{graphics}.
}
\description{
Plot a color palette in a \code{\link[=hyperSpec-class]{hyperSpec}} object,
 which was added with function \code{\link{hyAdd.color}}.
}
\examples{

# Example 1
sp <- hyAdd.color(Spectra3,"gr")
plot_hy.palette(sp, "gr")

#-----------------------------------------------------
# Example 2
plot_hy.palette(by = "RED color", palette = "#ee0000")

#-----------------------------------------------------
# Example 3
# Use data 'sp' from example 1
# The legend does not make sense in this context. It's just an example.

plot(flu)
plot_hy.palette(sp, "gr", as.legend = TRUE, cex = 1)


}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link[graphics]{legend}}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}

Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_colors}}, \code{\link{plot_spCompare}},
  \code{\link{plot_spDiff}}, \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

