% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infoDim.R
\name{plot_infoDim}
\alias{plot_infoDim}
\title{[!] Plot the result of \code{\link{infoDim}} (ggplot2)}
\usage{
plot_infoDim(Object, n.comp.SHOW = 20, selected = NA,
  Title = "Scree Plot", y.log = TRUE, show.legend = TRUE,
  ggtheme = theme_bw())
}
\arguments{
\item{Object}{Either an object (a list), generated by function
\code{\link{infoDim}} or a matrix (to be passed to \code{infoDim}).}

\item{n.comp.SHOW}{A number of components to show on x axis, default is 20.
This number is corrected if (a) vector of eigenvalues is smaller than 20
or (b) information dimension is higher than 15.}

\item{selected}{(optional parameter) A number of components sellected
(will be plotted as a separate vertical line).}

\item{Title}{The title of the plot.}

\item{y.log}{Logical. If \code{TRUE} (default) the scale of  y axis is
logarythmic.}

\item{show.legend}{Logical. If \code{TRUE} (default) the legend is displayed.}

\item{ggtheme}{A function of ggplot2 theme to apply (e.g.: \link[ggplot2]{ggtheme}).
Default is \code{theme_bw()}.}
}
\value{
A scree plot : plot which helps to
        determine the number of nenessary components (e.g. for PCA).\cr
        (A "ggplot" object.)
}
\description{
Plot the result generated by function \code{\link{infoDim}}.
}
\note{
http://www.originlab.com/doc%5Cen/Tutorial/images/Principal_Component_Analysis/Pca_scree_plot.png
}
\examples{
 infoDim(Spectra)

# ------------------------------------------------------
 my_matrix <- matrix(rexp(200, rate=.1), ncol=20)
 my_result <- infoDim(my_matrix)

 # Investigate the result
 str(my_result)
 my_result$exactDim
 my_result$dim

 #Plot
 my_plot <- plot_infoDim(my_result)
 my_plot

}
\seealso{
Other information dimension functions: \code{\link{InfoDim_plot}},
  \code{\link{infoDim}}

Other spHelper plots: \code{\link{InfoDim_plot}},
  \code{\link{plot_SpDiff}}, \code{\link{plot_confusion}},
  \code{\link{plot_kAmp}}, \code{\link{plot_kSp}},
  \code{\link{plot_stat}}
}

