% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kAmp.R
\name{plot_kAmp}
\alias{plot_kAmp}
\title{[!] Plot component amplitudes (a.k.a scores)}
\usage{
plot_kAmp(scores, Title = "Component amplitudes", subTitle = NULL,
  xLabel = labels(scores, ".wavelength"), yLabel = labels(scores, "spc"),
  by = "gr")
}
\arguments{
\item{scores}{- object of class \code{\link[=hyperSpec-class]{hyperSpec}}
with scores after factorisation/decomposition.}

\item{Title}{- Title}

\item{subTitle}{- Second line of title}

\item{xLabel}{- label of x axis}

\item{yLabel}{- label of y axis
@param by A name of grouping variable.}
}
\value{
object of class "ggplot"
}
\description{
Plot component amplitudes (a.k.a scores).
}
\note{
May be incorrect, if cals is not hyperSpec \cr
       xLabel = labels(scores, ".wavelength") \cr
       yLabel
}
\examples{

data(Scores)
plot_kAmp(Scores)

data(Scores3)
plot_kAmp(Scores3,by = "class")
}

