% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2 Fun Plotting Sp.R
\name{plot_kSp}
\alias{plot_kSp}
\title{[!] Plot spectra of components (a.k.a loadings) on one plot.}
\usage{
plot_kSp(loadings, Title = "Components", xLabel = labels(loadings, "spc"),
  yLabel = labels(loadings, ".wavelength"))
}
\arguments{
\item{loadings}{- \code{\link{hyperSpec}} object}

\item{Title}{- the title of graph}

\item{xLabel}{- label of x axis}

\item{yLabel}{- label of ordinate axis}
}
\value{
ggplot2 object
}
\description{
Plot spectra of components (a.k.a loadings) on one plot.
}
\examples{
plot_kSp(loadings)

}

