% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpFun.R
\name{plot_kSpFacets}
\alias{plot_kSpFacets}
\title{Plots spectra of components (a.k.a loadings) in separate graphs}
\usage{
plot_kSpFacets(loadings, Title = "Komponentai (atskirai)",
  ylab = "I, sant.vnt.", normalize = any(loadings$spc < 0) -
  any(loadings$spc > 0))
}
\arguments{
\item{loadings}{- \code{\link{hyperSpec}} object}

\item{Title}{- the title of graph}

\item{ylab}{- label of ordinate axis}

\item{normalize}{- flag if plot normalized components
default: normalize, if needed:
  Below0 <- any(loadings$spc < 0)
  Above0 <- any(loadings$spc > 0)

  if (Above0 - Below0):
  `0` - do not normalize;
 `+1` - normalize to max value;
 `-1` - normalize to min value}
}
\value{
object of class "ggplot2"
}
\description{
Plots spectra of components (a.k.a loadings) in separate graphs
}
\examples{
...

}

