% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LOF_by_class.R
\name{plot_LOF_hist}
\alias{plot_LOF_hist}
\title{Plot histogram of LOF values}
\usage{
plot_LOF_hist(LOF, LOF_threshold = 10, breaks = "Sturges",
  col = "tomato3", x = "topright", y = NULL)
}
\arguments{
\item{LOF}{- either "LOF_obj" object or vector with LOF values.}

\item{breaks}{one of:
    \itemize{
      \item a vector giving the breakpoints between histogram cells,
      \item a function to compute the vector of breakpoints,
      \item a single number giving the number of cells for the histogram,
      \item a character string naming an algorithm to compute the
      number of cells (see \sQuote{Details}),
      \item a function to compute the number of cells.
    }
    In the last three cases the number is a suggestion only; as the
    breakpoints will be set to \code{\link{pretty}} values, the number
    is limited to \code{1e6} (with a warning if it was larger).  If
    \code{breaks} is a function, the \code{x} vector is supplied to it
    as the only argument (and the number of breaks is only limited by
    the amount of available memory).
  }

\item{col}{a colour to be used to fill the bars.
    The default of \code{NULL} yields unfilled bars.}

\item{x, y}{position of legend. See \code{\link[graphics]{legend}}.}
}
\description{
Plot histogram of LOF values
}
\seealso{
Other LOF_obj: \code{\link{get_LOF_by_class}},
  \code{\link{plot_LOF_MDS}}, \code{\link{plot_LOF_sp}},
  \code{\link{plot_stacked}}
}
\concept{LOF_obj}
