% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LOF_by_class.R
\name{plot_LOF_MDS}
\alias{plot_LOF_MDS}
\title{Plot MDS of LOF_obj data}
\usage{
plot_LOF_MDS(obj, type = c("data", "differences", "centers"),
  LOF_threshold = 10, x = "topright", y = NULL, colors = c("black",
  "red"), LOF = NULL)
}
\arguments{
\item{obj}{- either "LOF_obj" object or "hyperspec" object.}

\item{type}{- type of plot, string wit one of the following: "data", "differences", "centers". May be abbreviated.}

\item{LOF_threshold}{- LOF threshold values. Rows (lines) that have threshold above}

\item{x, y}{position of legend. See \code{\link[graphics]{legend}}.}

\item{colors}{Colors names for LOF below and above threshold default are: \code{c("black", "red")}.}

\item{LOF}{- vector of lack-of-fit values for every row;}
}
\description{
Plot MDS of LOF_obj data
}
\examples{
library(spHelper)
obj <- get_LOF_by_class(Spectra2, "class")

plot_LOF_MDS(obj, "data",        x = NULL)
plot_LOF_MDS(obj, "differences", x = NULL)
plot_LOF_MDS(obj, "centers",     x = NULL)

}
\seealso{
Other LOF_obj: \code{\link{get_LOF_by_class}},
  \code{\link{plot_LOF_hist}}, \code{\link{plot_LOF_sp}},
  \code{\link{plot_stacked}}
}
\concept{LOF_obj}
