% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LOF_by_class.R
\name{plot_LOF_sp}
\alias{plot_LOF_sp}
\title{Plot spectra colored by LOF values}
\usage{
plot_LOF_sp(
  obj,
  type = c("data", "differences", "centers"),
  stacked = TRUE,
  LOF_threshold = 10,
  x = "topleft",
  y = NULL,
  colors = c("black", "red"),
  LOF = obj$LOF
)
}
\arguments{
\item{obj}{- \code{LOF_obj} object created with function \code{\link{get_LOF_by_class}}.}

\item{type}{- type of plot, string wit one of the following: "data", "differences", "centers". May be abbreviated.}

\item{stacked}{Either "TRUE" or vector with for stacking.
More details in \code{\link[hyperSpec]{plotspc}}.}

\item{LOF_threshold}{- LOF threshold values. Rows (lines) that have threshold above}

\item{x, y}{position of legend. See \code{\link[graphics]{legend}}.}

\item{colors}{Colors names for LOF below and above threshold default are: \code{c("black", "red")}.}

\item{LOF}{- vector of lack-of-fit values for every row;}
}
\description{
Plot spectra colored by LOF values
}
\examples{
library(spHelper)
obj <- get_LOF_by_class(Spectra2, "gr")

plot_LOF_sp(obj, "data", "gr")
plot_LOF_sp(obj, "differences", "gr")
plot_LOF_sp(obj, "centers", "gr")

}
\seealso{
Other LOF_obj: 
\code{\link{get_LOF_by_class}()},
\code{\link{plot_LOF_MDS}()},
\code{\link{plot_LOF_hist}()},
\code{\link{plot_stacked}()}
}
\concept{LOF_obj}
