% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_by_Ignas_Ciplys.R
\name{plot_multiDim}
\alias{plot_multiDim}
\title{[~!~] Groups dicrimination - test vs. prediction.}
\usage{
plot_multiDim(amp, gr_test, gr_predict, colors, k = 2)
}
\arguments{
\item{amp}{...}

\item{gr_test}{...}

\item{gr_predict}{...}

\item{colors}{...}

\item{k}{...}
}
\value{
...
}
\description{
/No description yet/
}
\examples{
 #NO EXAMPLES YET

}
\author{
Ignas Ciplys
}
\seealso{
Other functions by Ignas Ciplys: \code{\link{save_rds_unique}}
}

