% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ROC_plot_example.R
\name{plot_roc}
\alias{plot_roc}
\title{[.] Plot ROC - sample code}
\source{
\url{https://github.com/joyofdata/joyofdata-articles/blob/master/roc-auc/plot_roc.R}
}
\usage{
plot_roc(roc, threshold, cost_of_fp, cost_of_fn)
}
\arguments{
\item{roc}{...}

\item{threshold}{...}

\item{cost_of_fp}{...}

\item{cost_of_fn}{...}
}
\value{
...
}
\description{
[.] Plot ROC - sample code
}
\examples{
...
}
