% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spCompare.R
\name{plot_spCompare}
\alias{plot_spCompare}
\title{[!] Compare 2 spectroscopic signals}
\usage{
plot_spCompare(sp1, sp2, row = 1, colors = c("green4", "blue3", "red"),
  show.legend = TRUE, legend.title = "Spectra",
  legend.text = c(match.call()$sp1, match.call()$sp2, "Difference"),
  x = "topright", lwd = 1, ...)
}
\arguments{
\item{sp1, sp2}{hyperSpec objects of the same size}

\item{row}{An integer, that indicates which row must be plotted.}

\item{colors}{Vector of 3 colors for original, filtered and noise signals
respectively.}

\item{show.legend}{Logical}

\item{legend.title}{The title of the legend.}

\item{legend.text}{Character vertor of length = 3. The entries for the legend
for for original, filtered and noise signals respectively.}

\item{x, y}{the x and y co-ordinates to be used to position the legend.
They can be specified by keyword or in any way which is accepted
by \link[grDevices]{xy.coords}: See 'Details' in \code{\link[graphics]{legend}}.}

\item{lwd}{A line width. More details in \link[graphics]{par}.}

\item{...}{Other arguments to be passed to \code{\link[graphics]{legend}}
(except \code{legend}, \code{title}, \code{col} and \code{lty}).}
}
\value{
A plot made with R package \pkg{graphics}.
}
\description{
Compare 2 spectroscopic signals: plot 2 signals and their difference.
}
\examples{

# Apply running medians filter:
sp_filt <- apply(Spectra, 1, function(x) {runmed(x, 15)})

plot_spCompare(Spectra, sp_filt, row = 2)

# Modify the legend:
plot_spCompare(Spectra, sp_filt, row = 2,legend.text  = c("Original","Filtered","Noise"))

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_colors}}, \code{\link{plot_hy.palette}},
  \code{\link{plot_spDiff}}, \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

