% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SpDiff.R
\name{plot_SpDiff}
\alias{plot_SpDiff}
\title{[!] Plot Remainders After Subtracting Components}
\usage{
plot_SpDiff(loadings, scores, Spectra,
  Title = "Remainders After Subtracting Components", color = if (".color"
  \%in\% ls(Spectra$..)) Spectra$.color else "tan3", stacked = if ("ID" \%in\%
  ls(Spectra$..)) Spectra$ID else NULL, spc.nmax = 2000)
}
\arguments{
\item{loadings}{loadings}

\item{scores}{scores}

\item{Spectra}{Spectra (hyperSpec object)}

\item{Title}{Title of the plot.}

\item{color}{...}

\item{stacked}{...}

\item{spc.nmax}{max number of spectra to plot}
}
\value{
Plot of calculated difference between expected (original)
and reconstructed spectra.
}
\description{
Plot difference between experimental and reconstructed spectra \cr \cr
}
\details{
Plot difference between experimental (original) and reconstructed spectra.
Uses function \code{\link{getReconstructed}}, to calculate the reconstructed
spectra and subtracts it from original spectra. \cr

Difference between experimental and reconstructed spectra'
}
\examples{

plot_SpDiff(Loadings, Scores, Spectra)

}

