% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spDistribution.R
\name{plot_spDistribution}
\alias{plot_spDistribution}
\title{[!] Distribution of spectroscopoc data as percentiles}
\usage{
plot_spDistribution(sp, position.legend = "topright",
  show.legend = TRUE, cex.legend = 1, ...)
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{position.legend}{Position of the legend. More details in \link[graphics]{legend}}

\item{show.legend}{Logical. If \code{TRUE}, the legend is shown.}

\item{cex.legend}{Size of the legend. more details in \link[graphics]{par}.}

\item{...}{further parameters for \code{\link[hyperSpec]{plotspc}}.}
}
\value{
a plot
}
\description{
[!] Distribution of spectroscopoc data as percentiles
}
\examples{

plot_spDistribution(laser)
plot_spDistribution(flu, cex.legend = .7)
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{check_palette}},
  \code{\link{layer_spRangeMean}},
  \code{\link{plot_colors}}, \code{\link{plot_hyPalette}},
  \code{\link{plot_spCompare}}, \code{\link{plot_spDiff}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{qplot_spRangeCenter}},
  \code{\link{qplot_spRangeMedian}},
  \code{\link{qplot_spStat}}, \code{\link{qplot_spc}},
  \code{\link{rmExpr}}, \code{\link{rm_stripes}},
  \code{\link{stat_chull}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} plots}
