% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spFilt.R
\name{plot_spFilt}
\alias{plot_spFilt}
\title{[!] Plot original, filtered signals and the noise}
\usage{
plot_spFilt(sp, sp_filt, ind = 1, colors = c("green4", "blue1", "red"),
  show.legend = TRUE, legend.title = "Signals",
  legend.text = c("Original", "Filtered", "Noise"), lwd = 1)
}
\arguments{
\item{sp, sp_filt}{hyperSpec objects of the same size}

\item{ind}{An integer, that indicates which row must be plotted.}

\item{colors}{Vector of 3 colors for original, filtered and noise signals
respectively.}

\item{show.legend}{Logical}

\item{legend.title}{The title of the legend.}

\item{legend.text}{Character vertor of length = 3. The entries for the legend
for for original, filtered and noise signals respectively.}
}
\value{
A plot made with R \code{base} graphics system.
}
\description{
Plot 2 signals and their difference (the noise).
}
\examples{

# Construct and apply running medians filter
medianFilt <- function(x) {runmed(x, 15)}
   sp_filt <- apply(Spectra, 1, medianFilt)

plot_spFilt(Spectra, sp_filt, ind = 2)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_colors}}, \code{\link{plot_hy.palette}},
  \code{\link{plot_spDiff}}, \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

