% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LOF_by_class.R
\name{plot_stacked}
\alias{plot_stacked}
\title{Plot stacked spectra}
\usage{
plot_stacked(obj, stacked = TRUE, spc.nmax = nrow(obj), ...)
}
\arguments{
\item{obj}{- hyperSpec object.}

\item{stacked}{if not \code{NULL}, a "stacked" plot is produced, see the
example. \code{stacked} may be \code{TRUE} to stack single spectra.  A
numeric or factor is interpreted as giving the grouping, character is
interpreted as the name of the extra data column that holds the groups.}

\item{spc.nmax}{maximal number of spectra to be plotted (to avoid
accidentally plotting of large numbers of spectra).}

\item{...}{- further arguments to `plotspc()`.}
}
\description{
Convenience function to plot stacked spectra.
More details in \code{\link[hyperSpec]{plotspc}}
}
\examples{
library(spHelper)
plot_stacked(Spectra2,"gr")

}
\seealso{
Other LOF_obj: 
\code{\link{get_LOF_by_class}()},
\code{\link{plot_LOF_MDS}()},
\code{\link{plot_LOF_hist}()},
\code{\link{plot_LOF_sp}()}
}
\concept{LOF_obj}
