% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stat.R
\name{plot_stat}
\alias{plot_stat}
\alias{qplotStat}
\title{[!] Plot a summary statistic of spectroscopic data by group}
\usage{
plot_stat(sp, by = stop("Argument 'by' is missing."),
  FUN = stop("Argument 'FUN' is missing."),
  Title = fCap(as.character(match.call()$FUN)), subTitle = NULL,
  All = TRUE, fixed.colors = TRUE, All.color = "black",
  gr.color = RColorBrewer::brewer.pal(8, "Dark2"), All.linetype = "dashed",
  gr.linetype = "solid", All.size = 1.1, gr.size = 0.8,
  legend.title = element_blank())

qplotStat(..., FUN = mean)
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}

\item{by}{A grouping variable (\code{by <- as.factor(by)})}

\item{FUN}{A function that calculates one or several summary statistics.}

\item{Title}{A title of the plot.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{All}{Logical. If \code{TRUE}, plot additional corve for a statistic
of all spectra (\code{.ALL}), not devided to grroups.}

\item{fixed.colors}{- Logical. If \code{TRUE}, color scheme where \code{.ALL}
is plotted as a black line is be used.}

\item{All.color}{A line color for \code{.ALL}.}

\item{gr.color}{Line colors (a vector) for each level in \code{by}.}

\item{All.linetype}{A line type for \code{.ALL}.}

\item{gr.linetype}{Line types (a vector) for each level in \code{by}.}

\item{All.size}{A line width for \code{.ALL}.}

\item{gr.size}{Line widths (a vector) for each level in \code{by}.}

\item{legend.title}{A title for legend. Default is no title.}

\item{...}{(The same parameters as above).}
}
\value{
A plot.\cr (An object of class \code{ggplot} and
        \code{gg}. More details in package \pkg{ggplot2}).
}
\description{
Plot a summary of spectroscopic data in hyperSpec object by group.
The summary statistic internally is generated by either function
\code{\link{spStat}} or \code{\link[hyperSpec]{aggregate}}
}
\examples{


plot_stat(chondro,clusters,mean)
plot_stat(chondro,clusters,mean,All=FALSE)
plot_stat(chondro,clusters,mean_sd,All=FALSE) + facet_grid(.~clusters)

plot_stat(chondro,clusters,median,All=FALSE, fixed.colors=FALSE)
plot_stat(chondro,clusters,median, "My Title")


# Use `.aggregate` in making facets, to avoid facet called "NA":

plot_stat(chondro,clusters,mean_pm_sd) + facet_grid(.~clusters)
plot_stat(chondro,clusters,mean_pm_sd) + facet_grid(.~.aggregate)

}
\seealso{
\code{\link{spStat}}

Other spHelper plots: \code{\link{InfoDim_plot}},
  \code{\link{plot_SpDiff}}, \code{\link{plot_confusion}},
  \code{\link{plot_kAmp}}, \code{\link{plot_kSp}}
}

