% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpFun.R
\name{PlotConfusion}
\alias{PlotConfusion}
\title{Plots a confusion matrix, \code{PlotConfusion}}
\usage{
PlotConfusion(conf)
}
\arguments{
\item{conf}{- table, a square matrix}
}
\value{
...
}
\description{
Plots a confusion matrix, \code{PlotConfusion}
}
\examples{
Prediction <- sample(c("A","B","C"),1000, replace = T)
Reference  <- sample(c("A","B","C"),1000, replace = T)
conf <- table(Prediction,Reference)


PlotConfusion(conf)
PlotConfusion(prop.table(conf,1))
PlotConfusion(prop.table(conf,2))

}

