% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_PAP_RK_2014.R
\name{prepare_PAP_RK_2014__MATLAB_failui}
\alias{prepare_PAP_RK_2014__MATLAB_failui}
\title{[+] Transform dataset "PAP_RK_2014" and add labels}
\usage{
prepare_PAP_RK_2014__MATLAB_failui(sp, language = "EN")
}
\arguments{
\item{sp}{PAP_RK_2014 data, created by function \code{\link{read3csv2hy}} as a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{language}{A string, indicating a language of labels. Possible
entries are \code{EN} - English and \code{LT} - Lithuanian.
Default is \code{EN}.}
}
\value{
A transformed and labeled \code{hyperSpec} object.
}
\description{
Function is designed to transform data set (hyperSpec object) called
"PAP_RK_2014" and labels to it.
}
\details{
Select data columns, that are not removed, and add labels to
 \code{\link[=hyperSpec-class]{hyperSpec}} object of "PAP_RK_2014" data.
}
\examples{

# prepare_PAP_RK_2014(sp)

}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
