% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDuration.R
\name{printDuration}
\alias{printDuration}
\title{Print duration (difference of time)}
\usage{
printDuration(StartAt, Message = "Duration of analysis:")
}
\arguments{
\item{StartAt}{Time, when period of time started
(usually generated with \code{Sys.time()})}

\item{Message}{Message before time stamp. Default is
\code{"Duration of analysis:"}}
}
\value{
Text with time duration printed in console.
}
\description{
Print duration (difference of time)
}
\examples{
StartAt <-  Sys.time()

printDuration(StartAt)
printDuration(StartAt,"From start till now")

}

