% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_crosstab.R
\name{qplot_crosstab}
\alias{qplot_crosstab}
\alias{qplot_crosstab.table}
\alias{qplot_crosstab.numeric}
\alias{qplot_crosstab.hyperSpec}
\alias{qplot_crosstab.data.frame}
\alias{qplot_crosstab.matrix}
\alias{qplot_crosstab_sort}
\alias{qplot_crosstab0}
\alias{qplot_crosstab0s}
\title{[!.] Plot a cross-tabulation (classification table)}
\usage{
qplot_crosstab(obj, ...)

\method{qplot_crosstab}{table}(
  tabl,
  Title = "Cross-tabulation",
  xLabel = NULL,
  yLabel = NULL,
  subTitle = NULL,
  text.size = 5,
  sort = FALSE,
  bg.color = "skyblue1",
  max.color = "gold3",
  low.color = "grey80",
  decimals = 1,
  show.max = c("max", "colMax", "rowMax", FALSE),
  shades = TRUE,
  guide = c("legend", "colorbar", FALSE),
  zero.color = "grey60",
  text.color = "black",
  as.percentage = c(FALSE, TRUE, "row", "col"),
  ...
)

\method{qplot_crosstab}{numeric}(
  varX,
  varY,
  Title = "Cross-tabulation",
  xLabel = match.call()$varX,
  yLabel = match.call()$varY,
  ...
)

\method{qplot_crosstab}{hyperSpec}(
  obj,
  varX,
  varY,
  Title = "Cross-tabulation",
  xLabel = NULL,
  yLabel = NULL,
  ...
)

\method{qplot_crosstab}{data.frame}(
  obj,
  varX,
  varY,
  Title = "Cross-tabulation",
  xLabel = match.call()$varX,
  yLabel = match.call()$varY,
  ...
)

\method{qplot_crosstab}{matrix}(obj, Title = NULL, ...)

qplot_crosstab_sort(..., sort = "maxOnDiag", show.max = TRUE, shades = TRUE)

qplot_crosstab0(
  tabl,
  ...,
  sort = FALSE,
  bg.color = "wheat2",
  decimals = 2,
  show.max = FALSE,
  shades = FALSE,
  guide = FALSE
)

qplot_crosstab0s(
  tabl,
  ...,
  Title = "Cross-tabulation",
  xLabel = NULL,
  yLabel = NULL,
  subTitle = NULL,
  text.size = 5,
  sort = "maxOnDiag",
  bg.color = "wheat2",
  decimals = 2,
  show.max = FALSE,
  shades = FALSE,
  guide = FALSE
)
}
\arguments{
\item{...}{(The same parameters as above).}

\item{tabl}{A classification table, crosstabulation: either an object of
a class "table" or a square matrix.}

\item{Title}{The main title for the plot.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{text.size}{The size of text inside cells.}

\item{sort}{Default is \code{FALSE} matrix is plotted as is.
Other options: \cr
       Either \code{"maxOnDiag"} or \code{TRUE} - function
       \code{\link{sort_descOnDiag}} is applied (rows and columns are sorted
       in so that maximum values of frequencies were on diagonal, if it's
       possible);\cr
        \code{"rowSums"} - sort rows by row sums;\cr
        \code{"colSums"} - sort columns by column sums;\cr
        \code{"colMax"}  - sort rows in respect only to maximum value in
        each row;\cr
        \code{"rowMax"}  - sort columns in respect only to maximum value in
        each column.}

\item{bg.color}{(THIS PARAMETER DO NOT WORK if \code{as.percentage} is not
\code{FALSE}.)
 The main background color (if \code{shade = TRUE}, this color
 is used for cells with high values).}

\item{max.color}{A background for maximum values.}

\item{low.color}{A background for low values.}

\item{decimals}{The number of decimal positions in rounding. Default is 2
(i.e., precission is 0.01).}

\item{show.max}{The rule to highlighted top values. Possible entries:\cr
\code{"max"} highlight maximum of whole matrix;\cr
\code{"colMax"} (default) highlight maxima of every column ;\cr
\code{"rowMax"} highlight maxima of every row;\cr
\code{FALSE} maxima are \bold{not} highlighted.}

\item{shades}{Logical. If \code{TRUE}, color of a cell varies depending
not on its value (except the cells with top values if \code{show.max}
is \code{FALSE}).}

\item{guide}{A type of guide to display: either \code{"legend"},
\code{"colorbar"} or \code{FALSE} (if no guide is needed).}

\item{zero.color}{Color of text inside cells, that have value equal to zero.}

\item{as.percentage}{Show result as percentage:\cr
\code{TRUE} - show as fraction of total sum;\cr
\code{FALSE} (default) - show as frequency/counts;\cr
\code{"col"} column-wise persentage (sum of values of each column is 100\%);\cr
\code{"row"} row-wise persentage (sum of values of each row is 100\%).}

\item{Color}{of text inside cells, that have value not equal to zero.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
       More details in package \pkg{ggplot2}.
}
\description{
Plot a crosstabulation (classification table)
}
\examples{

 # BUG for this data: when 'sort = TRUE',
 #              IPPN LSIL HSIL
 #  H.Cluster 1    0    0  111
 #  H.Cluster 2    0    0   45
 #  H.Cluster 3    0    0   49
 #  H.Cluster 4    0    0    5

library(spHelper)


# Generate data: Random guess  ============================
 N <- 1000 # number of observations

Prediction <- sample(1:4, N, replace = TRUE)
Reference  <- sample(c("A", "B","C","D","E"),N, replace = TRUE)

tabl <- table(Prediction,Reference)
qplot_crosstab(tabl)
qplot_crosstab_sort(tabl)   # different order of columns and rows
qplot_crosstab0(tabl)       # no colors
qplot_crosstab0s(tabl)      # no colors, different order of columns and rows


qplot_crosstab(Prediction, Reference, as.percentage = FALSE, sort = TRUE)
qplot_crosstab(Prediction, Reference, as.percentage = "row", sort = TRUE)
qplot_crosstab(Prediction, Reference, as.percentage = "col", sort = TRUE)
qplot_crosstab(Prediction, Reference, as.percentage = TRUE, sort = TRUE)

}
\seealso{
To plot classification table (confusion matrix) use
 \code{\link{qplot_confusion}}.

Other \pkg{spHelper} plots: 
\code{\link{check_palette}()},
\code{\link{layer_spRangeMean}()},
\code{\link{plot_colors}()},
\code{\link{plot_hyPalette}()},
\code{\link{plot_spCompare}()},
\code{\link{plot_spDiff}()},
\code{\link{plot_spDistribution}()},
\code{\link{qplot_confusion}()},
\code{\link{qplot_infoDim}()},
\code{\link{qplot_kAmp}()},
\code{\link{qplot_kSp}()},
\code{\link{qplot_prediction}()},
\code{\link{qplot_spRangeCenter}()},
\code{\link{qplot_spRangeMedian}()},
\code{\link{qplot_spStat}()},
\code{\link{qplot_spc}()},
\code{\link{rmExpr}()},
\code{\link{rm_stripes}()},
\code{\link{stat_chull}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} plots}
