% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_kAmp.R, R/qplot_scores.R
\name{qplot_kAmp}
\alias{qplot_kAmp}
\alias{qplot_scores}
\title{[!+] Plot amplitudes (a.k.a scores) of spectroscopic components}
\usage{
qplot_kAmp(scores, Title = "Component amplitudes", subTitle = NULL,
  xLabel = labels(scores, ".wavelength"), yLabel = labels(scores, "spc"),
  by = "gr", palette = hyGet.palette(scores), add.violin = TRUE,
  add.jitter = FALSE, add.boxplot = TRUE, violin.alpha = 0.25,
  jitter.alpha = 0.3, jitter.size = 1)

qplot_scores(..., by = NULL, add.violin = FALSE)
}
\arguments{
\item{scores}{An object of class \code{\link[=hyperSpec-class]{hyperSpec}}
with factor scores, principal component scores, component amplitudes, etc.}

\item{Title}{The main title for the plot.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{by}{A name of grouping variable. If \code{NULL}, all data is ploted.
Default is \code{by = "gr"} for \code{qplot_kAmp} and
 \code{by = NULL} for \code{qplot_scores}.}

\item{palette}{A color palette to be used in plotting. (...)}

\item{add.violin}{Logical. If \code{TRUE} adds so called violin (i.e.,
symmetrical probarility density) plot.
 Default is \code{TRUE} for \code{qplot_kAmp}.}

\item{add.jitter}{Logical. If \code{TRUE} adds jitter plot.
Default is \code{FALSE}.}

\item{add.boxplot}{Logical. If \code{TRUE} adds boxplot.
Default is \code{TRUE}.}

\item{violin.alpha}{Transperency of violin plot.}

\item{jitter.alpha}{Transperency of jitter plot.}

\item{jitter.size}{Point size in jitter plot.}

\item{...}{(The same parameters as above).}
}
\value{
A plot: an object of class \code{\link[ggplot2]{ggplot}}.
       More details in package \pkg{ggplot2}).
}
\description{
Plot amplitudes (a.k.a scores) of spectroscopic components
 grouped by categorical variable \code{by}. Amplitudes are provided as
\code{\link[=hyperSpec-class]{hyperSpec}} object.\cr
}
\details{
Plots are drawn in this order: violin plot, jitter plot, boxplot.
}
\note{
If more information on matrix decomposition/factorisation is
needed it can be found
\href{https://en.wikipedia.org/wiki/Matrix_decomposition}{here} or
\href{http://www.r-bloggers.com/matrix-factorization/}{here}).
}
\examples{

data(Scores)
qplot_kAmp(Scores)
qplot_scores(Scores)

data(Scores3)
qplot_kAmp(Scores3, by = "class")
#'
sp3 <- hyAdd.color(Scores3, "class")
qplot_kAmp(sp3, by = "class")

p <- qplot_scores(Scores, add.jitter = TRUE)
p

p + theme_bw()

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_colors}}, \code{\link{plot_hy.palette}},
  \code{\link{plot_spDiff}}, \code{\link{plot_spFilt}},
  \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kSp}},
  \code{\link{qplot_prediction}}, \code{\link{rmExpr}},
  \code{\link{stat_chull}}

Other component analysis / factorisation related functions in \pkg{spHelper}: \code{\link{getScores}},
  \code{\link{infoDim}}, \code{\link{plot_spDiff}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kSp}},
  \code{\link{reconstructSp}}, \code{\link{sortLoadings}},
  \code{\link{unipeak}}, \code{\link{whichOutlier}}
}

