% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_kSp.R, R/qplot_kSpFacets.R, R/qplot_sp.R
\name{qplot_kSp}
\alias{qplot_kSp}
\alias{qplot_kSpFacets}
\alias{qplot_sp}
\title{[+] Plot spectroscopic curves and spectral components (a.k.a. loadings)}
\usage{
qplot_kSp(loadings, Title = "Components", xLabel = labels(loadings,
  ".wavelength"), yLabel = labels(loadings, "spc"), names.in = "kNames",
  legendName = FALSE, filled = TRUE, normalize = FALSE, facets = FALSE,
  subTitle = NULL)

qplot_kSpFacets(..., Title = "Components (stacked)", normalize = "auto",
  facets = TRUE)

qplot_sp(..., filled = FALSE)
}
\arguments{
\item{loadings}{Components' spectra (loadings) as a \code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{Title}{The main title for the plot.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{names.in}{A name of variable in \code{loadings} that contains variable names.
If indicated variable does not exist, row names are used instead.\cr
Default is \code{names = 'kNames'}}

\item{legendName}{A name of a legend. Possible entries: \enumerate{
\item {logical \code{FALSE}}{ - a legend without a name;}
\item {logical \code{TRUE}}{ - a label of a variable \code{name} is used as a name of a legend
    (\code{legendName <- labels(loadings,name}));}
\item {...}{manual input of the name.}
}}

\item{filled}{Logical. If \code{TRUE}, colored fill is used. If \code{FALSE}, no fill is used.
(Fill is an area between ordinate axis and the curve.)}

\item{normalize}{A flag that indicates whether components should be
normalized before plotting.
    Possible selections: \enumerate{
    \item {\code{FALSE}, \code{0}}{ - do not normalize;}
    \item {\code{TRUE}, \code{+1}}{ - normalize to max value;}
    \item {\code{-TRUE}, \code{-1}}{ - normalize to min value [! this choise can give unexpected results];}
    \item {\code{"auto"}}{ - one of the choices above is selected by determining if spectra have any
         possitive and any negative peaks:\itemize{
         \item \code{Below0 <- any(loadings$spc < 0);}
         \item \code{Above0 <- any(loadings$spc > 0);}
         \item \code{normalize <- (Above0 - Below0)}
         }
         }
    }}

\item{facets}{A logical flag. If \code{TRUE}, spectra are plotted on separate graphs/facets
       (implemented by function \code{\link[ggplot2]{facet_grid}}). If {\code{FALSE}, all spectra
        are plotted on one facet.
}}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{...}{(The same parameters as above).}
}
\value{
A plot: an object of class \code{\link[ggplot2]{ggplot}}.
       More details in package \pkg{ggplot2}).
}
\description{
Plot spectroscopic curves in different colors.

             Functions \code{qplot_kSp}, \code{qplot_kSpFacets} are designed
             to plot spectra of spectral components
            (a.k.a. loadings), extracted in principal component analysis
            (e.g., \code{\link[stats]{princomp}}), variaus matrix factorization
             methods (e.g., \code{\link[NMF]{nmf}})) and some other dimension
             reduction methods. \cr
             The function \code{qplot_sp} is convenient for all spectroscopic
             curves as uses no fill.
}
\details{
\code{qplot_kSp} plots spectra on one graph. \cr
         \code{qplot_kSpFacets} plots spectra on separate graphs (facets).\cr
}
\note{
Before using as an argument \code{loadings},
         a matrix of components/loadings (\code{loadings}) must be treated
         with function \code{\link[hyperSpec]{decomposition}} (or equivalent)
         which converts it to \code{\link[=hyperSpec-class]{hyperSpec}} object.
}
\examples{

qplot_kSp(Loadings)

data(flu, package = "hyperSpec")

qplot_kSpFacets(flu, Title = "Flu dataset")
qplot_kSpFacets(flu, Title = "Flu dataset", normalize = 1)
qplot_kSpFacets(flu, Title = "Flu dataset", normalize = FALSE)
qplot_kSpFacets(flu, Title = "Flu dataset", normalize = -1)

## Remove fill -----------------------------------------------------------------


qplot_kSp(flu, filled = FALSE)
qplot_sp(flu)

## Name of a legend ------------------------------------------------------------
flu$c2 <- as.factor(flu$c)

qplot_sp(flu, Title = "Flu dataset", names.in = 'c2', legendName = FALSE)
qplot_sp(flu, Title = "Flu dataset", names.in = 'c2', legendName = TRUE)
qplot_sp(flu, Title = "Flu dataset", names.in = 'c2', legendName = "Concentration")

## Example of line color transitions -------------------------------------------
  qplot_sp(laser)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_spDiff}}, \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_prediction}}, \code{\link{rmExpr}},
  \code{\link{stat_chull}}

Other component analysis / factorisation related functions: \code{\link{getScores}},
  \code{\link{infoDim}}, \code{\link{plot_spDiff}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{reconstructSp}}, \code{\link{sortLoadings}},
  \code{\link{unipeak}}, \code{\link{whichOutlier}}
}

