% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_prediction.R, R/qplot_proximity.R
\name{qplot_prediction}
\alias{qplot_prediction}
\alias{qplot_proximity}
\title{[!!!] Proximity plot created using multi-dimensiolal scaling (MDS)}
\usage{
qplot_prediction(
  scores,
  Prediction,
  Reference,
  xLabel = paste("Projection", xproj),
  yLabel = paste("Projection", yproj),
  Title = "Proximity of Groups",
  subTitle = NULL,
  palette = hyGet_palette(scores),
  type = c("Prediction", "Reference"),
  stat = c("chull", "ellipse", "none"),
  type.stat = type,
  MDS = c("metric", "isoMDS"),
  plot.scatter = TRUE,
  alpha = 0.5,
  k = 2,
  xproj = 1,
  yproj = 2
)

qplot_proximity(
  scores,
  by,
  xLabel = paste("Projection", xproj),
  yLabel = paste("Projection", yproj),
  Title = "Proximity of Groups",
  subTitle = NULL,
  palette = hyGet_palette(scores),
  stat = c("chull", "ellipse", "none"),
  MDS = c("metric", "isoMDS"),
  plot.scatter = TRUE,
  alpha = 0.5,
  k = 2,
  xproj = 1,
  yproj = 2
)
}
\arguments{
\item{scores}{Components' amplitudes (scores) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{Prediction}{Predicted groups. Either a variable name or a vector.}

\item{Reference}{Reference grpups. Either a variable name or a vector.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{Title}{The main title for the plot.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{palette}{A color palete to be used. Default is `ggplot2` default palette.}

\item{type}{A type of data for scatterplot: either "prediction" (default)
or "reference".}

\item{stat}{A statistic to be added, either "chull" - convex hull (default),
"ellipse" or "none"}

\item{type.stat}{A type of data for \code{stat}: either  "prediction"
(default) or "reference".}

\item{MDS}{A type of Multi-Dimensional Scaling.}

\item{plot.scatter}{Logical. If \code{true} (default), scatterplot
(`geom_point`)is plotted.}

\item{alpha}{Transperency of points.}

\item{k}{
  The desired dimension for the solution, passed to \code{cmdscale}.
}

\item{xproj, yproj}{An integer: a number of MDS projection to plot on x, y axis
(respectively).}

\item{by}{A grouping variable}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
       More details in package \pkg{ggplot2}.
}
\description{
Plot a 2D scatterplot that represents multidimensionally scaled data.
}
\examples{

# Examples with a `hyperSpec` object:

data(Scores2)
Scores2$Prediction <- sample(Scores2$gr)
Scores2 <- hyAdd_color(sp = Scores2, by = "gr", palette = c("tan3", "green4","skyblue"))

qplot_prediction(Scores2,Prediction = "Prediction", Reference = "gr")
qplot_prediction(Scores2,Prediction = "Prediction", Reference = "gr", type = "ref")

qplot_prediction(Scores2,"Prediction","gr", type.stat = "ref", MDS = "isoMDS")


sc <- Scores2[,,c(1,3),wl.index = TRUE]


qplot_prediction(sc,"Prediction","gr", type = "reference")
qplot_prediction(sc,"Prediction","gr", type = "prediction")
qplot_prediction(sc,"Prediction","gr", type = "prediction", type.stat = "ref")
qplot_prediction(sc,"Prediction","gr", type = "prediction", type.stat = "ref",stat ="ellipse")


sc <- Scores2[,,c(1,2),wl.index = TRUE]
sc <- hyAdd_color(sp = sc , by = "class", palette = c("tan3", "green4","skyblue","violet"))
sc$ID <- rownames(sc)

qplot_proximity(sc, "class")
qplot_proximity(sc, "class",  plot.scatter = FALSE) + geom_text(aes(label = sc$ID))

}
\seealso{
Other \pkg{spHelper} plots: 
\code{\link{check_palette}()},
\code{\link{layer_spRangeMean}()},
\code{\link{plot_colors}()},
\code{\link{plot_hyPalette}()},
\code{\link{plot_spCompare}()},
\code{\link{plot_spDiff}()},
\code{\link{plot_spDistribution}()},
\code{\link{qplot_confusion}()},
\code{\link{qplot_crosstab}()},
\code{\link{qplot_infoDim}()},
\code{\link{qplot_kAmp}()},
\code{\link{qplot_kSp}()},
\code{\link{qplot_spRangeCenter}()},
\code{\link{qplot_spRangeMedian}()},
\code{\link{qplot_spStat}()},
\code{\link{qplot_spc}()},
\code{\link{rmExpr}()},
\code{\link{rm_stripes}()},
\code{\link{stat_chull}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} plots}
