% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_spDistrib.R
\name{qplot_spDistrib}
\alias{qplot_spDistrib}
\title{[!] Plot distributions of spectroscopic data}
\usage{
qplot_spDistrib(sp, by = stop("`by` is missing"), percent = NULL,
  probs = NULL, palette = hyGet_palette(sp))
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{by}{A vector (factor variable) with indicated groups for each case:
either variable name inside the object \code{sp}, or a vector
of length \code{nrow(sp)}.}

\item{percent}{Numeric value between 0 and 100 that indicates
\emph{percentage} of samples nearest to median be plotted as a ribbon.}

\item{probs}{verctor of size 2 with values between 0 and 1 that indicates
\emph{probabilities} at which percentiles should be plotted.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
       More details in package \pkg{ggplot2}.
}
\description{
[!] Plot distributions of spectroscopic data
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} plots

 @examples
qplot_spDistrib(Spectra2, "gr")

# Both lines below gives identical plots with different default titles:
qplot_spDistrib(Spectra2, "gr", percent = 50)
qplot_spDistrib(Spectra2, "gr", probs = c(.25, .75))}
