% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_spStat.R
\name{qplot_spStat}
\alias{qplot_spStat}
\alias{layer_spStat}
\title{[+] Plot a summary statistic of spectroscopic data by group}
\usage{
qplot_spStat(sp, by = stop("Argument 'by' is missing."),
  FUN = stop("Argument 'FUN' is missing."),
  Title = fCap(as.character(match.call()$FUN)), subTitle = NULL,
  All = FALSE, fixed.colors = All, All.color = "black",
  gr.color = c(hyGet_palette(sp), RColorBrewer::brewer.pal(8, "Dark2")),
  All.linetype = "dashed", gr.linetype = "solid", All.size = 1.1,
  gr.size = 0.8, legend.title = element_blank(), ..., add = FALSE)

layer_spStat(sp, by = stop("Argument 'by' is missing."),
  FUN = stop("Argument 'FUN' is missing."),
  Title = fCap(as.character(match.call()$FUN)), subTitle = NULL,
  All = FALSE, ..., add = TRUE)
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}

\item{by}{A grouping variable (either a vector or a variable name in \code{sp}).}

\item{FUN}{A function that calculates one or several summary statistics.}

\item{Title}{The main title of the plot.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{All}{Logical. If \code{TRUE}, plot additional curve for a statistic
of all spectra (\code{.ALL}), not divided to groups.}

\item{fixed.colors}{- Logical. If \code{TRUE}, color scheme where
\code{.ALL} is plotted as a black line is be used.}

\item{All.color}{A line color for \code{.ALL}.}

\item{gr.color}{Line colors (a vector) for each level in \code{by}.}

\item{All.linetype}{A line type for \code{.ALL}.}

\item{gr.linetype}{Line types (a vector) for each level in \code{by}.}

\item{All.size}{A line width for \code{.ALL}.}

\item{gr.size}{Line widths (a vector) for each level in \code{by}.}

\item{legend.title}{A title for legend. Default is no title.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
       More details in package \pkg{ggplot2}.
}
\description{
Plot a summary of spectroscopic data in hyperSpec object by group.
The summary statistic internally is generated by either function
\code{\link{spStat}} or \code{\link[hyperSpec]{aggregate}}.
}
\examples{


qplot_spStat(chondro,"clusters",mean)
qplot_spStat(chondro,"clusters",mean,All=FALSE)
qplot_spStat(chondro,"clusters",mean_sd,All=FALSE) + facet_grid(.~clusters)

qplot_spStat(chondro,"clusters",median,All=FALSE, fixed.colors=FALSE)
qplot_spStat(chondro,"clusters",median, "My Title")


# Make facets with fewer ticks on the x axis:
qplot_spStat(chondro,"clusters",mean_pm_sd) +
     facet_grid(.~clusters) +
     nTick_x(2)


# Add as a layer:
qplot_spRange(chondro, "clusters") +
     layer_spStat(chondro,"clusters", mean, size = 1)


}
\seealso{
\code{\link{spStat}}

Other \pkg{spHelper} plots: \code{\link{check_palette}},
  \code{\link{layer_spRangeMean}},
  \code{\link{plot_colors}}, \code{\link{plot_hyPalette}},
  \code{\link{plot_spCompare}}, \code{\link{plot_spDiff}},
  \code{\link{plot_spDistribution}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{qplot_spRangeCenter}},
  \code{\link{qplot_spRangeMedian}},
  \code{\link{qplot_spc}}, \code{\link{rmExpr}},
  \code{\link{rm_stripes}}, \code{\link{stat_chull}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} plots}
