% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_spStat.R
\name{qplot_stat}
\alias{qplot_stat}
\title{[+] Plot a summary statistic of spectroscopic data by group}
\usage{
qplot_stat(sp, by = stop("Argument 'by' is missing."),
  FUN = stop("Argument 'FUN' is missing."),
  Title = fCap(as.character(match.call()$FUN)), subTitle = NULL,
  All = TRUE, fixed.colors = TRUE, All.color = "black",
  gr.color = c(hyGet.palette(sp), RColorBrewer::brewer.pal(8, "Dark2")),
  All.linetype = "dashed", gr.linetype = "solid", All.size = 1.1,
  gr.size = 0.8, legend.title = element_blank())
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}

\item{by}{A grouping variable (either a vertor or a variaable name in \code{sp}).}

\item{FUN}{A function that calculates one or several summary statistics.}

\item{Title}{The main title of the plot.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{All}{Logical. If \code{TRUE}, plot additional corve for a statistic
of all spectra (\code{.ALL}), not devided to grroups.}

\item{fixed.colors}{- Logical. If \code{TRUE}, color scheme where \code{.ALL}
is plotted as a black line is be used.}

\item{All.color}{A line color for \code{.ALL}.}

\item{gr.color}{Line colors (a vector) for each level in \code{by}.}

\item{All.linetype}{A line type for \code{.ALL}.}

\item{gr.linetype}{Line types (a vector) for each level in \code{by}.}

\item{All.size}{A line width for \code{.ALL}.}

\item{gr.size}{Line widths (a vector) for each level in \code{by}.}

\item{legend.title}{A title for legend. Default is no title.}
}
\value{
A plot: an object of class \code{\link[ggplot2]{ggplot}}.
       More details in package \pkg{ggplot2}).
}
\description{
Plot a summary of spectroscopic data in hyperSpec object by group.
The summary statistic internally is generated by either function
\code{\link{spStat}} or \code{\link[hyperSpec]{aggregate}}.
}
\examples{


qplot_stat(chondro,clusters,mean)
qplot_stat(chondro,clusters,mean,All=FALSE)
qplot_stat(chondro,clusters,mean_sd,All=FALSE) + facet_grid(.~clusters)

qplot_stat(chondro,clusters,median,All=FALSE, fixed.colors=FALSE)
qplot_stat(chondro,clusters,median, "My Title")


# Make facets:

qplot_stat(chondro,clusters,mean_pm_sd) + facet_grid(.~clusters) + nTick_x(2)

}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link{spStat}}

Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_spDiff}}, \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

