% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_of_fit.R
\name{quality_of_fit}
\alias{quality_of_fit}
\alias{spFitEval}
\title{[!] Calculate quality of fit parameters (for spectroscopic data)}
\usage{
quality_of_fit(obj = NULL, fit)

spFitEval(obj, fit)
}
\arguments{
\item{obj}{The first set of (\emph{original}) spectra. Either a matrix or a
hyperSpec object.}

\item{fit}{The second set of usually processed or reconstructed spectra, that
will be compared to the first set. Either a matrix or a hyperSpec object.}
}
\value{
A dataframe with following parameters for each row in \code{obj} and \code{fit}:
\itemize{
\item{\code{$LOF} - Lack of fit in percent (\%):
 \deqn{\frac{\sum{(obj - fit)^2}}{\sum{obj^2}} \cdot 100\%}{sum((obj - fit)^2)/sum((obj)^2) * 100\%}}
\item{\code{$GOF} - Goodness of fit in percent (\%): \deqn{100\% - LOF}{100 - LOF}}
\item{\code{$RMSE} - Root of mean squared error: \deqn{\sqrt{(obj - fit)^2}}{sqrt(obj - fit)^2}}
}
}
\description{
Calculate LOF (lack-of-fit), GOF (goodness-of-fit) and RMSE (root of mean
squared error) for original and reconstructed spectra.
}
\examples{
obj <- Spectra2
fit <- reconstructSp(Loadings2, Scores2)

evaluation <- quality_of_fit(obj, fit)

head(evaluation)

par(mfrow = c(3,1))
 plot(density(evaluation$RMSE), main = "RMSE", col = 2)
 plot(density(evaluation$LOF), main = "Lack-of-fit")
 plot(density(evaluation$GOF), main = "Goodness-of-fit")
par(mfrow = c(1,1))

# Density plot of LOF by group:
DATA <- cbind(obj$.., evaluation)
qplot(LOF, data = DATA, fill = gr, color = gr,
 geom = "density", alpha = I(.2))
}
