% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{quartiles}
\alias{quartiles}
\title{Median and quantiles}
\usage{
quartiles(x = NULL, MEDIAN = median(x), Q1 = quantile(x, 0.25),
  Q3 = quantile(x, 0.75))
}
\arguments{
\item{x}{data}
}
\value{
List with results... ???
}
\description{
....Confidence interval for median
}
\details{
....CI for case where variance of population is unknown.
}
\examples{

library(spHelper)

x <- c(3,0,2,1,0,4,3,2,1,2)
quartiles(x)

}
\seealso{
Other statistical functions: \code{\link{median_CI}},
  \code{\link{normal_var_CI}},
  \code{\link{poisson_lambda_CI}}
}
\concept{statistical functions}
