% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OceanView.header.R
\name{read.OceanView.header}
\alias{read.OceanView.header}
\title{[!] Read header lines of OceanView file}
\usage{
read.OceanView.header(
  file = NULL,
  dec = ".",
  n = 17,
  software = "OceanView",
  version_ = "1.5.2",
  software_info = paste(software, version_),
  last_headerline_text = ">>>>>Begin Spectral Data<<<<<",
  text
)
}
\arguments{
\item{file}{(string) A name of OcenView file .}

\item{dec}{("." | ",") The character used in the file for decimal points , e.g.
period (\code{"."}) or comma (\code{","}).}

\item{n}{(integer) Number of header lines to be scanned . Default is 17.
These lines include the header line indicated in
\code{last_headerline_text}.}

\item{software}{("OceanView" | "OOIBase32") Name of software, that created the file.}

\item{version_}{("1.5.2" | "1.5.0" | string) A version of OceanView file.
Curently supported  versions are "1.5.2" (or newer) and "1.5.0".}

\item{software_info}{(string) A combination of \code{software} and its \code{version_}.}

\item{last_headerline_text}{(string) A string, that indicates the last header line.
Default is \code{">>>>>Begin Spectral Data<<<<<"}.}

\item{text}{(strings) Text of header as if it was read with function
\code{\link[base]{readLines}}. If \code{text} is provided, \code{file}
is ignored.}
}
\value{
A dataframe with information, extracted from headerlines.
}
\description{
Read header lines with non-spectroscopic information of OceanView
(version 1.5.2) file.
}
\examples{

# Read from file: -------------------------------------------

\donttest{
\dontrun{
 read.OceanView.header("MySpectra.txt")
}}

# Read as text: ---------------------------------------------
header <- c(
     "Data from MySpectra.txt Node",
     "",
     "Date: Fri Jan 15 16:15:16 GMT 2014",
     "User: Scientist_1",
     "Spectrometer: USB2E2321",
     "Trigger mode: 4",
     "Integration Time (sec): 1.000000E1",
     "Scans to average: 1",
     "Electric dark correction enabled: true",
     "Nonlinearity correction enabled: true",
     "Boxcar width: 0",
     "XAxis mode: Wavelengths",
     "Number of Pixels in Spectrum: 2048",
     ">>>>>Begin Spectral Data<<<<<"
)

DF <- read.OceanView.header(text = header)

class(DF)
print(DF)


# Examples of file headerline formats: -----------------------

## An example of header lines in OcenView 1.5.2:

#    Data from MySpectra.txt Node
#
#    Date: Fri Jan 15 16:15:16 GMT 2014
#    User: Scientist_1
#    Spectrometer: USB2E2321
#    Trigger mode: 4
#    Integration Time (sec): 1.000000E1
#    Scans to average: 1
#    Electric dark correction enabled: true
#    Nonlinearity correction enabled: true
#    Boxcar width: 0
#    XAxis mode: Wavelengths
#    Number of Pixels in Spectrum: 2048
#    >>>>>Begin Spectral Data<<<<<

## An example of header lines in OcenView 1.5.0:

#    Data from MySpectra.txt Node
#
#    Date: Fri Jan 15 16:15:16 GMT 2016
#    User: Scientist_1
#    Spectrometer: USB2E2321
#    Autoset integration time: false
#    Trigger mode: 4
#    Integration Time (sec): 1.000000E1
#    Scans to average: 1
#    Electric dark correction enabled: true
#    Nonlinearity correction enabled: true
#    Boxcar width: 0
#    XAxis mode: Wavelengths
#    Stop averaging: false
#    Number of Pixels in Spectrum: 2048
#    >>>>>Begin Spectral Data<<<<<

}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}

Other functions to read spectroscopic data: 
\code{\link{file}},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
\concept{functions to read spectroscopic data}
