% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OceanView.ts.R
\name{read.OceanView.ts}
\alias{read.OceanView.ts}
\alias{read.OceanView.ts2}
\title{[!v0.1] Read OceanView timeseries spectra}
\usage{
read.OceanView.ts(file, dec = ".")

read.OceanView.ts2(file, dec = ",")
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{dec}{The character used in the file for decimal points.}
}
\value{
A \code{\link[=hyperSpec-class]{hyperSpec}} object with technical and
        spectroscopic information from file \code{file}.
}
\description{
Read contents of spectroscopic timeseries data file to
\code{\link[=hyperSpec-class]{hyperSpec}} object. File is created by
 software `OceanView 1.5`.
}
\examples{

\donttest{
\dontrun{

read.OceanView.ts("Spectra.txt")
read.OceanView.ts("Spectra.txt", dec = ",")
read.OceanView.ts2("Spectra.txt")


# Read several files to one hyperspec object:

Files   <- dir()[1:4]
sp_list <- lapply(Files, read.OceanView.ts2) # A list of objects
sp     <- collapse(sp_list)
plotmat(sp)
}}

}
\author{
Vilmantas Gegzna
}
\seealso{
Other functions for \pkg{hyperSpec}: \code{\link{hyGet.palette}},
  \code{\link{spStat}}
}

