% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OOIBase32.R
\name{read.OOIBase32}
\alias{read.OOIBase32}
\alias{read.OOIBase32_2}
\alias{read.OOIBase32_0}
\title{[!] Read data from OceanOptics OOI Base32 ASCII file with header lines}
\usage{
read.OOIBase32(
  file,
  dec = ".",
  n = 19,
  last_headerline_text = ">>>>>Begin Spectral Data<<<<<",
  version_ = "2.0.0.5",
  parse_filename = NULL,
  ignore.case = TRUE,
  xlab = NULL,
  ylab = "I, a.u."
)

read.OOIBase32_2(
  file,
  dec = ",",
  n = 19,
  last_headerline_text = ">>>>>Begin Spectral Data<<<<<",
  version_ = "2.0.0.5",
  parse_filename = NULL,
  ...
)

read.OOIBase32_0(
  file,
  dec = ".",
  n = 19,
  last_headerline_text = ">>>>>Begin Spectral Data<<<<<",
  version_ = "2.0.0.5",
  parse_filename = parser_TD2015,
  ...
)
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{dec}{("." | ",") The character used in the file for decimal points , e.g.
period (\code{"."}) or comma (\code{","}).}

\item{n}{(integer) Number of header lines to be scanned . Default is 17.
These lines include the header line indicated in
\code{last_headerline_text}.}

\item{last_headerline_text}{(string) A string, that indicates the last header line.
Default is \code{">>>>>Begin Spectral Data<<<<<"}.}

\item{version_}{version of OOI Base32 file. Curently version "2.0.0.5"
is supported.}

\item{parse_filename}{(function | string | NULL) Either a function that
parses string of filename and
extracts relevant information,\cr or a regular
expression to be used to parse file name and extract
relevant information,\cr or \code{NULL}.}

\item{ignore.case}{A logical that indicates if regular extression in
\code{parse_filename} is case sensitive.
Applies ONLY if \code{parse_filename} is a string with regular
expression.}

\item{xlab}{Label for x (wavelength) axis. If \code{NULL} (defaut) -
the label is automatically selected as "Wavelength, nm".}

\item{ylab}{(string) Label for y (intensity) axis. Defaut is \code{"I, a.u."}.}
}
\value{
A \code{\link[=hyperSpec-class]{hyperSpec}} object with technical and
        spectroscopic information from file \code{file}.
}
\description{
Read contents of spectroscopic ASCII data file with header lines that contain
metadata to \code{\link[=hyperSpec-class]{hyperSpec}} object. THe file is
created by software `OOI Base32 2.0.0.5`.
}
\details{
In \code{read.OOIBase32} the default decimal symbol is dot (\code{.}).\cr
         In \code{read.OOIBase32_2} - \emph{comma} (\code{,}).
}
\examples{

\donttest{
\dontrun{

read.OOIBase32("Spectra_ascii.txt")

# Read several files to one `hyperspec` object:

Files   <- dir()[1:4]                           # 4 files are selected
sp_list <- lapply(Files, read.OOIBase32)  # Make a list of objects
sp      <- collapse(sp_list)                    # Merge several spectra to one object

plotmat(sp)
}}

}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{sd_outliers}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}

Other functions to read spectroscopic data: 
\code{\link{file}},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
\concept{functions to read spectroscopic data}
