% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sp.csv.R
\name{read.sp.csv2}
\alias{read.sp.csv2}
\title{[!v0.1] Read spectroscopic CSV (comma separated values) file}
\usage{
read.sp.csv2(file, parse_filename = NULL, ylab = "Absorbance",
  xlab = "Wavelength, nm", ...)
}
\arguments{
\item{file}{The name of the file with specroscopic data.}

\item{parse_filename}{(function | string | NULL) Either a function that
parses string of filename and
extracts relevant information,\cr or a regular
expression to be used to parse file name and extract
relevant information,\cr or \code{NULL}.}

\item{ylab}{(string) Label for y (intensity) axis. Defaut is \code{"I, a.u."}.}

\item{xlab}{(string | NULL) Label for x (wavelength) axis.
 If \code{NULL} (defaut) -
the label is selected automatically.}

\item{...}{parameters to be passed to \code{\link[base]{read.csv}}.}
}
\value{
A \code{\link[=hyperSpec-class]{hyperSpec}} object with technical and
        spectroscopic information from file \code{file}.
}
\description{
[!v0.1] Read spectroscopic CSV (comma separated values) file
}
\details{
In \code{read.sp.csv} the default decimal symbol is dot (\code{.}).\cr
         In \code{read.sp.csv2} - \emph{comma} (\code{,}).
}
\examples{

\donttest{
\dontrun{

read.sp.csv2("Spectra_ascii.txt")

# Read several files to one `hyperspec` object:

Files   <- dir()[1:4]                           # 4 files are selected
sp_list <- lapply(Files, read.sp.csv)  # Make a list of objects
sp      <- collapse(sp_list)                    # Merge several spectra to one object

plotmat(sp)
}}

}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{IQR_outliers}},
  \code{\link{binning}}, \code{\link{file}},
  \code{\link{gapDer}}, \code{\link{hy2mat}},
  \code{\link{hyAdd_Label_wl}},
  \code{\link{hyAdd_Labels_PAP_PD_2014}},
  \code{\link{hyAdd_Labels_TD2009}}, \code{\link{hyAdd}},
  \code{\link{hyDrop_NA}}, \code{\link{hyGet_palette}},
  \code{\link{hyRm_palette}}, \code{\link{mad_outliers}},
  \code{\link{mean_Nsd}}, \code{\link{median_Nmad}},
  \code{\link{plot_hyPalette}},
  \code{\link{prepare_PAP_RK_2014__MATLAB_failui}},
  \code{\link{read.OOIBase32}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView}}, \code{\link{read3csv2hy}},
  \code{\link{replace_spc}}, \code{\link{sd_outliers}},
  \code{\link{spStat}}, \code{\link{sp_class_perform}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
