% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_OOspectra.R
\name{read_OOspectra}
\alias{read_OOspectra}
\title{Read OceanOptics spectra}
\usage{
read_OOspectra(path, pattern = ".*", read_fun = c("auto", "ascii",
  "ts", "base32", "a FUNCTION")[1], recursive = FALSE, ...)
}
\arguments{
\item{path}{(string)}

\item{pattern}{(string) regular extression to filter files to be selected as spectra.}

\item{read_fun}{(string | function)}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{...}{parameters to be passed to `read_fun`}
}
\value{
hyperSpec object
}
\description{
Read OceanOptics spectra
}
