% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcapturedmatches.R
\name{regcapturedmatches}
\alias{regcapturedmatches}
\title{Extract regular expression matches captured by names}
\usage{
regcapturedmatches(x, m)
}
\arguments{
\item{x}{- (a list of) strings.}

\item{m}{- ParsedData, aresult from a regular expression function.}
}
\value{
....
}
\description{
regcapturedmatches.R: extracts captured matches from match data obtained
 by regexpr, gregexpr or regexec.
}
\examples{
# usage
x <- c("larry:35,M","alison:22,F","dave","lily:55,F")
m <- regexpr("(.*):(\\\\\\\\d+),([MF])", x, perl=T)
D
regcapturedmatches(x,m)


}
\author{
Daniel Marcelino
\url{https://gist.github.com/danielmarcelino/8209cfdf2cd993eeb1b3}[github.com/danielmarcelino]
\url{https://gist.github.com/MrFlick/10413321}
\url{http://stackoverflow.com/questions/33288075/from-matlab-to-r-capture-named-fields-with-regular-expressions-to-a-dataframe}
}

