% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_spc.R
\name{replace_spc}
\alias{replace_spc}
\title{[!!!] Replace \code{object$spc} and \code{colnames(object$spc)}}
\usage{
replace_spc(object, x, ...)
}
\arguments{
\item{object}{A \code{hyperSpec} object.}

\item{x}{matrix with the new content for \code{object@data$spc}.

  Its size must correspond to rows (for \code{scores}) and to either columns
  or rows (for \code{loadings}) of \code{object}.}

\item{...}{further arguments to \link[hyperSpec]{decomposition}}
}
\value{
A \code{hyperSpec} object, updated according to x
}
\description{
[!!!] Replace \code{object$spc} and \code{colnames(object$spc)}
}
\examples{

ratios <- colRatios(Scores2[[]])
obj1 <- replace_spc(Scores2, ratios)

colnames(obj1$spc)  # preserved

obj2 <- decomposition(Scores2, ratios)
colnames(obj2$spc)  # NOT preserved


}
\seealso{
\link[hyperSpec]{decomposition}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{IQR_outliers}},
  \code{\link{binning}}, \code{\link{file}},
  \code{\link{gapDer}}, \code{\link{hy2mat}},
  \code{\link{hyAdd_Label_wl}},
  \code{\link{hyAdd_Labels_PAP_PD_2014}},
  \code{\link{hyAdd_Labels_TD2009}}, \code{\link{hyAdd}},
  \code{\link{hyDrop_NA}}, \code{\link{hyGet_palette}},
  \code{\link{hyRm_palette}}, \code{\link{mad_outliers}},
  \code{\link{mean_Nsd}}, \code{\link{median_Nmad}},
  \code{\link{plot_hyPalette}},
  \code{\link{prepare_PAP_RK_2014__MATLAB_failui}},
  \code{\link{read.OOIBase32}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView}}, \code{\link{read.sp.csv2}},
  \code{\link{read3csv2hy}}, \code{\link{sd_outliers}},
  \code{\link{spStat}}, \code{\link{sp_class_perform}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
