% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_stripes.R
\name{rm_stripes}
\alias{rm_stripes}
\title{[!] Remove facet stripes from `ggplot2` plot}
\usage{
rm_stripes(axis = "xy")
}
\arguments{
\item{axis}{String indicating axis from which the stripes should be removed:
either \code{"x"}, \code{"y"}, \code{"xy"}. Default is \code{"xy"}.}
}
\value{
Updated \pkg{ggplot2} object of class \code{\link[ggplot2]{ggplot}}.
}
\description{
A convenience function to remove facet stripes from `ggplot2` plot.
}
\details{
\code{rm_stripes} is a wrapper of function \code{\link[ggplot2]{theme}}
 with one of the 3 options (depending on value of input variable "axis")
 \code{strip.text.x}, \code{strip.text.y}, or
 \code{strip.text} set to \code{element_blank()}.
}
\examples{

p <- ggplot(Spectra2, wl.range = c(min ~ 420, 500~600)) +
     geom_line(aes(color = gr)) +
     facet_wl("class")

p
p + rm_stripes("x")
p + rm_stripes()

}
\seealso{
Function \code{\link[ggplot2]{theme}} from package \pkg{ggplot2}.

Other \pkg{spHelper} plots: \code{\link{check_palette}},
  \code{\link{layer_spRangeMean}},
  \code{\link{plot_colors}}, \code{\link{plot_hyPalette}},
  \code{\link{plot_spCompare}}, \code{\link{plot_spDiff}},
  \code{\link{plot_spDistribution}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{qplot_spRangeCenter}},
  \code{\link{qplot_spRangeMedian}},
  \code{\link{qplot_spStat}}, \code{\link{qplot_spc}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} plots}
