% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmExpr.R
\name{rmExpr}
\alias{rmExpr}
\title{[.] Convert expressions in \code{ggplot2} object labels to strings}
\usage{
rmExpr(p = ggplot2::last_plot())
}
\arguments{
\item{p}{ggplot2 object}
}
\value{
`ggplot2` object
}
\description{
Convert expressions in \code{\link[ggplot2]{ggplot}} object labels to strings, to
enable them to be plotted with \code{\link[plotly]{ggplotly}}.
}
\examples{

\donttest{
\dontrun{
data(flu, package = "hyperSpec")

qplotspc(flu)
ggplotly()

rmExpr()
ggplotly()

#--------------------------------
p1 <- plot_sp(Loadings)
p1$labels

p2 <- rmExpr(p1)
p2$labels

ggplotly(p1)
ggplotly(p2)

#----------------------------------
qplotspc(aggregate(chondro, chondro$clusters, mean_pm_sd),
mapping = aes(x = .wavelength,
              y = spc,
              colour = clusters)) +
    facet_grid(clusters ~ .) + ggtitle("Spectra of Chondrocytes")

 rmExpr()
 ggplotly()

}}

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_spDiff}}, \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{stat_chull}}

Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{clear}},
  \code{\link{expr2text}}, \code{\link{listAddRm}},
  \code{\link{make.filenames}},
  \code{\link{make.firstCapitals}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{st01}}, \code{\link{subt}},
  \code{\link{uncall}}
}

