% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_.R
\name{ROC_}
\alias{ROC_}
\alias{ROC_table}
\alias{ROC_plots}
\title{[.]  Do ROC analysis and plot it's results (ROC_)}
\usage{
ROC_(x, labels, label.ordering = NULL, make_plots = TRUE)

ROC_table(prediction.obj)

ROC_plots(prediction.obj)
}
\arguments{
\item{labels}{A vector, matrix, list, or data frame containing the
    true class labels. Must have the same dimensions as 'predictions'.}

\item{label.ordering}{The default
    ordering (cf.details)  of the classes can be changed by supplying a vector
    containing the negative and the positive class label.}

\item{make_plots}{Logical. If TRUE (default) Sensitivity-Specificity plot and
plot of sensitivity and specificity at each cot-off point are plotted.}

\item{prediction.obj}{An object of class \code{prediction}.}
}
\value{
A) plots as listed in description of \code{make_plots}.\cr
B) Table of performance measures a optimal cut of point.
}
\description{
[.]  Do ROC analysis and plot it's results (ROC_)
}
\examples{

library(ROCR)
library(spHelper)

data(ROCR.simple)
ROC_(ROCR.simple$predictions, ROCR.simple$labels)
  ##  Compared groups                      "0 vs. 1"
  ##  Group treated as positive            "1"
  ##  AUC                                  "0.83"
  ##  Cut-off                              "0.5015"
  ##  Sensitivity (True positive rate, Se) "0.85"
  ##  Specificity (True negative rate, Sp) "0.85"
  ##  Mean of Se and Sp                    "0.85"


}
\seealso{
This function is based on package \pkg{ROCR}.
}
\author{
Vilmantas Gegzna
}
