% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_by_Ignas_Ciplys.R
\name{save_rds_unique}
\alias{save_rds_unique}
\title{[~!~] Save object as `.RDS` with unique name (i.e., without overwritting)}
\usage{
save_rds_unique(what, path)
}
\arguments{
\item{what}{Object to save.}

\item{path}{File name with path to save to.}
}
\description{
/No description yet/
}
\examples{

#NO EXAMPLES YET

}
\author{
Ignas Ciplys
}
\seealso{
Other functions by Ignas Ciplys: \code{\link{plot_multiDim}}
}

