% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_outliers.R
\name{sd_outliers}
\alias{sd_outliers}
\alias{outside_mean_pm_Nsd}
\title{[!] Outlier detection based on standard deviations (sd)}
\usage{
sd_outliers(x, n = 2)

outside_mean_pm_Nsd(x, n = 2)
}
\arguments{
\item{x}{Matrix-like data or \code{hyperSpec} object.}

\item{n}{A number of standard deviations.}
}
\value{
A list with these elements \itemize{
\item{\code{matrix} - a logical matrix, with \code{TRUE} for elements which
      are outside column mean +/- n standard deiations;}
 \item{\code{rows} - a logical vector that indicate rows that have at least
     one such element.}
 }
}
\description{
Detect elements in each \emph{column} that are outside mean +/- n standard
 deviations and rows that contain at least one such element.
}
\seealso{
\link{mean_Nsd}

Other \pkg{spHelper} utilities: 
\code{\link{IQR_outliers}()},
\code{\link{expr2text}()},
\code{\link{mad_outliers}()},
\code{\link{rmExpr}()},
\code{\link{subt}()},
\code{\link{unGroup}()},
\code{\link{uncall}()}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: 
\code{\link{IQR_outliers}()},
\code{\link{binning}()},
\code{\link{file}},
\code{\link{gapDer}()},
\code{\link{hy2mat}()},
\code{\link{hyAdd_Label_wl}()},
\code{\link{hyAdd_Labels_PAP_PD_2014}()},
\code{\link{hyAdd_Labels_TD2009}()},
\code{\link{hyAdd}()},
\code{\link{hyDrop_NA}()},
\code{\link{hyGet_palette}()},
\code{\link{hyRm_palette}()},
\code{\link{mad_outliers}()},
\code{\link{mean_Nsd}()},
\code{\link{median_Nmad}()},
\code{\link{plot_hyPalette}()},
\code{\link{prepare_PAP_RK_2014__MATLAB_failui}()},
\code{\link{read.OOIBase32}()},
\code{\link{read.OceanView.header}()},
\code{\link{read.OceanView}()},
\code{\link{read.sp.csv2}()},
\code{\link{read3csv2hy}()},
\code{\link{replace_spc}()},
\code{\link{spStat}()},
\code{\link{sp_class_perform}()}

Other outlier detection functions in \pkg{spHelper}: 
\code{\link{IQR_outliers}()},
\code{\link{mad_outliers}()}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
\concept{\pkg{spHelper} utilities}
\concept{outlier detection functions in \pkg{spHelper}}
