% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCorrVars.R
\name{simCorrVars}
\alias{simCorrVars}
\title{[+] Simulate correlated continuous data from standard normal distribution}
\source{
\url{http://www.r-bloggers.com/simulating-random-multivariate-correlated-data-continuous-variables/}
}
\usage{
simCorrVars(corrMat, N = 100)
}
\arguments{
\item{corrMat}{A symmetricmatrix of correlation coefficients. Determinant of
\code{corrMat} must be positive. If \code{corrMat} is a vector,
it is passed to function \code{\link{corr_vec2mat}} to
construct a correlation matrix.\cr
Default \code{corrMat = corr_vec2mat(c(.8,.2,.7))}.}

\item{N}{A number of observations. Default is 100.}
}
\value{
A dataframe with correlated variables.
}
\description{
Simulate correlated continuaous data from standard normal distribution by
using Choleski factorization of correlation matrix.
}
\details{
Number of variables is determined by the size of correlation matrix
         \code{corrMat}.
}
\examples{

# ------------------------------------------------------------
simCorrVars()
simCorrVars(N = 10)

# ------------------------------------------------------------
# Generate 2 correlated variables with 10 observations:
# A. Imput is a matrix

corrMat <- matrix(c(1, .5, .5, 1), 2)
print(corrMat)
simCorrVars(corrMat, N = 10)

# B. Imput is a vector
simCorrVars(.5, N = 10)
simCorrVars(c(.1,.5,.8), N = 10)

# ------------------------------------------------------------
}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link{corr_vec2mat}}, \link[MASS]{mvrnorm}, \link[base]{chol}

Other simmulation functions: \code{\link{corr_vec2mat}},
  \code{\link{simSpectra}}
}

