% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpectra.R
\name{simSpectra}
\alias{simSpectra}
\title{[!] Generate spectroscopic data}
\usage{
simSpectra(x = 300:800, nGr = 2, nInGr = c(50, 50, 50), N = sum(nInGr),
  nDim = 4, w.possile = 5:150, plots = FALSE)
}
\arguments{
\item{x}{Values for x axis ("wavelengths").}

\item{nGr}{Number of groups.}

\item{nInGr}{Number of samples in a group (vector of values for each group).}

\item{N}{Total number of samples.}

\item{nDim}{Number of spectral components (dimensions0).}

\item{w.possile}{A vector of possible values of Gaussian curve parameter "w'
(i.e., width)}

\item{plots}{Logical. If \code{TRUE} - makes plots. Default is \code{FALSE}.}
}
\value{
List of hyperSpec objects:
\describe{
     \item{Spectra}{Spectra of each observation, made of \code{Scores * Loadings + NOISE}.}
     \item{Loadings}{Normalized spectra of \bold{components}.}
     \item{Scores}{\bold{Amplitudes} of components for each observation.}
}
}
\description{
Simulate spectroscopic data: generate spectra, loadings and scores.
}
\examples{

simSpectra()

simSpectra()$Spectra

}

