% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortDescOnDiag.R
\name{sortDescOnDiag}
\alias{sortDescOnDiag}
\title{[!!!] Sort rows and columns of crosstabulation by the best match}
\usage{
sortDescOnDiag(M)
}
\arguments{
\item{M}{A matrix.}
}
\description{
[!!!] Sort rows and columns of crosstabulation by the best match
}
\examples{

M <- matrix(c(2,10,10,10,8,9,4:6,4,1,8),byrow=TRUE,nc=3)
M <- pkgmaker::addnames(M)

M1 <- sortDescOnDiag(M)
qplot_crosstab0(M1)


qplot_confusion(sortDescOnDiag(M))

}
\author{
Vilmantas Gegzna
}
\seealso{
Other matrix operations in \pkg{spHelper}: \code{\link{corr_vec2mat}},
  \code{\link{ind.matrix}}, \code{\link{nDiag2nTri}},
  \code{\link{nTri2nDiag}}, \code{\link{new.matrix}},
  \code{\link{which.in}}, \code{\link{which.max.all}},
  \code{\link{which.max.perRow}}
}

