% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_data_labels.R
\name{sp_data_labels}
\alias{sp_data_labels}
\title{Lables for variables derived from spectroscopic data files}
\usage{
sp_data_labels()
}
\value{
A named list with labels
}
\description{
Lables for variables derived from spectroscopic data files
}
