% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_filter.R
\name{sp_filter}
\alias{sp_filter}
\title{Filter noise form a spectroscopic signal}
\usage{
sp_filter(sp, k = 3, n = 25, p = 9)
}
\arguments{
\item{sp}{hyperSpec object}

\item{k}{Window for a \emph{median} filter. If k = 0, median filter is not applied.}

\item{n}{filter length (must be odd) for a Savitzky-Golay (SG)smoothing filter.
If n = 0, SG filter is not applied.}

\item{p}{filter order for a Savitzky-Golay smoothing filter}
}
\description{
Function first applies median and then Savitzky-Golay smoothing filters for
each spectroscopic curve individually.
}
\examples{
library(spHelper)
library(hyperSpec)

Original
Spectra2[1]  \%>\% plotspc()

# Filtered
Spectra2[1]  \%>\% sp_filter() \%>\% plotspc()

}
