% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_normalize.R
\name{sp_normalize}
\alias{sp_normalize}
\title{Normalize spectra}
\usage{
sp_normalize(sp, norm = "max_I", in_range = min ~ max, at = NULL)
}
\arguments{
\item{sp}{\code{hyperSpec} object.}

\item{norm}{("I" | "I_max" | "I_mean" | "area" | "max" | "mean" )\cr
a way of normalization.}

\item{in_range, at}{(\code{numeric} | \code{formula}) \cr
either a single numeric value or
a pair of values describing the range (e.g., 300~350)
of wavelengths to be used for normalization.
Default is \code{range = min~max}.}
}
\value{
\code{sp} with normalized spectra in (\code{sp@data$spc}).
}
\description{
Convenience function to normalize spectroscopic data in a \code{hyperSpec} object.
}
\examples{

library(magrittr)


Spectra2 \%>\% plotspc()

Spectra2 \%>\% sp_normalize() \%>\% plotspc()

Spectra2 \%>\% sp_normalize("I", 720) \%>\% plotspc()

Spectra2 \%>\% sp_normalize("area") \%>\% plotspc()
Spectra2 \%>\% sp_normalize("area", 300~400) \%>\% plotspc()
}
