% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_normalize2.R
\name{sp_normalize2}
\alias{sp_normalize2}
\title{Normalize spectra}
\usage{
sp_normalize2(sp, norm = "max", in_range = min ~ max, at = NULL)
}
\arguments{
\item{sp}{\code{hyperSpec} object.}

\item{norm}{("max" | "mean" | "area" | "area_abs" | "nim")\cr
a way of normalization.}

\item{in_range, at}{(\code{numeric} | \code{formula}) \cr
either a single numeric value or
a pair of values describing the range (e.g., 300~350)
of wavelengths to be used for normalization.
Default is \code{range = min~max}.}
}
\value{
\code{sp} with normalized spectra in (\code{sp@data$spc}).
}
\description{
Convenience function to normalize spectroscopic data in a \code{hyperSpec} object.
}
\examples{

library(magrittr)

Spectra2 <- hyperSpec::chondro
Spectra2 \%>\% plotspc()

Spectra2 \%>\% hyperSpec::normalize01() \%>\% plotspc()

Spectra2 \%>\% sp_normalize2(at = 700) \%>\% plotspc()

Spectra2 \%>\% sp_normalize2("area") \%>\% plotspc()

Spectra2 \%>\% sp_normalize2("area", in_range = 1400~1700) \%>\% plotspc()
}
