% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_remove_offset.R
\name{sp_remove_offset}
\alias{sp_remove_offset}
\title{Remove offset}
\usage{
sp_remove_offset(
  sp,
  in_range = min ~ max,
  offset = 0,
  method = "not_used_yet",
  at = NULL
)
}
\arguments{
\item{sp}{\code{hyperSpec} object.}

\item{in_range, at}{(\code{numeric} | \code{formula}) \cr
either a single numeric value or
a pair of values describing the range (e.g., 300~350)
of wavelengths to be used for normalization.
Default is \code{range = min~max}.}

\item{offset}{(number)\cr
value of desired offset.}
}
\value{
\code{sp} with y-axis offset removed in (\code{sp@data$spc}).
}
\description{
Convenience function to remove offset from spectroscopic data in a \code{hyperSpec} object.
}
\examples{

library(magrittr)

Spectra2 \%>\% plotspc()

Spectra2 \%>\% sp_remove_offset() \%>\% plotspc()

Spectra2 \%>\% sp_remove_offset(in_range = 300~400) \%>\% plotspc()
}
