% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spStat.R
\name{spStat}
\alias{spStat}
\title{[+] Calculate summary statistic of spectroscopic data for all spectra and by groups}
\usage{
spStat(sp, by, FUN = mean, Name_of.by = as.character(match.call()$by),
  ...)
}
\arguments{
\item{sp}{A hyperSpec object.}

\item{by}{A grouping variable (either variable name of \code{sp},
or vector (factor) of length \code{length(sp)}}

\item{FUN}{A function to apply.}

\item{Name_of.by}{(Do not change this!) The name of input \code{by}.}

\item{...further}{arguments passed to \code{FUN}.}
}
\value{
A hyperSpec object that contains summary statistic. Additional
        column \code{.aggregate} is added. In this column levels of
        grouping variable are indicated.
}
\description{
Calculate summary statistic spectra in \code{hyperSpec} object for all
spectra and by levels of factor variable \code{by}.
}
\note{
This function is similar to \code{\link[hyperSpec]{aggregate}},
just calculates additional statistic for all spectra. This
statistic is labeled by adding level \code{.ALL}) in variable
\code{.aggregate} and variable, that has the same name as \code{by}
(if \code{Name_of.by} si not modified).
}
\examples{
spStat(Spectra2, gr)
spStat(Spectra2, by = gr,          FUN = IQR)

## BUG:
# spStat(Spectra2, by = Spectra2$gr, FUN = IQR)

}
\seealso{
\code{\link[hyperSpec]{aggregate}}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{IQR_outliers}},
  \code{\link{binning}}, \code{\link{file}},
  \code{\link{gapDer}}, \code{\link{hy2mat}},
  \code{\link{hyAdd_Label_wl}},
  \code{\link{hyAdd_Labels_PAP_PD_2014}},
  \code{\link{hyAdd_Labels_TD2009}}, \code{\link{hyAdd}},
  \code{\link{hyDrop_NA}}, \code{\link{hyGet_palette}},
  \code{\link{hyRm_palette}}, \code{\link{mad_outliers}},
  \code{\link{mean_Nsd}}, \code{\link{median_Nmad}},
  \code{\link{plot_hyPalette}},
  \code{\link{prepare_PAP_RK_2014__MATLAB_failui}},
  \code{\link{read.OOIBase32}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView}}, \code{\link{read.sp.csv2}},
  \code{\link{read3csv2hy}}, \code{\link{replace_spc}},
  \code{\link{sd_outliers}}, \code{\link{sp_class_perform}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}}
