% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spStat_ci.R
\name{spStat_ci}
\alias{spStat_ci}
\alias{spStat_ci_corr}
\alias{ggplot_ci_rez}
\title{Calculate a statistic}
\usage{
spStat_ci(obj, FUN = mean, label = as.character(match.call()$FUN)
  \%if_null_or_len0\% "mean")

spStat_ci_corr(obj, y = NULL, FUN = mean, method = c("spearman",
  "kendall", "pearson")[1], use = "everything", conf = 0.95,
  R = 1000, sim = "balanced", type = c("norm"),
  label = paste0(spMisc::fCap(method), "'s corr. coeff."))

ggplot_ci_rez(rez, linetype = 1)
}
\arguments{
\item{obj}{hyperSpec object.}

\item{FUN}{a function that takes a vector and results in a single number,
e.g., mean, median, etc.}

\item{label}{(string) a label for function to be used as column name.}

\item{y}{a vector.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}

\item{conf}{
    A scalar or vector containing the confidence level(s) of the required
    interval(s).
  }

\item{R}{
    The number of bootstrap replicates.  Usually this will be a single
    positive integer.  For importance resampling, some resamples may use
    one set of weights and others use a different set of weights.  In
    this case \code{R} would be a vector of integers where each
    component gives the number of resamples from each of the rows of
    weights.
  }

\item{sim}{
    A character string indicating the type of simulation required.
    Possible values are \code{"ordinary"} (the default),
    \code{"parametric"}, \code{"balanced"}, \code{"permutation"}, or
    \code{"antithetic"}.  Importance resampling is specified by
    including importance weights; the type of importance resampling must
    still be specified but may only be \code{"ordinary"} or
    \code{"balanced"} in this case.
  }

\item{type}{
    A vector of character strings representing the type of intervals
    required. The value should be any subset of the values
    \code{c("norm","basic", "stud", "perc", "bca")} or simply \code{"all"}
    which will compute all five types of intervals.
  }

\item{rez}{a rezult of function \code{spStat_ci_corr} or \code{spStat_ci}}
}
\description{
### NEBAIGTA ###
}
\examples{

rez <- spStat_ci(Spectra2)
rez

qplotspc(rez)

ggplot(Spectra2) + geom_line()

 data <- hy_spc2df(rez)
 ggplot(data,
        aes_string(x = "wl",
                   y = names(data)[3],
                   ymin = "ci_lower",
                   ymax = "ci_upper")
 ) +
     geom_ribbon(alpha = 0.2) +
     geom_line(linetype = 2)

ggplot(Spectra2[1:20,,300~400])

hyperSpec::aggregate(Spectra2, by = "gr", FUN = mean)


set.seed(1)
amzius   <- rnorm(nrow(Spectra2))
spektrai <- Spectra2[,,400~500]

set.seed(1)
rez <- spStat_ci_corr(spektrai, y = amzius)
ggplot_ci_rez(rez)

}
