% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plotting_helpers.R
\name{spStat_ldf}
\alias{spStat_ldf}
\title{Calculate a summary statistic and return long-format data frame}
\usage{
spStat_ldf(sp, by, FUN, ..., var_names = NULL, name_if_by_is_NULL = "All Data")
}
\arguments{
\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{by}{(string) name of grouping variable.}

\item{FUN}{Function of summray statistics.}

\item{...}{Arguments to be passed to \code{FUN}.}

\item{var_names}{Names of summary statisticsm generated with \code{FUN}.}

\item{name_if_by_is_NULL}{(string) Name of whole dataset in case \code{by}
is NULL. Default is "All Data"}
}
\value{
A long-format data frame.
}
\description{
Calculate a summary statistic of spectroscopic data in
\code{hyperSpec} object and return the result as a
long-format data frame.
}
\examples{
spStat_ldf(Spectra2, FUN = mean)

spStat_ldf(sp, "gr", FUN = quantile, probs = c(0,1),
           var_names = c("ymin","ymax"))

}
