% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spZone.R
\name{spZone}
\alias{spZone}
\alias{spZone.default}
\alias{spZone.formula}
\title{[!] Annotate x axis range in ggplot2 graphs}
\usage{
spZone(xmin, xmax, fill = "red", alpha = 0.2, ...)

\method{spZone}{default}(xmin, xmax, fill = "red", alpha = 0.2, ...)

\method{spZone}{formula}(x, fill = "red", alpha = 0.2, ...)
}
\arguments{
\item{xmin}{Minimum value of x axis.}

\item{xmax}{Maximum value of x axis.}

\item{fill}{Fill color.}

\item{alpha}{Fill transperency.}

\item{...}{further arguments to \code{\link[ggplot2]{annotate}}.}

\item{x}{formula indicating min and max values (\code{xmin ~ xmax}).}
}
\description{
A wrapper function to annotate x axis range for ggplot2 graphs
}
\examples{

qplot_sp(flu) + spZone(425~450)
qplot_sp(flu) + spZone(460,490, fill = "green")

}
