% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st01.R
\name{st01}
\alias{st01}
\title{[+] Standardize values to intervat from min to 1}
\usage{
st01(x, m = min(x))
}
\arguments{
\item{x}{A numeric vector vector}

\item{m}{A value treated as a minimum.}
}
\description{
Apply formula: \eqn{(x-m)/(1-m)}{(x-m)/(1-m)}
}
\examples{

x <- seq(.3,1,.1)
st01(x)
st01(x, m = .2)

}
\seealso{
Other family \pkg{spHelper} utilities: \code{\link{clc}},
  \code{\link{make.nameswin}}
}

