% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Fun_Plotting_Sp.R
\name{stat_chull}
\alias{stat_chull}
\title{[!] Convex hull layer for ggplot2}
\usage{
stat_chull(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Source: vignette \href {https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}{Extending ggplot2}
}
\examples{

ggplot(mpg, aes(displ, hwy)) +
             geom_point() +
             stat_chull(fill = NA, colour = "black")


ggplot(mpg, aes(displ, hwy, colour = drv)) +
             geom_point() +
             stat_chull(fill = NA)


ggplot(mpg, aes(displ, hwy)) +
             stat_chull(geom = "point", size = 4, colour = "red") +
             geom_point()

}

