% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subt.R
\name{subt}
\alias{subt}
\title{[+] Add bold title and subtitle to a plot}
\usage{
subt(Title = NULL, subTitle = NULL)
}
\arguments{
\item{Title}{The first line of title, which will be in bold.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}
}
\value{
Formated title.
}
\description{
Add bold title with second line in smaller font with subtitle.
              Functions \code{\link[base]{bquote}} and
               \code{\link[grDevices]{atop}} are used to achieve
               this effect.
}
\section{Warning!}{


May not work with plotting functions from package \pkg{hyperSpec}.
}
\examples{

subt("Cars")
## bold("Cars")

subt("Cars","Distance vs. speed")
## atop(bold("Cars"), atop(italic("Distance vs. speed")))

# ----------------------------------------------------------------

plot(cars[,1:2], main = "Cars")
plot(cars[,1:2], main = subt("Cars")) # the same as in previous line
plot(cars[,1:2], main = subt("Cars","Distance vs. speed"))
plot(cars[,1:2], main = subt(subTitle = "Distance vs. speed"))

# ----------------------------------------------------------------

library(ggplot2)

g <- qplot(mpg, wt, data=mtcars)
g + ggtitle("Cars") # non-bold title
g + ggtitle(subt("Cars")) # bold title
g + ggtitle(subt("Cars","Distance vs. speed"))
g + ggtitle(subt(subTitle = "Distance vs. speed"))


# ----------------------------------------------------------------

library(lattice)

xyplot(eruptions~waiting, data = faithful)

xyplot(eruptions~waiting, data = faithful,
 main = "Old Faithful Geyser Data")

xyplot(eruptions~waiting, data = faithful,
 main = subt("Old Faithful Geyser Data"))

xyplot(eruptions~waiting, data = faithful,
 main = subt("Old Faithful Geyser", "Data"))

xyplot(eruptions~waiting, data = faithful,
 main = subt(subTitle = "Old Faithful Geyser Data"))

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clear}}, \code{\link{expr2text}},
  \code{\link{make.firstCapitals}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{uncall}}
}

