% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_hyData.R
\name{summary_hyData}
\alias{summary_hyData}
\alias{summary_hyData.default}
\alias{summary_hyData.hyperSpec}
\title{[!] Summary statistics of non-spectroscopic data in `hyperSpec` objecet}
\usage{
summary_hyData(object, ...)

\method{summary_hyData}{default}(object, ...)

\method{summary_hyData}{hyperSpec}(sp, ..., unique_ID = FALSE)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{further parameters to be passed to function \code{summary}.}

\item{sp}{\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{unique_ID}{(FALSE | character) either variable name
that contain ID numbers (to select only the first rows with
unique ID) or \code{FALSE} (default) to use all rows.}
}
\value{
A table with summary statistics.
}
\description{
An alternative way to summarize non spectroscopic data (in \code{@data} slot)
in a \code{hyperSpec} object.
}
\details{
If class of summarized object is:\cr
 - "hyperSpec", lablels of variables (slot \code{@label}) are used as
  headers of columns.\cr

- not "hyperSpec", the function behaves in a similar way as function
 \code{summary}. The differences are indicated in section "Examples".
}
\examples{

library(spHelper)
library(pander)


# Summary of "hyperSpec" object
 summary_hyData(Spectra2)


# Summary of "hyperSpec" object + `pander` (useful if `knitr` is used)
 summary_hyData(Spectra2) \%>\%  pander


# ======= `summary_hyData(sp)` vs `summary(sp$..)' vs `summary(sp)' ========
sp <- Spectra2
labels(sp) <- list(gr = "--- Group ---", class = "--- Class ---")

summary_hyData(sp)  # Column names are appropriate values of `labels(sp)`
#>   --- Group --- --- Class ---
#>    A:52          K :50
#>    B:55          l :30
#>    C:43          N :19
#>    S1:51

summary(sp$..)      # Column names are appropriate values of `colnames(sp)`
#>    gr     class
#>    A:52   K :50
#>    B:55   l :30
#>    C:43   N :19
#>    S1:51

summary(sp)         # Default summary of whole `hyperSpec` object
#>   hyperSpec object
#>   150 spectra
#>   3 data columns
#>   501 data points / spectrum
#>   wavelength:  [integer] 300 301 ... 800
#>   data:  (150 rows x 3 columns)
#>   1. gr: --- Group --- [factor] B B ... A
#>   2. class: --- Class --- [factor] N l ... S1
#>   3. spc:  [matrix501] 159.8996 139.9296 ... 12.11558

# ======= Summary of factor variables in a data frame ======================
# (if printed using function `pander` there are differences in column
# `Species` ):

iris[,4:5]  \%>\%  summary         \%>\%  pander
iris[,4:5]  \%>\%  summary_hyData  \%>\%  pander


}
\seealso{
\code{\link[base]{summary}}
}
