% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFolds.R
\name{TestFolds}
\alias{TestFolds}
\title{[TMP] Test if data in folds is stratified and blocked.}
\usage{
TestFolds(FoldsOBJ, DATA, stratBy = "gr", blockBy = "ID")
}
\arguments{
\item{FoldsOBJ}{list with indices in folds.}

\item{DATA}{A data set, for wich \code(FoldsOBJ) was created.}

\item{stratBy}{A name of variable used for stratification.}

\item{blockBy}{A name of variable used for blocking.}
}
\value{
Print tables to evaluate if data is (A) stratified, (B) blocked.
}
\description{
[TMP] Test if data in folds is stratified and blocked.
}

