% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3__math2latex.R
\name{to_math}
\alias{to_math}
\alias{to_math_q}
\title{A partial implementation of a R expression -> latex math converter}
\source{
\href{https://gist.github.com/hadley/5576263}{latex-math.r}
}
\usage{
to_math(x)

to_math_q(x)
}
\arguments{
\item{x}{A supported R plotmath expression.}
}
\value{
A LaTeX expression
}
\description{
A partial implementation of a R expression -> latex math converter
}
\examples{


to_math(x_1 + 1^{2 + 4} + 5 + sqrt(y) / 5 \%/\% 10)
to_math(paste(x^2, y - 1, z_i))
to_math(hat(tilde(ring(x))))

to_math_q(expression(x+1))


# ERROR (known bug):

\donttest{
\dontrun{
to_math_q(expression(x+1))
to_math(+1)
to_math(x*+1)
to_math(x^-a)

}}

}
\author{
Hadley Wickham
}

