% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmExpression.R
\name{uncall}
\alias{uncall}
\title{[.] Unwrap text from `call` object
Assumption: only the second element in a `call` can be text
first element is an object of class "name", that can be dropped.
This algorithm can loose some text elements.}
\usage{
uncall(x)
}
\arguments{
\item{x}{A `call` object}
}
\value{
A string
}
\description{
[.] Unwrap text from `call` object
Assumption: only the second element in a `call` can be text
first element is an object of class "name", that can be dropped.
This algorithm can loose some text elements.
}

