% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup.R
\name{unGroup}
\alias{unGroup}
\title{[!] Mix values of factor variable levels in equal proportions}
\usage{
unGroup(gr, n = NULL, base = "new_group_", new.names = NULL)
}
\arguments{
\item{gr}{A vector indicating original groups.}

\item{n}{Number of new groups. If \code{NULL}, default number is used
which is either \code{nlevels(gr)} or \code{nlevels(gr)+1}, if \code{NA}'s
are presented (see "Details" section).}

\item{base}{Base for names of new groups. Default is \code{"new_group_"}.}

\item{new.names}{Labels for new groups. If value is not \code{NULL},
\code{n} and \code{base} are ignored.}
}
\value{
A vertor with new groups, which are a mix of proportionally
        distributed (ungrouped) original groups.
}
\description{
Mix values of factor variable levels in (approximately) equal proportions
to disrupt existing structure of levels, i.e. "ungroup".
}
\details{
Default number of new groups \code{n} is equal to number of factor levels.
 \code{n = nlevels(gr)}. If NA's exist, they are treated as one
 more additional group thus \code{n = nlevels(gr) + 1}.
}
\examples{
gr <- Scores2$gr

# -------------------------------------------------------
set.seed(1)
newGr <- unGroup(gr)
table(gr,newGr, useNA = "ifany")

# NA's are treated as an additional level ---------------
gr[5:10] <- NA

set.seed(1)
newGr <- unGroup(gr)
table(gr,newGr, useNA = "ifany")

}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{IQR_outliers}},
  \code{\link{expr2text}}, \code{\link{mad_outliers}},
  \code{\link{rmExpr}}, \code{\link{sd_outliers}},
  \code{\link{subt}}, \code{\link{uncall}}
}
\author{
Vilmantas Gegzna
}
\concept{\pkg{spHelper} utilities}
