% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.in.R
\name{which.in}
\alias{which.in}
\alias{which.in.col}
\alias{which.in.diag}
\alias{which.in.offdiag}
\alias{which.in.row}
\alias{which.in.trilow}
\alias{which.in.triupp}
\title{Indices of certain matrix elements}
\usage{
which.in(type, x, ...)

which.in.diag(x)

which.in.offdiag(x)

which.in.col(x, col)

which.in.row(x, row)

which.in.trilow(x, diag = FALSE)

which.in.triupp(x, diag = FALSE)
}
\arguments{
\item{type}{(just for \code{which.in}) Either \code{diag}, \code{offdiag},
\code{col}, \code{row}, \code{trilow} or \code{triupp}.}

\item{x}{A matrix or an object, which can be coerced to a marix.}

\item{...}{Other appropriate parameters, as indicated below.}

\item{col}{An index/Indices of columns that contain elements of interest.}

\item{row}{An index/Indices of rows that contain elements of interest.}

\item{diag}{logical.  Should the diagonal be included?}
}
\value{
Indices of indicated* elements of a matrix (*that belong to either
        diagonal, offdiagonal, upper/lower triange of matrix, certain row(s)
        or column(s)).
}
\description{
Calculate indices of certain (e.g., diagonal) matrix elements.
These elements can be in:
\enumerate{
 \item{diag}{diagonal;}
 \item{offdiag}{offdiagonal;}
 \item{col}{certain columns;}
 \item{row}{certain rows;}
 \item{trilow}{lower triangle;}
 \item{triupp}{upper triangle.}
}
}
\examples{

m1 <- matrix(NA, 5, 5)
which.in.diag(m1)
 ## [1]  1  7 13 19 25

m2 <- matrix(NA, 5, 2)
which.in.diag(m2)
 ## [1]  1  7

#================================

 which.in(diag, m1)

}
\seealso{
Other Matrix operations: \code{\link{corr_vec2mat}},
  \code{\link{nRow2nTri}}, \code{\link{nTri2nRow}}
}

