% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.max.perRow.R
\name{which.max.perRow}
\alias{which.max.perCol}
\alias{which.max.perRow}
\title{[!] Indices of maximum values in each row/column}
\usage{
which.max.perRow(x)

which.max.perCol(x)
}
\arguments{
\item{x}{A matrix}
}
\description{
[!] Indices of maximum values in each row/column
}
\author{
Vilmantas Gegzna
}
\seealso{
Other matrix operations in \pkg{spHelper}: \code{\link{corr_vec2mat}},
  \code{\link{ind.matrix}}, \code{\link{nDiag2nTri}},
  \code{\link{nTri2nDiag}}, \code{\link{new.matrix}},
  \code{\link{sortMaxOnDiag}}, \code{\link{which.in}},
  \code{\link{which.max.all}}
}

