% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_whichOutlier.R
\name{whichOutlier}
\alias{whichOutlier}
\title{[! DEPRECATED] Find indices of rows that contain outlier column-wise outliersscores}
\usage{
whichOutlier(x, zLimit = 2)
}
\arguments{
\item{x}{Matrix-like data.}

\item{zLimit}{A threshold for standardized (scaled) values
(i.e., \href{https://en.wikipedia.org/wiki/Standard_score}{z-scores})
 to be treated as an outlier. If \code{(-zLimit) < scale(scores)} or
       \code{scale(scores) > (+zLimit)} it is treated as an outlier.\cr
      Default \code{zLimit = 2}.}
}
\value{
Vector of indices that indicate rows containing outliers.
}
\description{
[! DEPRECATED] Return indices of rows in scores matrix that contain outliers. A row is
treated as having an oultier if any
\href{http://wiki.eigenvector.com/index.php?title=Advanced_Preprocessing:_Variable_Scaling#Autoscale}{autoscaled}
score in that row is not between \code{±zLimit}.
}
\examples{


whichOutlier(Scores2)
whichOutlier(Scores2,3)

}
\seealso{
Other component analysis / factorisation related functions in \pkg{spHelper}: 
\code{\link{getScores}()},
\code{\link{infoDim}()},
\code{\link{plot_spDiff}()},
\code{\link{qplot_infoDim}()},
\code{\link{qplot_kAmp}()},
\code{\link{qplot_kSp}()},
\code{\link{qplot_spc}()},
\code{\link{reconstructSp}()},
\code{\link{sortLoadings}()},
\code{\link{unipeak}()}
}
\author{
Vilmantas Gegzna
}
\concept{component analysis / factorisation related functions in \pkg{spHelper}}
