% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Fun_Plotting_Sp.R
\name{withSubTitle}
\alias{subt}
\alias{withSubTitle}
\title{Make title with subtitle

Make title with second line as subtitle. Function
uses \code{\link[base]{bquote}} and \code{\link[grDevices]{atop}}.}
\usage{
withSubTitle(Title = NULL, subTitle = NULL)

subt(Title = NULL, subTitle = NULL)
}
\arguments{
\item{Title}{- the first line of title, which will be in bold.}

\item{subTitle}{- the second line of title, which will have smaller
font size and will be in italic.}
}
\value{
Formated title
}
\examples{

identical(subt("Cars"), withSubTitle("Cars"))
## TRUE#'

plot(cars[,1:2], main = "Cars")
plot(cars[,1:2], main = withSubTitle("Cars"))
plot(cars[,1:2], main = subt("Cars")) # the same as previous line
plot(cars[,1:2], main = withSubTitle("Cars","Distance vs. speed"))
plot(cars[,1:2], main = withSubTitle(subTitle = "Distance vs. speed"))


library(ggplot2)
qplot(mpg, wt, data=mtcars) + ggtitle("Cars") # non-bold title
qplot(mpg, wt, data=mtcars) + ggtitle(withSubTitle("Cars")) # bold title
qplot(mpg, wt, data=mtcars) + ggtitle(subt("Cars")) # bold title
qplot(mpg, wt, data=mtcars) + ggtitle(subt("Cars","Distance vs. speed"))
qplot(mpg, wt, data=mtcars) + ggtitle(subt(subTitle = "Distance vs. speed"))

}

