% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_internal_functions.R
\name{x90}
\alias{x90}
\alias{x30}
\title{Rotate x axis tick labels}
\usage{
x90()

x30()
}
\description{
Convenience functions to rotate x axis tick labels in \code{ggplot} plots:\cr
\code{x30()} rotates in 30 degrees.\cr
\code{x90()} rotates in 90 degrees.\cr
}
\details{
These functions correct \code{ggplot2} theme elements and may not respond if
new theme is applied after e.g. \code{x30()} is apllied. I.e. INCORRECT:\cr
\code{qplot(mpg, wt, data = mtcars) + x30() + theme_bw()}.\cr
CORRECT:\cr
\cr \code{qplot(mpg, wt, data = mtcars) + theme_bw() + x30()}.
}
\examples{
library(ggplot2)
library(spHelper)

qplot(mpg, wt, data = mtcars) + x30()
qplot(mpg, wt, data = mtcars) + x90()

}
