% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bru.R
\name{bru}
\alias{bru}
\alias{bru0}
\title{[+] Generate a string of repeated symbols}
\usage{
bru(symbol = "=", n = 60, after = (if (print) 1 else 0), before = 0,
  print = TRUE)

bru0(..., print = FALSE)
}
\arguments{
\item{symbol}{A desired symbol or sequence of symbols. Default is \code{"="}.}

\item{n}{A number of symbols: a length of a string. Default is 60 symbols.}

\item{after}{A number of new (empty) lines/rows to be added afterwards.
\code{0} means that following text continues in the same row.
Default is \cr\code{if (print==TRUE) 1 else 0}.}

\item{before}{A number of peceeding (empty) rows to be added. Default is 0.}

\item{print}{If \code{TRUE} (default for \code{bru}), result is printed to
console using \code{\link[base]{cat}} method.
       If \code{FALSE}, a string is returned.}

\item{...}{(The same parameters as above).}
}
\value{
String of repeated symbols.
}
\description{
Generate a string of repeated symbols.
 Useful for R Markdown (.rmd) files and console output to highlight results.\cr\cr
 \code{bru} prints the result. \cr
 \code{bru0} returns a string.
}
\examples{

bru()
## ============================================================

bru("-")
## ------------------------------------------------------------

bru("= ")
## = = = = = = = = = = = = = = = = = = = = = = = = = = = = = =

bru(12345, 3) # length = 3 symbols
## 123

bru(12345, 6) # length = 6 symbols
## 123451

bru0() # return a string
## [1] "============================================================"

# Several sequences in separate rows
bru(n=5);bru("*",n=5)
## =====
## *****

# Several sequences in one row
bru(n=5, after=0);bru("*",n=5, after=0)
## =====*****

# The same as previous, just short name of "after"
bru(n=5, a=0);bru("*",n=5, a=0)
## =====*****


paste(bru(":", 5),  bru("*", 5))
## :::::
## *****
## character(0)

paste(bru0(":", 5),  bru0("*", 5))
## [1] "::::: *****"

bru0('*', 10, before = 5)
## [1] "\\n\\n\\n\\n\\n**********"

bru('*', 10, before = 5) # 5 empty rows are added
##
##
##
##
##
## **********

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{clear}},
  \code{\link{expr2text}}, \code{\link{makeFirstCapital}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{subt}},
  \code{\link{uncall}}
}

