% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\alias{clear.class}
\alias{clear.except}
\alias{clear.except.class}
\alias{clear.fun}
\title{[!] Clear variables from workspace (global environment)}
\source{
Ideas to create \code{clear} is taken from \code{\link[pracma]{clear}}
        in \pkg{pracma}.
}
\usage{
clear(..., list = NULL, except = NULL)

clear.except(..., list = NULL)

clear.class(clrClass = NULL, except = NULL)

clear.except.class(exceptClass = NULL)

clear.fun()
}
\arguments{
\item{...}{The objects in the global environment as names (unquoted) or
character strings (quoted).}

\item{list}{A character vector naming objects used instead of `\code{...}`.
If \code{list} is not \code{NULL}, `\code{...}` is ignored.}

\item{except}{Names of \bold{variables} (as a character vector) in the global
environment \bold{to be kept} (to be cleared). Default is \code{NULL}.}

\item{clrClass}{Names of classes (as a character vector). Objects of indicated classes
will be removed from workspace (global environment).}

\item{exceptClass}{Names of classes (as a character vector) in the global environment
that has \bold{not} to be cleared. Default is \code{NULL}.}
}
\description{
Remove items from the workspace, i.e. the global environment,
and free up system memory. Explanation in section \bold{"Details"}.
}
\details{
\code{clear} removes either listed or all variables (if none is listed) from
 workspace (global environment). Default is to clear all objects.\cr\cr
\code{clear.except} clears all variables except listed ones.\cr\cr
\code{clear.class} removes objects of indicated class(es), except those which
names provided as argument \code{except}.\cr\cr
\code{clear.except.class} keeps objects of indicated class(es), others are
 cleared.\cr\cr
\code{clear.fun} removes \emph{all} functions only. \cr\cr
}
\examples{

clear()

  A <- 5
  B <- "s"
 D1 <- "string2"
 D2 <- "string3"
  L <- list(A,B)
FUN <- function(x) x
ls()

clear.class("numeric")
ls()

clear.fun()
ls()

clear(except = c("D1", "D2"))
ls()

clear(except = c("B"))
ls()

clear()
ls()

A1 <- 5
A2 <- 5
A3 <- 5
B <- list("a","A")
c <- "ABC"
ls()

clear.except.class(c("numeric", "list"))
ls()

clear.class("numeric", except = "A1")
ls()


}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link[base]{rm}}

Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{expr2text}},
  \code{\link{listAddRm}}, \code{\link{make.filenames}},
  \code{\link{make.firstCapitals}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{st01}},
  \code{\link{subt}}, \code{\link{uncall}}
}

