% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\alias{clear.class}
\alias{clear.fun}
\title{[.] Clear variables from workspace (global environment)}
\source{
Function \code{clear} is copied from \code{\link[pracma]{clear}}
        in \pkg{pracma}.
}
\usage{
clear(lst)

clear.class(className)

clear.fun()
}
\arguments{
\item{lst}{Character vector of names of variables in the global environment.

\code{clear} removes either all ol listed  variables from workspace
              (global environment).\cr
\code{clear.class} removes objects of indicated class(es).\cr
\code{clear.fun} removes functions only.}

\item{className}{Name of class. Objects of this class will be removed
from workspace (global environment).
Remove all items from the workspace, i.e. the global environment,
and freeing up system memory.}
}
\description{
Clear variables from workspace (global environment).
}
\examples{

A <- 5
B <- "s"
L <- list(A,B)
FUN <- function(x) x

clear.class("numeric")
clear.fun()
clear()

}

