% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpFun.R
\name{createFolds_strat}
\alias{createFolds_strat}
\title{Creates folds for k-fold cross-validation stratified by levels in "gr"}
\usage{
createFolds_strat(df_all, k = 5, returnTrain = TRUE)
}
\arguments{
\item{df_all}{- a data frame, that contains variables "ID"  and "gr"

names(df_all)
##   [1] "ID" "gr"}

\item{k}{- number of folds, default k = 5}

\item{returnTrain}{- returnTrain=TRUE}
}
\value{
Result is the same as produced by function \code{\link{caret::createFolds}}
}
\description{
Observations are blocked by "ID" (i.e. observations with the same "ID" are treated as one observation and are always in the same fold)
}
\details{
\code{\link{createFolds_strat}}
}
\examples{


   df_all <- Object$..[,c("ID","gr")]
   nFolds  = 5
   createFolds_strat(df_all, nFolds)

   @export
}

