% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_datasets.R
\docType{data}
\name{DataSet1}
\alias{DataSet1}
\title{[+] Dataset: a data frame for illustrations}
\format{A data frame with 40 rows and 3 variables:
\describe{
  \item{ID}{An identification code.}
  \item{gr}{A factor variable with 4 groups.}
  \item{.row}{A row number.}
}}
\source{
Artificially generated in R.
}
\usage{
DataSet1
}
\description{
A dataset containing example data with variables \code{ID}, \code{gr}
and \code{.row}.
}
\examples{
   DataSet1 <- data.frame(ID = gl(n = 20, k = 2),
                         gr = gl(n = 4, labels = LETTERS[1:4], k = 10))
   DataSet1$.row <- 1:nrow(DataSet1)

   # devtools::use_data(DataSet1)
}
\author{
Vilmantas Gegzna
}
\keyword{datasets}

