% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCol.R
\name{dropCol}
\alias{dropCol}
\title{[!] Remove hyperSpec object columns that contain only \code{NA} values}
\usage{
dropCol(sp)
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}
}
\value{
Object without columns that contain all NA values.
}
\description{
[!] Remove hyperSpec object columns that contain only \code{NA} values
}
\seealso{
Other \code{spHelper} functions for \code{hyperSpec}: \code{\link{addLabels_PAP_PD2014}},
  \code{\link{addLabels_TD2009}}, \code{\link{hy2mat}},
  \code{\link{hyAdd.color}}, \code{\link{read3csv2hy}}
}

