% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropCol.R
\name{dropCol}
\alias{dropCol}
\title{[!] Remove colums of hyperSpec object with all NA values}
\usage{
dropCol(Spectra)
}
\arguments{
\item{Spectra}{- hyperSpec object}
}
\description{
[!] Remove colums of hyperSpec object with all NA values
}
\examples{

 dropCol(Spectra)

}

