% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmExpr.R
\name{expr2text}
\alias{expr2text}
\title{[.] Convert `expression` and `call` to text and remove quotes}
\usage{
expr2text(x)
}
\arguments{
\item{x}{`Call` object or `expression`}
}
\description{
[.] Convert `expression` and `call` to text and remove quotes
}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clear}}, \code{\link{make.firstCapitals}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{subt}},
  \code{\link{uncall}}
}

