% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmExpression.R
\name{expr2text}
\alias{expr2text}
\title{[.] Convert `expression` and `call` to text and remove quotes}
\usage{
expr2text(x)
}
\arguments{
\item{x}{`Call` object or `expression`}
}
\description{
[.] Convert `expression` and `call` to text and remove quotes
}

