% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussAmp.R
\name{GaussAmp}
\alias{GaussAmp}
\title{Generate Gausian curve(s) (GaussAmp)}
\usage{
GaussAmp(x, xc = 0, w = 1, A = 1, y0 = 0)
}
\arguments{
\item{x}{vector of x values}

\item{xc}{vector with centers of Gaussian curves}

\item{w}{vector with parameter w, which determines the width of Gaussian curves}

\item{A}{vector with Amplitudes of Gaussian curves}

\item{y0}{vector with offsets on y axis}
}
\value{
y values of Gaussian curve
}
\description{
Generate Gausian curve(s) (GaussAmp)
}
\note{
The number of curves is determined by maximal length of any of 4
Gausian curve parameters' (xc, w, A, y0) vector. Other parameters are
recycled as shown in example 2 (parameter "A")
}
\examples{
# Example 1
x <- seq(-9.9, 10, 0.2)
y <- GaussAmp(x)

plot(x,y, type = "l", col = "green3"); grid()

# Example 2

require(hyperSpec)

# Make 7 lines
y <- GaussAmp(x, xc = 1:7,A = c(1,2))

dim(y)
##[1]   7 100

Obj <- new("hyperSpec",spc = y,    wavelength = x,
         label = list (spc = "y", .wavelength = "x"))
plot(Obj, col = 1:nrow(Obj)); grid()

}

