% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Fun_Component_Analysis.R
\name{getReconstructed}
\alias{getReconstructed}
\title{[!] Reconstruct spectra from loadings and scores.}
\usage{
getReconstructed(loadings, scores, Sp = NULL)
}
\arguments{
\item{loadings}{??? loadings}

\item{scores}{??? scores}

\item{Sp}{original \code{hyperSpec} object. If \code{Sp} is provided,
the result of this function will be the \code{Sp} in which Sp$spc
will be replaced with \code{reconstructed} spectra.}
}
\value{
\code{reconstructed <- scores \%*\% loadings}
}
\description{
Reconstruct spectra from loadings and scores (i.e. )
}
\examples{
function(loadings, scores)
function(loadings, scores, Sp)

}

