% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScores.R
\name{getScores}
\alias{getScores}
\title{[!] Calculate component amplitudes (a.k.a. scores) by matrix multiplication}
\usage{
getScores(sp, loadings, xLabel = "Component", yLabel = "Amplitude")
}
\arguments{
\item{sp}{- matrix with experimental spectra}

\item{loadings}{- matrix with components' spectra}

\item{xLabel}{- label for loadings for plotting ???}

\item{yLabel}{}
}
\value{
scores - amplitudes of the components
}
\description{
[!] Calculate component amplitudes (a.k.a. scores) by matrix multiplication
}
\details{
\deqn{scores = sp * loadings * inv(loadings' * loadings)}

 formula is taken  and adapted from [1]
}
\examples{
# e.g.:
    sp = Object
    loadings = loadings

getScores(sp, loadings)

}
\references{
[1] M. Brydegaard et al. IEEE Photonics J 2011:3(3);406-21.
}

