% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3 Fun Component Analysis.R
\name{getScores}
\alias{getScores}
\title{[!] Calculate component amplitudes (a.k.a scores) by matrix multiplication}
\usage{
getScores(y, k_sp, xLabel = "Component", yLabel = "Amplitude")
}
\arguments{
\item{y}{- matrix with experimental spectra}

\item{k_sp}{- matrix with components' spectra}

\item{label}{- label for components for plotting ???}
}
\value{
k_amp - amplitudes of the components
}
\description{
[!] Calculate component amplitudes (a.k.a scores) by matrix multiplication
}
\details{
\deqn{k_amp = y * k_sp * inv(k_sp' * k_sp)}

 formula is taken  and adapted from [1]
}
\examples{
# e.g.:
    y = Object
    k_sp = loadings

getScores(y, k_sp)

}
\references{
[1] M. Brydegaard et al. IEEE Photonics J 2011:3(3);406-21.
}

