% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScores.R
\name{getScores}
\alias{getScores}
\title{[!] Calculate component amplitudes (a.k.a. scores) by matrix multiplication}
\usage{
getScores(sp, loadings, xLabel = "Component", yLabel = "Amplitude",
  scores = NULL)
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}

\item{loadings}{Components' spectra (loadings) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{xLabel}{- label that will be used for plotting x axis
\code{labels(scores,".wavelength") <- xLabel}}

\item{yLabel}{- label that will be used for plotting y axis
\code{labels(scores,"spc") <- yLabel}}

\item{scores}{Known scores (do not need to calculate.)}
}
\value{
scores - amplitudes of the components (scores)
}
\description{
[!] Calculate component amplitudes (a.k.a. scores) by matrix multiplication
}
\details{
\deqn{scores = sp * loadings * inv(loadings' * loadings)}

 formula is taken  and adapted from [1]
}
\examples{
# e.g.:
    sp = Object
    loadings = loadings

getScores(sp, loadings)

}
\references{
[1] M. Brydegaard et al. IEEE Photonics J 2011:3(3);406-21.
}

