% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4 Fun for hyperSpec.R
\name{hy2mat}
\alias{hy2mat}
\title{Extract matrix from either hyperSpec object or a matrix}
\usage{
hy2mat(Object)
}
\arguments{
\item{Object}{- object of classes either  \code{\link{hyperSpec}} or  \code{\link{matrix}}}
}
\value{
A matrix or an error if it's impossible to extract a matrix.
}
\description{
hy2mat
}
\examples{
hy2mat(Object)

}

