% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy2mat.R
\name{hy2mat}
\alias{hy2mat}
\title{Extract matrix from either \code{hyperSpec} object or a matrix}
\usage{
hy2mat(sp)
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}
}
\value{
Either a matrix or an error, if it's impossible to extract a matrix.
}
\description{
hy2mat
}
\examples{

data(flu, package = "hyperSpec")

flu
a <- hy2mat(flu)
b <- hy2mat(flu$spc)

identical(a,b)
## [1] TRUE


hy2mat(matrix(NA,5,10))

}
\seealso{
Other \code{spHelper} functions for \code{hyperSpec}: \code{\link{addLabels_PAP_PD2014}},
  \code{\link{addLabels_TD2009}}, \code{\link{dropCol}},
  \code{\link{hyAdd.color}}, \code{\link{read3csv2hy}}
}

