% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd.color.R
\name{hyAdd.color}
\alias{hyAdd.color}
\title{[!] Add a variable with color names to hyperSpec object}
\usage{
hyAdd.color(sp = NULL, by, palette = c("#377EB8", "#4DAF4A", "#984EA3",
  "#FF7F00", "#A65628", "#F781BF", "#999999"))
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}

\item{by}{Variale which levels will correspond to colors in \code{.color}.}

\item{palette}{A color palette (vector with colors for each level in \code{by}).}
}
\value{
HyperSpec object with added/replaced column \code{.color}. Lavel of
       \code{.color} indicate unique names of used colors.
}
\description{
Add (or overwrite, if already exists) a column \code{.color} with color
names that correspond to values of factor variable \code{by}.
}
\examples{

data(Spectra)
Spectra2 <- hyAdd.color(Spectra, "class")

colnames(Spectra)
## [1] "gr"    "class" "spc"

colnames(Spectra2)
## [1] "gr"     "class"  "spc"    ".color"

}

