% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAdd.R
\name{hyAdd}
\alias{hyAdd}
\title{[!] Add a variable to \code{hyperSpec} object}
\usage{
hyAdd(sp, var.name, by, unique.values)
}
\arguments{
\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}

\item{var.name}{The name of variable to be added.}

\item{by}{A factor variale which levels will correspond to values in
\code{unique.values}.}

\item{unique.values}{Unique values used to create the variavle.}
}
\value{
\code{HyperSpec} object with added/replaced variable.
             Lables of the variable indicate unique values used.
}
\description{
Add (or overwrite, if already exists) a variable which name is passed as
 \code{var.name} with values in \code{unique.values}
that correspond to levels of factor variable \code{by}.
}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyDrop.NA}},
  \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

