% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyDrop.NA.R
\name{hyDrop.NA}
\alias{hyDrop.NA}
\title{[+] Remove variables of \code{hyperSpec} object that contain only \code{NA} values}
\usage{
hyDrop.NA(sp)
}
\arguments{
\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}
}
\value{
Object without columns with all \code{NA} values.
}
\description{
Remove variables (columns) of \code{\link[=hyperSpec-class]{hyperSpec}}
object that contain only \code{NA} values.
}
\seealso{
Other \code{spHelper} functions for \code{hyperSpec}: \code{\link{addLabels_PAP_PD2014}},
  \code{\link{addLabels_TD2009}}, \code{\link{hy2mat}},
  \code{\link{hyAdd.color}}, \code{\link{read3csv2hy}}
}

