% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyGet.palette.R
\name{hyGet.palette}
\alias{hyGet.palette}
\title{[+] Get color palette used to create variable'.color'}
\usage{
hyGet.palette(sp)
}
\arguments{
\item{sp}{A \code{hyperSpec} object, that contais variable \code{.color}
added with function \code{\link{hyAdd.color}}}
}
\value{
Color palette used to create variable \code{.color}.
}
\description{
[+] Get color palette used to create variable'.color'
}
\note{
If class of \code{sp} is not \code{hyperSpec}, function returns
      \code{NULL}.
}
\examples{

data(Spectra)
Spectra2 <- hyAdd.color(Spectra, "class")

colnames(Spectra)
   #> [1] "gr"    "class" "spc"

colnames(Spectra2)
   #> [1] "gr"     "class"  "spc"    ".color"

# Names of colors, used for each level of factor variable
labels(Spectra2,".color")
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

hyGet.palette(Spectra2)
   #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

# ATTENTION -------------------------------------------------

# Preserve labels/ color palette (operator ` ]<-`):
Spectra2[1,".color"] <- "red"
hyGet.palette(Spectra2)
    #>  "#377EB8" "#4DAF4A" "#984EA3" "#FF7F00"

# Overwrite labels/ color palette (operator `$<-`):
Spectra2$.color[1] <- "red"
hyGet.palette(Spectra2)
    #>  Warning message:
    #>  In hyGet.palette(Spectra2) : Values of pallete do not exist.
    #>  Most probably they are overwriten by operation `$.color<-`.

labels(Spectra2,".color")
    #>  ".color"

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

