% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoDim.R
\name{infoDim}
\alias{InfoDim}
\alias{infoDim}
\title{Calculate information dimension of a matix}
\usage{
infoDim(Matrix)

InfoDim(Matrix)
}
\arguments{
\item{Matrix}{A matrix with data (rows = observations, columns = variables).}
}
\value{
A list (classes "list" and "infoDim") with fields:
 \describe{
 \item{$dim }{Information dimension, rounded towards positive infinitive;}
 \item{$exactDim  }{Information dimension (fractional, not rounded);}
 \item{$explained  }{A vector of eigenvalues, normalized by sum of eigenvalues,
  which can be used to determine the importance of (principal) components;}
 \item{$eigenvalues }{A vector of eigenvalues;}
 \item{$n.comp }{A vector with integers from 1 to length(eigenvalues).}
}
}
\description{
The function calculates a measure, called "information dimension".
}
\section{WARNING!!!}{

\code{InfoDim} will be removed in future versions. Use \code{infoDim} instead.
}
\examples{
 my_matrix <- matrix(rexp(200, rate=.1), ncol=20)

 my_result <- InfoDim(my_matrix)

 # Investigate the result
 str(my_result)
 my_result$exactDim
 my_result$dim

 #Plot
 my_plot <- plot_infoDim(my_result)
 my_plot

}
\references{
[1]	R. Cangelosi and A. Goriely, Component retention in principal
      component analysis with application to cDNA microarray data.
      Biol Direct, 2, 2 (2007), \url{http://dx.doi.org/10.1186/1745-6150-2-2}
}
\seealso{
Other component analysis / factorisation related functions: \code{\link{getScores}},
  \code{\link{plot_infoDim}}, \code{\link{plot_kAmp}},
  \code{\link{plot_kSp}}, \code{\link{plot_spDiff}},
  \code{\link{reconstructSp}}

Other information dimension functions: \code{\link{InfoDim_plot}},
  \code{\link{plot_infoDim}}
}

