% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoDim.R
\name{InfoDim}
\alias{InfoDim}
\title{Calculate information dimension of a matix.}
\usage{
InfoDim(Matrix)
}
\arguments{
\item{Matrix}{- data matrix (rows = observations, columns = variables)}
}
\value{
A list with fields:
 \itemize{
 \item{$dim }{- information dimension ,rounded towards positive infinitive}
 \item{$exactDim  }{- information dimension (fractional, not rounded)}
 \item{$explained  }{- a vector of eigenvalues, normalized by sum of eigenvalues,
  which can be used to determine the importance of (principal) components}
 \item{$eigenvalues }{- a vector of eigenvalues}
 \item{$n.comp }{- avector with integers from 1 to length(eigenvalues)}
}
}
\description{
The function calculates a measure, called "information dimension".
}
\note{
PrieÅ<U+FFFD> pradedant vykdyti operacijÄ…, svarbus Å¾ingsnis pasirinkti tinkamÄ…
normavimo buda. To nepadarius gausime klaidingÄ… atsakymÄ…. \cr
sp = sp_normuok(sp,x,'1',495);

Taip pat labai svarbus ir triukÅ<U+FFFD>mo lygis. DidÄ—jant triukÅ<U+FFFD>mui atitinkamai
padidinamas maksimalus dimensijÅ³ skaiÄius.

eigenvalues - Tikrines reiksmes / Singular values
pk - tikimybines dimensiju vertes, skirtos entropijos ivertinimui.
explain = pk;
}
\examples{
 my_matrix <- matrix(rexp(200, rate=.1), ncol=20)

 my_result <- InfoDim(my_matrix)

 # Investigate the result
 str(my_result)
 my_result$exactDim
 my_result$dim

 #Plot
 my_plot <- InfoDim_plot(my_result)
 my_plot

}
\references{
[1]	R. Cangelosi and A. Goriely, Component retention in principal
      component analysis with application to cDNA microarray data.
      Biol Direct, 2, 2 (2007), \url{http://dx.doi.org/10.1186/1745-6150-2-2}
}
\seealso{
InfoDim_plot
}

