% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoDim_plot.R
\name{InfoDim_plot}
\alias{InfoDim_plot}
\title{[! ]Plot the result of \code{\link{InfoDim}}}
\usage{
InfoDim_plot(Object, n.comp.SHOW = 20, selected = NULL,
  Title = "Scree Plot")
}
\arguments{
\item{Object}{An object (list), generated by function \code{\link{InfoDim}}}

\item{n.comp.SHOW}{A number of components to show, default is 20. This
number can be corrected if either vector of eigenvalues is smaller than 20
or information dimension is higher than 15.}

\item{selected}{(optional parameter) A number of components sellected
(will be plotted as a separate vertical line).}

\item{Title}{The title of the plot.}
}
\value{
A scree plot: a "lattice" (class "trellis") plot which helps to
        determine the number of nenessary components (e.g. for PCA).
}
\description{
Plot the result of \code{\link{InfoDim}}
}
\note{
http://www.originlab.com/doc%5Cen/Tutorial/images/Principal_Component_Analysis/Pca_scree_plot.png
}
\examples{
 my_matrix <- matrix(rexp(200, rate=.1), ncol=20)

 my_result <- InfoDim(my_matrix)

 # Investigate the result
 str(my_result)
 my_result$exactDim
 my_result$dim

 #Plot
 my_plot <- InfoDim_plot(my_result)
 my_plot

}
\seealso{
Other spHelper plots: \code{\link{plot_SpDiff}},
  \code{\link{plot_confusion}}, \code{\link{plot_kAmp}},
  \code{\link{plot_kSp}}, \code{\link{plot_stat}}
}

