% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2 plotSpFun.R
\name{InfoDim_plot}
\alias{InfoDim_plot}
\title{Plot the result of \code{\link{InfoDim}}}
\usage{
InfoDim_plot(Object, n.comp.SHOW = 20)
}
\arguments{
\item{Object}{- an object (list), generated by function \code{\link{InfoDim}}}

\item{n.comp.SHOW}{- number of components to show, default is 20. This
number can be corrected if either vector of eigenvalues is smaller than 20
or information dimension is higher than 15.}
}
\value{
A plot of class "trellis" which helps to estimate the number of
 nenessary components
}
\description{
Plot the result of \code{\link{InfoDim}}
}
\examples{
 my_matrix <- matrix(rexp(200, rate=.1), ncol=20)

 my_result <- InfoDim(my_matrix)

 # Investigate the result
 str(my_result)
 my_result$exactDim
 my_result$dim

 #Plot
 my_plot <- InfoDim_plot(my_result)
 my_plot

}

