% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Various_functions.R
\name{list.functions}
\alias{list.functions}
\title{[!] List all functions in a package.}
\usage{
list.functions(Package = "spHelper", print.table = TRUE)
}
\arguments{
\item{Package}{- name of package. Default \code{Package = "spHelper"}.}

\item{print.table}{- print the result as a table using
\code{\link[pander]{pander}}.
Default is true.}
}
\value{
A list of functions in a package.
}
\description{
List all functions in a package.
}
\examples{

list.functions()
list.functions(Package = "tidyr", print.table = F)

}

