% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.filenames.R
\name{make.filenames}
\alias{make.filenames}
\title{[+] Convert string into Windows compatible filename}
\usage{
make.filenames(s, allow.space = TRUE)
}
\arguments{
\item{s}{A string.}

\item{allow.space}{Logical. If \code{FALSE}, space symbols will be repalaced
to underscores (\code{_}).}
}
\value{
A string without special symbols.
}
\description{
Function replaces special symbols \emph{\ / : * ? " < > |} in a string \code{s}
 with underscore (\code{_}). After replacement string can be used as a file
 name in Windows.
}
\examples{
 s <- '\\\\ / : * ? " < > |'
 make.filenames(s)
 #> "_ _ _ _ _ _ _ _ _"

 make.filenames(s,allow.space = FALSE)
 #> "_________________"


 s2 <- "Hello?"
 make.filenames(s2)
 #> "Hello_"


}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{clear}},
  \code{\link{expr2text}}, \code{\link{listAddRm}},
  \code{\link{make.firstCapitals}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{st01}},
  \code{\link{subt}}, \code{\link{uncall}}
}

