% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFirstCapital.R
\name{makeFirstCapital}
\alias{fCap}
\alias{makeFirstCapital}
\title{[+] Convert first letters of all words to capitals}
\usage{
makeFirstCapital(x)

fCap(x)
}
\arguments{
\item{x}{Either a string or a vector of strings.}
}
\value{
The same string as input with all words starting in capital letters.
}
\description{
Convert first letters of all words in a string to capitals.
}
\details{
\code{fCap} is a wrapper of \code{makeFirstCapital}.
}
\examples{

makeFirstCapital('laa laa laa')
##[1] "Laa Laa Laa"

fCap('laa laa laa')
##[1] "Laa Laa Laa"

}

