% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5 Various functions.R
\name{makeFirstCapital}
\alias{makeFirstCapital}
\title{Convert the first letter to capital.}
\usage{
makeFirstCapital(x)
}
\arguments{
\item{x}{- a string}
}
\value{
The same string with all words starting with capital letter.
}
\description{
Convert the first letter to capital.
}
\examples{

makeFirstCapital('laa laa laa')

##[1] "Laa Laa Laa"

}

