% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDiag2nTri.R
\name{nDiag2nTri}
\alias{nDiag2nTri}
\title{[+] Matrix: calculate number of elements in a lower/upper triangle}
\usage{
nDiag2nTri(n)
}
\arguments{
\item{n}{A number of diagonal elements in a symmetric matrix.}
}
\value{
A number of elements in a lower/upper triangle of a matrix
        (excluding the diagonal).
}
\description{
Calculate a number of elements in a lower/upper triangle of a symmetric
matrix of size n*n when n is given.
}
\details{
nCoeff = (n^2-n)/2, thus
n = (1+sqrt(1+8*nCoeff))/2
}
\examples{

nDiag2nTri(5)
nDiag2nTri(7)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other matrix operations: \code{\link{corr_vec2mat}},
  \code{\link{ind.matrix}}, \code{\link{nTri2nDiag}},
  \code{\link{new.matrix}}, \code{\link{sortMaxOnDiag}},
  \code{\link{which.in}}, \code{\link{which.max.all}},
  \code{\link{which.max.perRow}}
}

