% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nID_nSp.R
\name{nID_nSp}
\alias{nID_nSp}
\title{[!] Calculate number and percentage of specimens and their spectra.}
\usage{
nID_nSp(data, ID, gr)
}
\arguments{
\item{data}{A data frame with variables, that names are denoted by \code{ID} and \code{gr}}

\item{ID}{...}

\item{gr}{...}
}
\description{
Calculate number and percentage of specimens and their spectra. number(spectra) >= number(specimens)
}
\details{
data     - a data frame with variables denoted by \code{ID} and \code{gr}
ID, gr   - varable names in \code{data} with variables, tad contain IDs and group names (factor levels) respectively.
}
\examples{







nID_nSp(data, ID, gr)
pander::pander(nID_nSp(data, ID, gr))

# For hyperSpec object
nID_nSp(Spectra$.., ID, gr)

}

