% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Fun_for_hyperSpec.R
\name{nID_nSp}
\alias{nID_nSp}
\title{[!] Calculate number and percentage of specimens and their spectra.}
\usage{
nID_nSp(data, ID, gr)
}
\arguments{
\item{data}{- a data frame with variables, that names are denoted by \code{ID} and \code{gr}}

\item{ID}{-...}

\item{gr}{-...}
}
\description{
Calculate number and percentage of specimens and their spectra.
}

