% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nTick.R
\name{nTick}
\alias{nTick}
\alias{nTick_x}
\alias{nTick_y}
\title{[+] Control number of ticks in \code{ggplot2} plots with continuous scale}
\usage{
nTick_x(n = 2, min.n = 2, ...)

nTick_y(n = 2, min.n = 2, ...)
}
\arguments{
\item{n}{An Integer giving the desired number of intervals. Non-integer
values are rounded down.}

\item{min.n}{nonnegative integer giving the minimal number of intervals.
If \code{min.n == 0}, \code{pretty(.)} may return a single value.}

\item{...}{Other parameters to be passed to function \code{pretty}.}
}
\description{
Convenience function to control number of ticks in \code{ggplot2}
 plots with continuous scale.
}
\examples{

# Make a plot but do not print
p <- qplolt_spStat(Spectra, gr, mean, All.linetype = "solid") + facet_grid(.~gr)

#Print the plot
p

# Correct number of ticks:
p + nTick_x(2)
p + nTick_x(2) + nTick_y(8)

}
\author{
Vilmantas Gegzna
}
\seealso{
Pretty breakpoints: \code{\link[base]{pretty}}.

Other \pkg{spHelper} plots: \code{\link{plot_spDiff}},
  \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

