% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hy.palette.R
\name{plot_hy.palette}
\alias{plot_hy.palette}
\title{[!] Plot a color palette used in a hyperSpec object}
\usage{
plot_hy.palette(sp = NULL, by = ".color", palette = hyGet.palette(sp),
  cex = 1.2, Title = if (!is.null(sp)) "Colors of The Groups" else "Colors")
}
\arguments{
\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}

\item{by}{\bold{Either} a name of factor variale in \code{sp} which levels
correspond to colors in \code{palette} \bold{or}, if \code{sp} is
not present, a charter vector of names to be ploted as a text.}

\item{palette}{A list of color names or color codes.}
}
\value{
A plot made with R \code{base} graphics system.
}
\description{
Plot color palette in a \code{\link[=hyperSpec-class]{hyperSpec}} object,
 added with function \code{link{hyAdd.color}}.
}
\examples{

sp <- hyAdd.color(Spectra3,"gr")
plot_hy.palette(sp, "gr")

#--------------------------------------------------
plot_hy.palette(by = "RED color", palette = "#ee0000")

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}

Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_colors}}, \code{\link{plot_spDiff}},
  \code{\link{plot_spFilt}}, \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

