% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2 plotSpFun.R
\name{plot_kAmp}
\alias{plot_kAmp}
\title{Plot component amlitudes (a.k.a scores)}
\usage{
plot_kAmp(scores, Title = "Komponentø amplitudþiø skirstiniai grupëmis")
}
\arguments{
\item{scores}{- object of class \code{\link{hyperSpec}} with scores after factorisation/decomposition}
}
\value{
object of class "ggplot"
}
\description{
Plot component amlitudes (a.k.a scores)
}
\examples{
plot_kAmp(scores)
}

