% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kAmp.R
\name{plot_kAmp}
\alias{plot_kAmp}
\title{[!] Plot component amplitudes (a.k.a scores)}
\usage{
plot_kAmp(scores, Title = "Component amplitudes", subTitle = NULL,
  xLabel = labels(scores, ".wavelength"), yLabel = labels(scores, "spc"),
  by = "gr")
}
\arguments{
\item{scores}{An object of class \code{\link[=hyperSpec-class]{hyperSpec}}
with scores after factorisation/decomposition.}

\item{Title}{The main title for the plot.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{by}{A name of grouping variable. Default is \code{by = "gr"}.}
}
\value{
object of class "ggplot"
}
\description{
Plot component amplitudes (a.k.a scores).
}
\note{
May be incorrect, if cals is not hyperSpec \cr
       xLabel = labels(scores, ".wavelength") \cr
       yLabel
}
\examples{

data(Scores)
plot_kAmp(Scores)

data(Scores3)
plot_kAmp(Scores3, by = "class")

}
\seealso{
Other spHelper plots: \code{\link{InfoDim_plot}},
  \code{\link{plot_SpDiff}}, \code{\link{plot_confusion}},
  \code{\link{plot_kSp}}, \code{\link{plot_stat}}
}

