% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kAmp.R, R/plot_scores.R
\name{plot_kAmp}
\alias{plot_kAmp}
\alias{plot_scores}
\title{[+] Plot amplitudes (a.k.a scores) of spectroscopic components}
\usage{
plot_kAmp(scores, Title = "Component amplitudes", subTitle = NULL,
  xLabel = labels(scores, ".wavelength"), yLabel = labels(scores, "spc"),
  by = "gr", add.violin = TRUE, add.jitter = FALSE, add.boxplot = TRUE,
  violin.alpha = 0.25, jitter.alpha = 0.3, jitter.size = 1)

plot_scores(..., by = NULL, add.violin = FALSE)
}
\arguments{
\item{scores}{An object of class \code{\link[=hyperSpec-class]{hyperSpec}}
with factor scores, principal component scores, component amplitudes, etc.}

\item{Title}{The main title for the plot.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{by}{A name of grouping variable. If \code{NULL}, all data is ploted.
Default is \code{by = "gr"} for \code{plot_kAmp} and
 \code{by = NULL} for \code{plot_scores}.}

\item{add.violin}{Logical. If \code{TRUE} adds so called violin (i.e.,
symmetrical probarility density) plot.
 Default is \code{TRUE} for \code{plot_kAmp}.}

\item{add.jitter}{Logical. If \code{TRUE} adds jitter plot.
Default is \code{FALSE}.}

\item{add.boxplot}{Logical. If \code{TRUE} adds boxplot.
Default is \code{TRUE}.}

\item{violin.alpha}{Transperency of violin plot.}

\item{jitter.alpha}{Transperency of jitter plot.}

\item{jitter.size}{Point size in jitter plot.}

\item{...}{(The same parameters as above).}
}
\value{
A plot.\cr (An object of class \code{ggplot} and
        \code{gg}. More details in package \pkg{ggplot2}).
}
\description{
Plot amplitudes (a.k.a scores) of spectroscopic components
 grouped by categorical variable \code{by}. Amplitudes are provided as
\code{\link[=hyperSpec-class]{hyperSpec}} object.\cr
}
\details{
Plots are drawn in this order: violin plot, jitter plot, boxplot.
}
\note{
If more information on matrix decomposition/factorisation is
needed it can be found
\href{https://en.wikipedia.org/wiki/Matrix_decomposition}{here} or
\href{http://www.r-bloggers.com/matrix-factorization/}{here}).
}
\examples{

data(Scores)
plot_kAmp(Scores)
plot_scores(Scores)

data(Scores3)
plot_kAmp(Scores3, by = "class")

p <- plot_scores(Scores, add.jitter = TRUE)
p

p + theme_bw()


}
\seealso{
Other component analysis / factorisation related functions: \code{\link{getScores}},
  \code{\link{infoDim}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kSp}}, \code{\link{plot_spDiff}},
  \code{\link{reconstructSp}}

Other spHelper plots: \code{\link{nTick}},
  \code{\link{plot_confusion}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kSp}}, \code{\link{plot_spDiff}},
  \code{\link{plot_stat}}, \code{\link{rmExpr}},
  \code{\link{stat_chull}}

Other spHelper plots: \code{\link{nTick}},
  \code{\link{plot_confusion}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kSp}}, \code{\link{plot_spDiff}},
  \code{\link{plot_stat}}, \code{\link{rmExpr}},
  \code{\link{stat_chull}}
}

