% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2 plotSpFun.R
\name{plot_kSp}
\alias{plot_kSp}
\title{Plots spectra of components (a.k.a loadings) on one plot.}
\usage{
plot_kSp(loadings, Title = "Komponentai (kartu)", ylab = "I, sant.vnt")
}
\arguments{
\item{loadings}{- \code{\link{hyperSpec}} object}

\item{Title}{- the title of graph}

\item{ylab}{- label of ordinate axis}
}
\value{
ggplot2 object
}
\description{
Plots spectra of components (a.k.a loadings) on one plot.
}
\examples{
plot_kSp(loadings)

}

