% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kSp.R, R/plot_kSpFacets.R
\name{plot_kSp}
\alias{plot_kSp}
\alias{plot_kSpFacets}
\alias{plot_sp}
\title{[+] Plot spectroscopic curves and spectral components (a.k.a. loadings)}
\usage{
plot_kSp(loadings, Title = "Components", xLabel = labels(loadings,
  ".wavelength"), yLabel = labels(loadings, "spc"), names = "kNames",
  legendName = FALSE, filled = TRUE, normalize = FALSE, Facets = FALSE,
  subTitle = NULL)

plot_sp(..., filled = FALSE)

plot_kSpFacets(..., Title = "Components (stacked)", normalize = "auto",
  Facets = TRUE)
}
\arguments{
\item{loadings}{Components' spectra (loadings) as a \code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{Title}{The main title for the plot.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{names}{A name of variable in \code{loadings} that contains variable names.
If indicated variable does not exist, row names are used instead.\cr
Default is \code{names = 'kNames'}}

\item{legendName}{A name of a legend. Possible entries: \enumerate{
\item {logical \code{FALSE}}{ - a legend without a name;}
\item {logical \code{TRUE}}{ - a label of a variable \code{name} is used as a name of a legend
    (\code{legendName <- labels(loadings,name}));}
\item {...}{manual input of the name.}
}}

\item{filled}{Logical. If \code{TRUE}, colored fill is used. If \code{FALSE}, no fill is used.
(Fill is an area between ordinate axis and the curve.)}

\item{normalize}{A flag that indicates whether components should be
normalized before plotting.
    Possible selections: \enumerate{
    \item {\code{FALSE}, \code{0}}{ - do not normalize;}
    \item {\code{TRUE}, \code{+1}}{ - normalize to max value;}
    \item {\code{-TRUE}, \code{-1}}{ - normalize to min value [! this choise can give unexpected results];}
    \item {\code{"auto"}}{ - one of the choices above is selected by determining if spectra have any
         possitive and any negative peaks:\itemize{
         \item \code{Below0 <- any(loadings$spc < 0);}
         \item \code{Above0 <- any(loadings$spc > 0);}
         \item \code{normalize <- (Above0 - Below0)}
         }
         }
    }}

\item{Facets}{A logical flag. If \code{TRUE}, spectra are plotted on separate graphs/facets
       (implemented by function \code{\link[ggplot2]{facet_grid}}). If {\code{FALSE}, all spectra
        are plotted on one facet.
}}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{...}{(The same parameters as above).}
}
\value{
A plot.\cr (An object of class \code{ggplot} and
        \code{gg}. More details in package \pkg{ggplot2}).
}
\description{
Plot spectroscopic curves in different colors.

             Functions \code{plot_kSp}, \code{plot_kSpFacets} are designed
             to plot spectra of spectral components
            (a.k.a. loadings), extracted in principal component analysis
            (e.g., \code{\link[stats]{princomp}}), variaus matrix factorization
             methods (e.g., \code{\link[NMF]{nmf}})) and some other dimension
             reduction methods. \cr
             The function \code{plot_sp} is convenient for all spectroscopic
             curves as uses no fill.
}
\details{
\code{plot_kSp} plots spectra on one graph. \cr
         \code{plot_kSpFacets} plots spectra on separate graphs (facets).\cr
}
\note{
Before using as an argument \code{loadings},
         a matrix of components/loadings (\code{loadings}) must be treated
         with function \code{\link[hyperSpec]{decomposition}} (or equivalent)
         which converts it to \code{\link[=hyperSpec-class]{hyperSpec}} object.
}
\examples{

plot_kSp(Loadings)

data(flu, package = "hyperSpec")

plot_kSpFacets(flu, Title = "Flu dataset")
plot_kSpFacets(flu, Title = "Flu dataset", normalize = 1)
plot_kSpFacets(flu, Title = "Flu dataset", normalize = FALSE)
plot_kSpFacets(flu, Title = "Flu dataset", normalize = -1)

## Remove fill -----------------------------------------------------------------


plot_kSp(flu, filled = FALSE)
plot_sp(flu)

## Name of a legend ------------------------------------------------------------
flu$c2 <- as.factor(flu$c)

plot_sp(flu, Title = "Flu dataset", names = 'c2', legendName = FALSE)
plot_sp(flu, Title = "Flu dataset", names = 'c2', legendName = TRUE)
plot_sp(flu, Title = "Flu dataset", names = 'c2', legendName = "Concentration")

## Example of line color transitions -------------------------------------------
  plot_sp(laser)

}
\seealso{
Other component analysis / factorisation related functions: \code{\link{getScores}},
  \code{\link{infoDim}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kAmp}}, \code{\link{plot_spDiff}},
  \code{\link{reconstructSp}}

Other spHelper plots: \code{\link{InfoDim_plot}},
  \code{\link{plot_confusion}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kAmp}}, \code{\link{plot_spDiff}},
  \code{\link{plot_stat}}
}

