% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spCompare.R
\name{plot_spCompare}
\alias{plot_spCompare}
\title{[!] Compare 2 spectroscopic signals}
\usage{
plot_spCompare(sp1, sp2, row = 1, colors = c("green4", "blue3", "red"),
  show.legend = TRUE, legend.title = "Signals",
  legend.text = c("Original", "Filtered", "Noise"), lwd = 1)
}
\arguments{
\item{sp1, sp2}{hyperSpec objects of the same size}

\item{row}{An integer, that indicates which row must be plotted.}

\item{colors}{Vector of 3 colors for original, filtered and noise signals
respectively.}

\item{show.legend}{Logical}

\item{legend.title}{The title of the legend.}

\item{legend.text}{Character vertor of length = 3. The entries for the legend
for for original, filtered and noise signals respectively.}

\item{lwd}{A line width. More details in \link[graphics]{par}.}
}
\value{
A plot made with R \code{base} graphics system.
}
\description{
Compare 2 spectroscopic signals: plot 2 signals and their difference.
}
\examples{

# Construct and apply running medians filter
Original <- Spectra
Filtered <- apply(Original, 1, function(x) {runmed(x, 15)})

plot_spCompare(Original, Filtered, row = 2)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_colors}}, \code{\link{plot_hy.palette}},
  \code{\link{plot_spDiff}}, \code{\link{qplolt_spStat}},
  \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

