% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Fun_Plotting_Sp.R
\name{Plot_SpDiff}
\alias{Plot_SpDiff}
\title{[!] Plot difference between experimental and reconstructed spectra}
\usage{
Plot_SpDiff(loadings, scores, Spectra,
  Title = "Difference between experimental and reconstructed spectra",
  spc.nmax = 2000)
}
\arguments{
\item{loadings}{loadings}

\item{scores}{scores}

\item{Spectra}{Spectra (hyperSpec object)}

\item{Title}{Title of the plot.}

\item{spc.nmax}{max number of spectra to plot}
}
\description{
Plot difference between experimental (original) and reconstructed spectra.
Uses function \code{\link{getReconstructed}}, to calculate the reconstructed
spectra and subtracts it from original spectra.
}
\examples{

Plot_SpDiff
}

