% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spDiff.R
\name{plot_spDiff}
\alias{plot_spDiff}
\title{[+] Plot difference between original and reconstructed spectra}
\usage{
plot_spDiff(loadings, scores, sp,
  Title = "Difference between oginal and reconstructed spectra", color = if
  (".color" \%in\% ls(Spectra$..)) sp$.color else "tan3", stacked = if ("ID"
  \%in\% ls(Spectra$..)) sp$ID else NULL, spc.nmax = 2000, ...)
}
\arguments{
\item{loadings}{Components' spectra (loadings) as a \code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{scores}{Components' amplitudes (scores) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{sp}{Spectroscopic data as a \code{\link[=hyperSpec-class]{hyperSpec}}
object.}

\item{Title}{The main title for the plot.}

\item{color}{see \code{\link[graphics]{par} col}. Might be a vector giving
individual colors for the spectra.
\bold{Default} values: vector in \code{sp$.color}, if does not exist,
"tan3" is used as a default color.}

\item{stacked}{if not \code{NULL}, a "stacked" plot is produced.
\code{stacked} may be  \code{TRUE} to stack single spectra. A numeric
 or factor is
  be interpreted as giving the grouping, character is interpreted as
  the name of the extra data column that holds the groups.
  \bold{Default} stacking is by \code{sp$ID}, and \code{NULL} if this
  variable is missing.}

\item{spc.nmax}{maximal number of spectra to be plotted (to avoid
accidentally plotting of large numbers of spectra).}

\item{...}{Other parameters to be passed to function
\code{\link[hyperSpec]{plotspc}}.}
}
\value{
Plot of calculated difference between expected (original)
and reconstructed spectra.
}
\description{
Plot difference between original (e.g., experimental) and
reconstructed spectra. Function \code{\link{reconstructSp}} is used
to reconstruct spectra, that are subtracted from original spectra afrterwards.
}
\note{
Function \code{\link[hyperSpec]{plotspc}} is used to make a plot. R \code{base}
plotting system annotations can be used to enhance the plot.
}
\examples{
plot_spDiff(Loadings, Scores[1:10,,], Spectra[1:10,,], stacked = TRUE)

}
\author{
Vilmantas Gegzna
}
\seealso{
Other component analysis / factorisation related functions: \code{\link{getScores}},
  \code{\link{infoDim}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kAmp}}, \code{\link{plot_kSp}},
  \code{\link{reconstructSp}}, \code{\link{sortLoadings}},
  \code{\link{unipeak}}, \code{\link{whichOutlier}}

Other spHelper plots: \code{\link{nTick}},
  \code{\link{plot_confusion}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kAmp}}, \code{\link{plot_kSp}},
  \code{\link{plot_stat}}, \code{\link{rmExpr}},
  \code{\link{stat_chull}}
}

