% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotConfusion.R
\name{PlotConfusion}
\alias{PlotConfusion}
\title{Plot a confusion matrix, \code{PlotConfusion}}
\usage{
PlotConfusion(conf)
}
\arguments{
\item{conf}{- a table, a square matrix.}

\item{Title}{... /not implemented yet/ Title.}

\item{subTitle}{... /not implemented yet/ Subtitle.}
}
\value{
Plot of confusion matrix. ...
}
\description{
Plot a confusion matrix, \code{PlotConfusion}
}
\examples{
Prediction <- sample(c("A","B","C"),1000, replace = T)
Reference  <- sample(c("A","B","C"),1000, replace = T)
conf <- table(Prediction,Reference)

PlotConfusion(conf)
PlotConfusion(prop.table(conf,1))
PlotConfusion(prop.table(conf,2))

}

