% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDuration.R
\name{printDuration}
\alias{printDuration}
\title{[+] Evaluate how much time passed}
\usage{
printDuration(Start, Message = "Analysis completed in",
  returnString = FALSE)
}
\arguments{
\item{Start}{Moment of time which is treated as a beggining (object of class
\code{\link[=POSIXct-class]{POSIXct}}).}

\item{Message}{Message before time stamp that describes it. Default is
\code{"Duration of analysis:"}}

\item{returnString}{If \code{TRUE}, returns result as a string.
If \code{FALSE} (default), function \code{\link[pander]{pander}} prints the
result.}
}
\value{
Text indicating how much time has passed (either printed in console or as a string).
}
\description{
Evaluate and print difference between moment of time at which time
interval began (usually captured by function \code{\link[base]{Sys.time}} some time ago) and
the present (i.e., moment when function \code{printDuration} is called).
}
\examples{
Start <-  Sys.time()

Start
## [1] "2016-02-12 16:15:09 UTC"

class(Start)
## [1] "POSIXct" "POSIXt"

printDuration(Start)
## Duration of analysis: 23.3 secs

printDuration(Start,"From start till now")
## From start till now 39.2 secs

my_duration <- printDuration(Start)
my_duration
## NULL

my_duration <- printDuration(Start, returnString = TRUE)
my_duration
## Analysis completed in 2.4 mins

}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link[base]{difftime}}

Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clear}}, \code{\link{expr2text}},
  \code{\link{make.firstCapitals}},
  \code{\link{regexp2df}}, \code{\link{rmExpr}},
  \code{\link{subt}}, \code{\link{uncall}}
}

