% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_confusion.R
\name{qplot_confusion}
\alias{qplot_confusion}
\alias{qplot_confusion.default}
\alias{qplot_confusion.matrix}
\alias{qplot_confusion.table}
\title{[!+] Plot a confusion matrix (a.k.a. classification table)}
\usage{
qplot_confusion(...)

\method{qplot_confusion}{default}(Prediction, Reference,
  Title = "Classification table", xLabel = NULL, yLabel = NULL,
  subTitle = NULL, shades = c("prop", "max", "const", "none"),
  guide = FALSE, text.size = 5, decimals = 2)

\method{qplot_confusion}{matrix}(mat, ...)

\method{qplot_confusion}{table}(conf, Title = "Classification table",
  xLabel = NULL, yLabel = NULL, subTitle = NULL, shades = c("prop",
  "max", "const", "none"), guide = FALSE, text.size = 5, decimals = 2)
}
\arguments{
\item{...}{Appropriate parameters (described below).}

\item{Prediction}{A factor variable with \bold{predicted} groups.}

\item{Reference}{A factor variable  with \bold{reference} groups.}

\item{Title}{The main title for the plot.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{shades}{A function how intensities of cell colors in the main confsion
  matrix depend on cell values:
\describe{
    \item{"prop"}{An intensity of a cell color is proportianal
          to a value in the cell. Distribution of color intensities is
           balanced according to number of classes (but not balanced
           according to number of observations per class.)}
    \item{"max"}{Cell with the absolute maximum value is represented by the
    most intensive color. Other cells are proportionally less intensive.
    }
    \item{"const"}{Constant red and green colors.}
    \item{"none"}{All cells are grey.}
}}

\item{guide}{Logical. If \code{TRUE}, a legend is plotted.}

\item{text.size}{The size of a text inside cells.}

\item{decimals}{The number of decimal positions in rounding. Default is 2
(i.e., precission is 0.01).}

\item{conf, mat}{A confusion matrix (classification table): either an object of
a class "table" or a square matrix.}
}
\value{
A plot of confusion matrix and additional statistics (`ggplot` object).
}
\description{
Plot a confusion matrix (classification table) with
             additional statistics (sensitivity (Se) a.k.a. true positive
             rate, positive predictive value (PPV) and Cohens' Kappa (k)). \cr

         Colors in the main matrix: \bold{diagonal} cells vary from grey to
         \bold{green} and \bold{offdiagonal} elements vary from grey to
         \bold{red}. The color intensity in certain cell is determined by
         parameter \code{shades}.\cr

              Colors of cells with statistics Se, PPV and k vary depening on
          values of these statistics from \bold{red} (\emph{low} values) to
          \bold{grey} (\emph{middle} values), to \bold{green} (\emph{high} values).\cr

            \bold{Exception:} when \code{shades="const"} and \code{shades="none"},
         color intensities are constant.
}
\examples{

# Generate data: Random guess  ============================
 N <- 1000 # number of observations

Prediction <- sample(c("A","B","C","D"), N, replace = TRUE)
Reference  <- sample(c("A", "B","C","D"),N, replace = TRUE)

# This function:
qplot_confusion(Prediction,Reference)

# does the same as:
conf <- table(Prediction,Reference)
qplot_confusion(conf)

# At least 50\% of the cases agree =========================
ind <- sample(1:N,round(0.50*N))
Reference[ind] <- Prediction[ind]
conf2 <- table(Prediction,Reference)

qplot_confusion(conf2)

# Most of the cases agree =================================
ind <- sample(1:N,round(N*.80))
Reference[ind] <- Prediction[ind]
conf3 <- table(Prediction,Reference)

qplot_confusion(conf3)

# Proportions =============================================

qplot_confusion(conf3)

# Shades: proportional =====================================

qplot_confusion(conf,shades = "prop",  subTitle = "shades: 'prop', correct by chance")
qplot_confusion(conf,shades = "max",   subTitle = "shades: 'max', correct by chance")

qplot_confusion(conf2,shades = "prop", subTitle = "shades: 'prop', correct >50\%")
qplot_confusion(conf2,shades = "max",  subTitle = "shades: 'max', correct >50\%")

qplot_confusion(conf3,shades = "prop", subTitle = "shades: 'prop', correct >80\%")
qplot_confusion(conf3,shades = "max",  subTitle = "shades: 'max', correct >80\%")

# Shades: constant and none ================================

qplot_confusion(conf3,shades = "const",subTitle = "shades: constant")
qplot_confusion(conf3,shades = "none", subTitle = "shades: none")


}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_colors}}, \code{\link{plot_hy.palette}},
  \code{\link{plot_spCompare}}, \code{\link{plot_spDiff}},
  \code{\link{qplolt_spStat}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

