% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_crosstab.R
\name{qplot_crosstab}
\alias{qplot_crosstab}
\title{[!!!] Plot a crosstabulation (classification table)}
\usage{
qplot_crosstab(tabl, Title = "Cross-tabulation", xLabel = NULL,
  yLabel = NULL, subTitle = NULL, text.size = 5, bgcolor = "wheat2",
  decimals = 2)
}
\arguments{
\item{tabl}{A classification table, crosstabulation: either an object of
a class "table" or a square matrix.}

\item{Title}{The main title for the plot.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{subTitle}{The second line of title, which will be smaller and
and in italics.}

\item{text.size}{The size of a text inside cells.}

\item{bgcolor}{A background color.}

\item{decimals}{The number of decimal positions in rounding. Default is 2
(i.e., precission is 0.01).}
}
\value{
A plot: an object of class \code{\link[ggplot2]{ggplot}}.
       More details in package \pkg{ggplot2}).
}
\description{
Plot a crosstabulation (classification table)
}
\examples{

# Generate data: Random guess  ============================
 N <- 1000 # number of observations

Prediction <- sample(c("A","B","C","D"), N, replace = TRUE)
Reference  <- sample(c("A", "B","C","D"),N, replace = TRUE)

tabl <- table(Prediction,Reference)
qplot_crosstab(tabl)

qplot_crosstab(prop.table(table(Prediction,Reference)))

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_spDiff}}, \code{\link{qplot_confusion}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_prediction}},
  \code{\link{qplot_stat}}, \code{\link{rmExpr}},
  \code{\link{stat_chull}}
}

