% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infoDim.R, R/qplot_infoDim.R, R/qplot_screeplot.R
\name{plot_infoDim}
\alias{plot_infoDim}
\alias{qplot_infoDim}
\alias{qplot_scree}
\alias{qplot_screeplot}
\title{[+] Scree plot with indicated information dimension (ggplot2)}
\usage{
plot_infoDim(data, ...)

qplot_infoDim(Object, n.comp.SHOW = 20, selected = NA,
  Title = "Scree Plot", y.log = TRUE, show.legend = TRUE,
  legend.position = c(0.8, 0.8), ggtheme = theme_bw())

qplot_screeplot(data, ...)

qplot_scree(data, ...)
}
\arguments{
\item{...}{(The same parameters as above).}

\item{Object}{Either a matrix (rows = observations, columns = variables;
 to be passed to \code{infoDim}) or
an object (a list) generated by function \code{\link{infoDim}}.}

\item{n.comp.SHOW}{A number of components to show on x axis, default is 20.
This number is corrected if (a) vector of eigenvalues is smaller than 20
or (b) information dimension is higher than 15.}

\item{selected}{A number of components sellected, will be plotted as a
separate vertical line (optional parameter).}

\item{Title}{The main title of the plot.}

\item{y.log}{Logical. If \code{TRUE} (default) the scale of  y axis is
logarythmic.}

\item{show.legend}{Logical. If \code{TRUE} (default) the legend is displayed.}

\item{legend.position}{the position of the legend ("none", "left", "right",
"bottom", "top", or two-element numeric vector).}

\item{ggtheme}{A function of ggplot2 theme to apply (e.g.:
\link[ggplot2]{ggtheme}). Default is \code{theme_bw()}.}

\item{...}{(The same parameters as above).}

\item{...}{(The same parameters as above).}
}
\value{
A scree plot : plot which helps to
        determine the number of nenessary components (e.g. for PCA).\cr
        (A "ggplot" object.)
}
\description{
Plot a screeplot and with indicated information dimension.
}
\note{
http://www.originlab.com/doc%5Cen/Tutorial/images/Principal_Component_Analysis/Pca_scree_plot.png
}
\examples{
 data(Spectra, package ="spHelper")
 qplot_infoDim(Spectra)

# ------------------------------------------------------
 my_matrix <- matrix(rexp(200, rate=.1), ncol=20)
 my_result <- infoDim(my_matrix)

 # Investigate the result
 str(my_result)
 my_result$exactDim
 my_result$dim

 #Plot
 my_plot <- qplot_infoDim(my_result)
 my_plot


qplot_infoDim(my_matrix)

}
\seealso{
Other component analysis / factorisation related functions: \code{\link{getScores}},
  \code{\link{infoDim}}, \code{\link{plot_kAmp}},
  \code{\link{plot_kSp}}, \code{\link{plot_spDiff}},
  \code{\link{reconstructSp}}

Other information dimension functions: \code{\link{infoDim}}

Other spHelper plots: \code{\link{nTick}},
  \code{\link{plot_confusion}}, \code{\link{plot_kAmp}},
  \code{\link{plot_kSp}}, \code{\link{plot_spDiff}},
  \code{\link{plot_stat}}, \code{\link{rmExpr}},
  \code{\link{stat_chull}}
}

