% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_prediction.R
\name{qplot_prediction}
\alias{qplot_prediction}
\alias{qplot_scatterMDS}
\title{[!!!] Plot a multi-dimensiolally scaled (MDS) scatterplot}
\usage{
qplot_prediction(scores, Prediction, Reference, xLabel = paste("Projection",
  xproj), yLabel = paste("Projection", yproj),
  Title = "Proximity of Groups", subTitle = NULL,
  palette = hyGet.palette(scores), type = c("Prediction", "Reference"),
  stat = c("chull", "ellipse", "none"), type.stat = type,
  MDS = c("metric", "isoMDS"), k = 2, xproj = 1, yproj = 2)

qplot_scatterMDS(scores, by, ...)
}
\arguments{
\item{scores}{Components' amplitudes (scores) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{Prediction}{Predicted groups. Either a variable name or a vector.}

\item{Reference}{Reference grpups. Either a variable name or a vector.}

\item{xLabel}{A label for abscisa (x) axis.}

\item{yLabel}{A label for ordinate (y) axis.}

\item{Title}{The main title for the plot.}

\item{palette}{A color palete to be used. Default is `ggplot2` default palette.}

\item{type}{A type of data for scatterplot: either "prediction" (default)
or "reference".}

\item{stat}{A statistic to be added, either "chull" - convex hull (default),
"ellipse" or "none"}

\item{type.stat}{A type of data for \code{stat}: either  "prediction"
(default) or "reference".}

\item{MDS}{A type of Multi-Dimensional Scaling.}

\item{k}{
  The desired dimension for the solution, passed to \code{cmdscale}.
}

\item{xproj, yproj}{An integer: a number of MDS projection to plot on x, y axis
(respectively).}

\item{by}{A grouping variable}

\item{...}{(The same parameters as above).}
}
\value{
A plot.\cr (An object of class \code{ggplot} and
        \code{gg}. More details in package \pkg{ggplot2}).
}
\description{
/No description yet/ multidimensional scaling is used
}
\examples{

# Example with a `hyperSpec` object:

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} plots: \code{\link{nTick}},
  \code{\link{plot_spDiff}}, \code{\link{qplot_confusion}},
  \code{\link{qplot_crosstab}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{qplot_stat}},
  \code{\link{rmExpr}}, \code{\link{stat_chull}}
}

