% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_resamples.R
\name{qplot_resamples}
\alias{qplot_resamples}
\title{ggplot Function for Visualizing Resampling Results of Models With 1 Tuned Parameter}
\usage{
qplot_resamples(x, data = NULL, metric = x$metric, ...,
  LIMITS = "auto")
}
\arguments{
\item{x}{an object generated by \code{resamples}}

\item{data}{Only used for the \code{ggplot} method; an object generated by
\code{resamples}}

\item{metric}{a character string for which metrics to use as conditioning
variables in the plot. \code{splom} requires exactly one metric when
\code{variables = "models"} and at least two when \code{variables =
"metrics"}.}

\item{...}{further arguments to pass to either
\code{\link[lattice:histogram]{histogram}},
\code{\link[lattice:histogram]{densityplot}},
\code{\link[lattice:xyplot]{xyplot}}, \code{\link[lattice:xyplot]{dotplot}}
or \code{\link[lattice:splom]{splom}}}

\item{LIMITS}{limits for y axis (numeric vector with 2 elements).}
}
\value{
A \code{ggplot2} object.
}
\description{
Function plots a \pkg{ggplot2} plot to visualize resampling results of \pkg{caret} models with 1 tuned parameter.
}
