% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplotStat.R
\name{qplotStat}
\alias{qplotStat}
\title{[!] Plot a summary statistic of spectroscopic data by group.}
\usage{
qplotStat(sp, by = stop("Argument 'by' is missing."),
  FUN = stop("Argument 'FUN' is missing."),
  Title = fCap(as.character(match.call()$FUN)), All = TRUE,
  fixed.colors = TRUE, All.color = "black",
  gr.color = RColorBrewer::brewer.pal(8, "Dark2"), All.linetype = "dashed",
  gr.linetype = "solid", All.size = 1.1, gr.size = 0.8,
  legend.title = element_blank())
}
\arguments{
\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}

\item{by}{by <- as.factor(by)}

\item{FUN}{- ...}

\item{Title}{- ...}

\item{All}{- plot additional statistic of all spectra}

\item{fixed.colors}{- flag if color scheme where ".All" is plotted as
a black line shpuld be used.}

\item{All.color}{- ...}

\item{gr.color}{- ...}

\item{All.linetype}{- ...}

\item{gr.linetype}{- ...}

\item{All.size}{- ...}

\item{gr.size}{- ...}

\item{legend.title}{- ...}
}
\value{
A plot.\cr (An object of class \code{\link[=ggplot2-class]{ggplot}} and
        \code{\link[=ggplot2-class]{gg}}. More details in package \pkg{ggplot2}).
}
\description{
Plot a summary of spectroscopic data in hyperSpec object by group.
The summary statistic internally is generated by either function
\code{\link{spStat}} or \code{\link[hyperSpec]{aggregate}}
}
\examples{


qplotStat(chondro,clusters,mean)
qplotStat(chondro,clusters,mean,All=FALSE)
qplotStat(chondro,clusters,mean_sd,All=FALSE) + facet_grid(.~clusters)

qplotStat(chondro,clusters,median,All=FALSE, fixed.colors=FALSE)
qplotStat(chondro,clusters,median, "My Title")


# Use .aggregate in make facets, to avoid facet called "NA":

qplotStat(chondro,clusters,mean_pm_sd) + facet_grid(.~clusters)
qplotStat(chondro,clusters,mean_pm_sd) + facet_grid(.~.aggregate)

}
\seealso{
\code{\link{spStat}}
}

