% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OceanView.header.R
\name{read.OceanView.header}
\alias{read.OceanView.header}
\title{[!] Read header lines of OceanView file}
\usage{
read.OceanView.header(file = NULL, dec = ".", n = 17,
  indicator = ">>>>>Begin Spectral Data<<<<<", text)
}
\arguments{
\item{file}{The name of the file which header lines are to be read from.}

\item{dec}{The character used in the file for decimal points. Default is
a point (\code{.}).}

\item{n}{Number of lines to be read. Default is 17.}

\item{indicator}{The text, that indicates the last line of the header
lines and the beggining of spectroscopic information.
 Default is \code{">>>>>Begin Spectral Data<<<<<"}.}

\item{text}{Text of header as if it was read with function
\code{\link[base]{readLines}}. If \code{text} is provided, \code{file}
is ignored.}
}
\value{
A list with 2 enties:
\enumerate{
     \item{$data}{A dataframe with information}
     \item{$last_line} {The number of the last header line}
}
}
\description{
Read header lines with non-spectroscopic information of OceanView
(version 1.5) file.
}
\examples{

\donttest{
\dontrun{

 read.OceanView.header("MySpectra.txt")
 read.OceanView.header("MySpectra.txt")$data

}}

## An example of header lines:

#    Data from MySpectra.txt Node
#
#    Date: Fri Jan 15 16:15:16 GMT 2016
#    User: Scientist1
#    Spectrometer: USB2E2222
#    Autoset integration time: false
#    Trigger mode: 4
#    Integration Time (sec): 1.000000E1
#    Scans to average: 1
#    Electric dark correction enabled: true
#    Nonlinearity correction enabled: true
#    Boxcar width: 0
#    XAxis mode: Wavelengths
#    Stop averaging: false
#    Number of Pixels in Spectrum: 2048
#    >>>>>Begin Spectral Data<<<<<

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

