% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OceanView.ts.R
\name{read.OceanView.ts}
\alias{read.OceanView.ts}
\alias{read.OceanView.ts2}
\title{[!v0.1] Read OceanView timeseries spectra}
\usage{
read.OceanView.ts(file, dec = ".", n = 17,
  indicator = ">>>>>Begin Spectral Data<<<<<")

read.OceanView.ts2(file, dec = ",", n = 17,
  indicator = ">>>>>Begin Spectral Data<<<<<")
}
\arguments{
\item{file}{The name of the file which the data are to be read from.}

\item{dec}{The character used in the file for decimal points.}

\item{n}{The number of file lines to scan to search for indicator of
spectral data beginning.}

\item{indicator}{The text, that indicates the last line of the header
lines and the beggining of spectroscopic information.
 Default is \code{">>>>>Begin Spectral Data<<<<<"}.}
}
\value{
A \code{\link[=hyperSpec-class]{hyperSpec}} object with technical and
        spectroscopic information from file \code{file}.
}
\description{
Read contents of spectroscopic timeseries data file to
\code{\link[=hyperSpec-class]{hyperSpec}} object. File is created by
 software `OceanView 1.5`.
}
\details{
In \code{read.OceanView.ts} the default decimal symbol is dot (\code{.}).\cr
         In \code{read.OceanView.ts2} - \emph{comma} (\code{,}).
}
\examples{

\donttest{
\dontrun{

read.OceanView.ts("Spectra.txt")
read.OceanView.ts("Spectra.txt", dec = ",")
read.OceanView.ts2("Spectra.txt")

# Read several files to one `hyperspec` object:

Files   <- dir()[1:4]                         # 4 files are sellected
sp_list <- lapply(Files, read.OceanView.ts2)  # Makes a list of objects
sp      <- collapse(sp_list)                  # Makes one object

plotmat(sp)
}}

}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{plot_hy.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read3csv2hy}}, \code{\link{spStat}}
}

