% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read3csv2hy.R
\name{read3csv2hy}
\alias{read3csv2hy}
\title{[+] Read spectroscopic data from 3 CSV files ("data", "wavelengths" and "spectra") to \code{hyperSpec} object}
\usage{
read3csv2hy(FileName_base)
}
\arguments{
\item{FileName_base}{A base of file name (a string). See section "Details".}
}
\value{
Object of class \code{\link[=hyperSpec-class]{hyperSpec}}.
}
\description{
Read spectroscopic data  ("data", "wavelengths" and
 "spectra") stored in 3 CSV files to hyperSpec object. These fileas are
 usually exported from Matlab.
}
\details{
If base name is "TD_2009", then CSV file names must be:\cr
[1] "TD_2009 (wavelengths).csv"\cr
[2] "TD_2009 (spectra).csv"\cr
[3] "TD_2009 (data).csv"

Base name can be a path, e.g. "D:/spectra/TD_2009"
}
\section{MATLAB code}{


The MATLAB code used to generate csv files:
\href{https://github.com/GegznaV/spHelper/tree/master/inst/doc/Save_DS_for_R.m}{Save_DS_for_R.m}
}
\author{
Vilmantas Gegzna
}
\seealso{
Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}}, \code{\link{spStat}}
}

