% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4 Fun for hyperSpec.R
\name{read3csv2hy}
\alias{read3csv2hy}
\title{Read spectroscopic data in 3 CSV files ("data", "wavelengths" and "spectra") to hyperSpec object}
\usage{
read3csv2hy(FileName_base)
}
\arguments{
\item{FileName_base}{- base of file name (a string). See more in section "Details"}
}
\value{
Object of class \code{\link[hyperSpec]{hyperSpec}}
}
\description{
These fileas are usually exported from Matlab.
}
\details{
If base name is "TD_2009", then CSV files must be:
[1] "TD_2009 (wavelengths).csv"
[2] "TD_2009 (spectra).csv"
[3] "TD_2009 (data).csv"

Base name can be a path, e.g.  "D:/spectra/TD_2009"
}
\note{
Data columns, that have unique values are removed.
}

