% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpFun.R
\name{read3csv2hy}
\alias{read3csv2hy}
\title{Function reads spectra, exported from Matlad to 3 CSV files.
Data columns, that have unique values are removed.}
\usage{
read3csv2hy(FileName_base)
}
\arguments{
\item{FileName_base}{- base of file name (a string). See more in section "Details"}
}
\value{
Object of class \code{\link{hyperSpec}}
}
\description{
Function reads spectra, exported from Matlad to 3 CSV files.
Data columns, that have unique values are removed.
}
\details{
If base name is "TD_2009", then CSV files must be:
[1] "TD_2009 (wavelengths).csv"
[2] "TD_2009 (spectra).csv"
[3] "TD_2009 (data).csv"

Base name can be a path, e.g.  "D:/spectra/TD_2009"
}

