% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructSp.R
\name{reconstructSp}
\alias{getReconstructed}
\alias{reconstructSp}
\title{[!] Reconstruct spectra from loadings and scores.}
\usage{
reconstructSp(loadings, scores, sp = NULL)

getReconstructed(...)
}
\arguments{
\item{loadings}{Components' spectra (loadings) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{scores}{Components' amplitudes (scores) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{sp}{original \code{hyperSpec} object. If \code{sp} is provided,
the result of this function will be the \code{sp} in which sp$spc
will be replaced with \code{reconstructed} spectra.}
}
\value{
\code{reconstructed <- scores \%*\% loadings}
}
\description{
Reconstruct spectra from loadings and scores (i.e. )
}
\examples{
function(loadings, scores)
function(loadings, scores, sp)

}
\seealso{
\code{\link[ChemometricsWithR]{reconstruct}} in \pkg{ChemometricsWithR}, \cr
         \code{\link[ChemometricsWithR]{project}} in \pkg{ChemometricsWithR},\cr
         \code{\link[Rssa]{reconstruct}} in \pkg{Rssa},\cr
         \code{\link[wmtsa]{reconstruct}} in \pkg{wmtsa}\cr
}

