% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructSp.R
\name{reconstructSp}
\alias{getReconstructed}
\alias{reconstructSp}
\title{[!] Reconstruct spectra from loadings and scores.

Reconstruct spectra from loadings and scores (i.e. )}
\usage{
reconstructSp(loadings, scores, sp = NULL)

getReconstructed(...)
}
\arguments{
\item{loadings}{??? loadings}

\item{scores}{??? scores}

\item{sp}{original \code{hyperSpec} object. If \code{sp} is provided,
the result of this function will be the \code{sp} in which sp$spc
will be replaced with \code{reconstructed} spectra.}
}
\value{
\code{reconstructed <- scores \%*\% loadings}
}
\examples{
function(loadings, scores)
function(loadings, scores, sp)

}

