% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmExpression.R
\name{rmExpr}
\alias{rmExpr}
\title{[.] Convert expressions in \code{ggplot2} object labels to strings}
\usage{
rmExpr(p = ggplot2::last_plot())
}
\arguments{
\item{p}{ggplot2 object}
}
\value{
`ggplot2` object
}
\description{
Convert expressions in \code{\link[ggplot2]{ggplot}} object labels to strings, to
enable them to be plotted with \code{\link[plotly]{ggplotly}}.
}
\examples{
data(flu, package = "hyperSpec")

qplotspc(flu)
ggplotly()

rmExpr()
ggplotly()

#--------------------------------
p1 <- plot_sp(Loadings)
p1$labels

p2 <- rmExpr(p1)
p2$labels

ggplotly(p1)
ggplotly(p2)

#----------------------------------
\donttest{
\dontrun{
qplotspc(aggregate(chondro, chondro$clusters, mean_pm_sd),
mapping = aes(x = .wavelength,
              y = spc,
              colour = clusters)) +
    facet_grid(clusters ~ .) + ggtitle("Spectra of Chondrocytes")

 rmExpr()
 ggplotly()

}}

}
\author{
Vilmantas Gegzna
}
\seealso{
Other `spHelper` utilities: \code{\link{clear}},
  \code{\link{expr2text}}, \code{\link{makeFirstCapital}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{subt}}, \code{\link{uncall}}

Other spHelper plots: \code{\link{nTick}},
  \code{\link{plot_confusion}}, \code{\link{plot_infoDim}},
  \code{\link{plot_kAmp}}, \code{\link{plot_kSp}},
  \code{\link{plot_spDiff}}, \code{\link{plot_stat}},
  \code{\link{stat_chull}}
}

