% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCorrVars.R
\name{simCorrVars}
\alias{simCorrVars}
\title{[!!!] Simulate correlated continuous data from standard normal distribution}
\source{
\url{http://www.r-bloggers.com/simulating-random-multivariate-correlated-data-continuous-variables/}
}
\usage{
simCorrVars(corrMat, N = 100)
}
\arguments{
\item{corrMat}{A symmetricmatrix of correlation coefficients.}

\item{N}{A number of observations.}
}
\value{
A dataframe of correlated variables.
}
\description{
Simulate correlated continuaous data from standard normal distribution.
}
\examples{

# ------------------------------------------------------------
simCorrVars()
simCorrVars(N = 10)

# ------------------------------------------------------------
# Generate 2 correlated variables with 10 observations

corrMat <- matrix(c(1, .5, .5, 1), 2)
print(corrMat)
simCorrVars(corrMat, N = 10)

# ------------------------------------------------------------
corrMat <- simCorrMat(3)
simCorrVars(corrMat, N = 10)


}
\seealso{
\link{corrVec2Mat}, \link[MASS]{mvrnorm}
}

