% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpectra.R
\name{simSpectra}
\alias{simSpectra}
\title{[!] Simulate spectroscopic data}
\usage{
simSpectra(x = 300:800, nGr = 2, nInGr = c(50, 50, 50), N = sum(nInGr),
  nDim = 4, w.possile = 5:150, plots = FALSE)
}
\arguments{
\item{x}{Values for x axis ("wavelengths").}

\item{nGr}{Number of groups.}

\item{nInGr}{Number of samples in a group (vector of values for each group).}

\item{N}{Total number of samples.}

\item{nDim}{Number of spectral components (dimensions).}

\item{w.possile}{A vector of possible values of Gaussian curve parameter "w'
(i.e., width). Values of w will be randomly sampled from this vector.}

\item{plots}{Logical. If \code{TRUE} - makes plots. Default is \code{FALSE}.}
}
\value{
List of \code{\link[=hyperSpec-class]{hyperSpec}} objects:
\describe{
     \item{Spectra}{Spectra of each observation, made of \code{(Scores * Loadings) + NOISE}.}
     \item{Loadings}{Normalized spectra of \bold{components}.}
     \item{Scores}{\bold{Amplitudes} of components for each observation.}
}

In the list the \code{hyperSpec} objects contain spectroscopic data and
additional variables.

Additional variables for \code{Spectra} and \code{Scores}:

\describe{
  \item{class}{A factor variable for classification.}
  \item{gr}{A factor variable for other type classification.}
}

Additional variables for \code{Loadings}:
\describe{
  \item{PeakAt}{Position of components top peak.}
  \item{kNames}{Names of components.}
  \item{order.of.rows}{Original order of components before sorting.}
}
}
\description{
Simulation of spectroscopic data: Loadings (spectra of spectroscopic
components), Scores (amplitudes of these components) an Spectra (mixture of
the spectroscopic components multiplied by the amplitudes with random noise
added).
}
\details{
[\emph{This function works, but is not well documented yet.}]
}
\examples{

simSpectra()

simSpectra()$Spectra

}
\author{
Vilmantas Gegzna
}
\seealso{
Other simmulation functions in \pkg{spHelper}: \code{\link{corr_vec2mat}},
  \code{\link{simCorrVars}}
}

