% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Fun_Component_Analysis.R
\name{sortLoadings}
\alias{sortLoadings}
\title{[!] Sort component spectra (a.k.a. Loadings) by possition of top peak}
\usage{
sortLoadings(loadings, Sp = NULL, PCA = FALSE)
}
\arguments{
\item{loadings}{- matrix of loadings (components).}

\item{Sp}{- spectra (object of class \code{\link[=hyperSpec-class]{hyperSpec}})
which will be used to convert sorted loadings into
\code{\link[=hyperSpec-class]{hyperSpec}} object.
More information at \code{\link[=hyperSpec-class]{decomposition}}}

\item{PCA}{- if TRUE, some components are flipped. ... Set to TRUE
if PCA loadings are used. Default \code{PCA = FALSE}}
}
\value{
Either matrix (if \code{Sp} is not provided) or
hyperSpec object with sorted loadings.
In case of hyperSpec object, 3 additional columns
(PeakAt, order.of.rows, kNames) are added.
}
\description{
[!] = [INCOMPLETE DESCRIPTION] Sort component spectra (a.k.a. Loadings) by
 possition of top peak and return
}
\examples{

sortLoadings(loadings)        # returns a matrix

sortLoadings(loadings,Sp)     # returns a hyperSpec object

}

