% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortLoadings.R
\name{sortLoadings}
\alias{sortLoadings}
\title{[+] Sort component spectra (a.k.a. loadings) by of top peak}
\usage{
sortLoadings(loadings, sp = NULL, PCA = FALSE, sort = TRUE)
}
\arguments{
\item{loadings}{Components' spectra (loadings) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{sp}{Spectroscopic data
(either a \code{\link[=hyperSpec-class]{hyperSpec}} object or a matrix).}

\item{PCA}{Logical. If \code{TRUE}, some components are flipped. Set to TRUE
if PCA loadings are used. Default \code{PCA = FALSE}. The flipping follows
the rule:
\deqn{loading_i * (-score_i) = (-loading_i) * score_i}{loading[i] * (-score[i]) = (-loading[i]) * score[i]}
where \eqn{-loading-i}{-loading[i]} represents the i-th flipped loading.}

\item{sort}{Logical. Indicates if returned componenst must be sorted.
If \code{FALSE}, only additional tasks are performed.
Default is \code{TRUE}.}
}
\value{
Either matrix (if \code{sp} is not provided) or
\code{hyperSpec} object with sorted loadings.
In case of \code{hyperSpec} object, 3 additional columns
(PeakAt, order.of.rows, kNames) are added.
}
\description{
Sort rows of \code{hyperSpoec} object by possition of maximum
 value in rows (in other words, sort component spectra, a.k.a. loadings, by
 possition of top peak) and do additional tasks:
 \enumerate{
     \item{If \code{sp} is provided, convert resulting matrix to corresponding
         \code{\link[=hyperSpec-class]{hyperSpec}} object by using function
         \code{\link[hyperSpec]{decomposition}}.}

     \item{If \code{PCA = TRUE}, \code{sp} is provided, and if the mean of
     i-th component's scores is negative, flip the loadings ot that component:
      (\code{sign(mean(Scores_of_component_i)) < 0})
         \code{loadings} and \code{sp} are used to calculate the scores.}
 }
}
\note{
spectra (object of class \code{\link[=hyperSpec-class]{hyperSpec}})
         which will be used to convert sorted loadings into
         \code{\link[=hyperSpec-class]{hyperSpec}} object.
}
\examples{
# ======================================================================

sortLoadings(Loadings[c(2,3,5,1,4),,])

# ======================================================================

loadings <- Loadings[[]] # Exrtract matrix of spectra

L1 <- sortLoadings(loadings)         # returns a matrix
class(L1)
## [1] "matrix"

L2 <- sortLoadings(loadings,Spectra) # returns a hyperSpec object
class(L2)
## [1] "hyperSpec"
# ======================================================================
}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link[hyperSpec]{decomposition}}

Other component analysis / factorisation related functions: \code{\link{getScores}},
  \code{\link{infoDim}}, \code{\link{plot_spDiff}},
  \code{\link{qplot_infoDim}}, \code{\link{qplot_kAmp}},
  \code{\link{qplot_kSp}}, \code{\link{reconstructSp}},
  \code{\link{unipeak}}, \code{\link{whichOutlier}}
}

