% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1 spHelper description.R
\docType{package}
\name{spHelper}
\alias{spHelper}
\alias{spHelper-package}
\title{spHelper: Various functions by V. Gegzna, et al.}
\description{
spHelper: Various functions by V. Gegzna, et al.
}
\section{Functions in spHelper}{

addLabels_TD2009
createFolds_strat
createFolds_stratified
getScores
hy2mat
info_dim
makeFirstCapital
plot_kAmp
plot_kSp
plot_kSpFacets
PlotConfusion
read3csv2hy
regcapturedmatches
}
\examples{
# List all functions in package:

FunctionList <- unclass(lsf.str(envir = asNamespace("spHelper"), all = T))
pander::pander(as.data.frame(FunctionList))

# Other things to remember
devtools::install('D:/Data/R/spHelper')
library(spHelper)


plotc(ObjectName[,,500],model = spc~Integration_time)

}

