% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Fun_for_hyperSpec.R
\name{spStat}
\alias{spStat}
\title{[!] Calculate summary statistic by group and for all data}
\usage{
spStat(Spectra, by, fun = IQR)
}
\arguments{
\item{Spectra}{- hyperSpec object.}

\item{by}{-}

\item{fun}{- function to apply.}
}
\description{
[!] Calculate summary statistic by group and for all data
}
\examples{

spStat(Spectra, by=Spectra$gr, fun = IQR)

}

