% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spStat.R
\name{spStat}
\alias{spStat}
\title{[!] Calculate summary statistic by group and for all data}
\usage{
spStat(sp, by = gr, FUN = mean)
}
\arguments{
\item{sp}{- hyperSpec object.}

\item{by}{- grouping variable (either variable name of \code{sp},
or vector (factor) of length \code{length(sp)}}

\item{FUN}{- function to apply.}
}
\description{
[!] Calculate summary statistic by group and for all data
}
\examples{

spStat(sp, by=sp$gr, FUN = IQR)
spStat(sp, by=gr,    FUN = IQR)

}

