% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spStat.R
\name{spStat}
\alias{spStat}
\title{[+] Calculate summary statistic of spectroscopic data for all spectra and by groups}
\usage{
spStat(sp, by, FUN = mean, Name_of.by = as.character(match.call()$by))
}
\arguments{
\item{sp}{A hyperSpec object.}

\item{by}{A grouping variable (either variable name of \code{sp},
or vector (factor) of length \code{length(sp)}}

\item{FUN}{A function to apply.}

\item{Name_of.by}{(Do not change this!) The name of input \code{by}.}
}
\value{
A hyperSpec object that contains summary statistic. Additional
        column \code{.aggregate} is added. In this column levels of
        grouping variable are indicated.
}
\description{
Calculate summary statistic spectra in \code{hyperSpec} object for all
spectra and by levels of factor variable \code{by}.
}
\note{
This function is similar to \code{\link[hyperSpec]{aggregate}},
just calculates additional statistic for all spectra. This
statistic is labeled by adding level \code{.ALL}) in variable
\code{.aggregate} and variable, that has the same name as \code{by}
(if \code{Name_of.by} si not modified).
}
\examples{
spStat(Spectra, gr)
spStat(Spectra, by = gr,         FUN = IQR)
spStat(Spectra, by = Spectra$gr, FUN = IQR)

}
\author{
Vilmantas Gegzna
}
\seealso{
\code{\link[hyperSpec]{aggregate}}

Other \pkg{spHelper} functions for spectroscopy and \pkg{hyperSpec}: \code{\link{hy2mat}},
  \code{\link{hyAdd.Label.wl}},
  \code{\link{hyAdd.Labels_PAP_PD_2014}},
  \code{\link{hyAdd.Labels_PAP_RK_2014}},
  \code{\link{hyAdd.Labels_TD2009}},
  \code{\link{hyAdd.Labels_TD2015}},
  \code{\link{hyAdd.color}}, \code{\link{hyAdd}},
  \code{\link{hyDrop.NA}}, \code{\link{hyGet.palette}},
  \code{\link{plot_hy.palette}},
  \code{\link{read.OceanView.header}},
  \code{\link{read.OceanView.ts}},
  \code{\link{read3csv2hy}}
}

