% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st01.R
\name{st01}
\alias{st01}
\alias{stMinMax}
\title{[+] Standardize values to intervat from 0 to 1}
\usage{
st01(x, MIN = min(x), MAX = 1)

stMinMax(x, MIN = min(x), MAX = max(x))
}
\arguments{
\item{x}{A numeric vector vector}

\item{MIN}{A value treated as a minimum.}

\item{MAX}{A value treated as a maximum.}
}
\description{
Apply formula: \eqn{(x-MIN)/(MAX-MIN)}{(x-MIN)/(MAX-MIN)}
}
\examples{

x <- seq(.3,1,.1)
st01(x)
st01(x, MIN = .2)

}
\seealso{
Other \pkg{spHelper} utilities: \code{\link{bru}},
  \code{\link{clc}}, \code{\link{clear}},
  \code{\link{expr2text}}, \code{\link{listAddRm}},
  \code{\link{make.filenames}},
  \code{\link{make.firstCapitals}},
  \code{\link{printDuration}}, \code{\link{regexp2df}},
  \code{\link{rmExpr}}, \code{\link{subt}},
  \code{\link{uncall}}
}

