% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_chull.R
\name{stat_chull}
\alias{stat_chull}
\title{[+] A convex hull layer for \code{ggplot2}}
\usage{
stat_chull(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}
or \code{\link[ggplot2]{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data.}

\item{position}{The position adjustment to use for overlapping points on
this layer.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a warning.
If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{...}{Other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}.}
}
\value{
Layer of convex hull for \pkg{ggplot2} plots.
}
\description{
Create a layer of convex hull for \pkg{ggplot2} plots.\cr
Source: vignette \href{https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}{Extending ggplot2}
}
\examples{

ggplot(mpg, aes(displ, hwy)) +
             geom_point() +
             stat_chull(fill = NA, colour = "black")


ggplot(mpg, aes(displ, hwy, colour = drv)) +
             geom_point() +
             stat_chull(fill = NA)


ggplot(mpg, aes(displ, hwy)) +
             stat_chull(geom = "point", size = 4, colour = "red") +
             geom_point()

}
\seealso{
Other spHelper plots: \code{\link{InfoDim_plot}},
  \code{\link{nTick}}, \code{\link{plot_confusion}},
  \code{\link{plot_infoDim}}, \code{\link{plot_kAmp}},
  \code{\link{plot_kSp}}, \code{\link{plot_spDiff}},
  \code{\link{plot_stat}}
}

