% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5 Various functions.R
\name{stratifiedFolds}
\alias{createFolds_stratified}
\alias{stratifiedFolds}
\title{[!] Create stratified and blocked folds for k-fold cross-validation


Create folds that are stratified by levels in variable "gr" in which
observations are blocked by "ID" (i.e. observations with the same "ID"
are treated as one observation and are always in the same fold). \cr

Folds are used for for k-fold cross-validation.}
\usage{
stratifiedFolds(df_all, k = 5, returnTrain = TRUE)

createFolds_stratified(ID = NULL, groups = NULL, k = 5,
  returnTrain = TRUE)
}
\arguments{
\item{df_all}{- a data frame, that contains variables "ID"  and "gr" \cr
\cr
names(df_all) \cr
##   [1] "ID" "gr"}

\item{k}{- number of folds, default k = 5}

\item{returnTrain}{- returnTrain=TRUE}

\item{ID}{- vector}

\item{groups}{- vector}

\item{returnTrain}{- ....returnTrain=TRUE}
}
\value{
Result is a list of folds. In each fold indices observations are
 listed (). (i.e. the structure of outpus is the same as if it was created by
 \code{\link[caret]{createFolds}})
}
\examples{

# Make data with 20 different ID's and 4 different groups:
   df <- data.frame(ID = gl(n = 20, k = 2),
                    gr = gl(n = 4, labels = LETTERS[1:4], k = 10)
                    )

  df

   nFolds = 5

   Folds1 <- createFolds_stratified(df, nFolds)
   Folds2 <- stratifiedFolds(df$ID, df$groups, nFolds)

   @export

}

