% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedFolds.R
\name{stratifiedFolds}
\alias{createFolds2}
\alias{stratifiedFolds}
\title{[+] Create folds with stratification and blocking for k-fold
           cross-validation}
\usage{
stratifiedFolds(data = NULL, gr = NULL, ID = NULL, k = 5,
  returnTrain = TRUE)

createFolds2(..., k = 5)
}
\arguments{
\item{data}{A data frame, that contains variables which names are denoted
by arguments \code{ID} and by \code{gr}.}

\item{gr}{A vector or a name of factor variable in \code{data}, which levels
will be used for \emph{stratification}. E.g., vector with
medical groups.}

\item{ID}{A vector or a name of variable in \code{data}, that contains
identification codes/numbers (ID). These codes will be used bor blocking.}

\item{k}{A number of folds, default k = 5.}

\item{returnTrain}{Logical. If \code{TRUE}, returns indices of variables in training
set. If \code{FALSE}, returns indices of variables in
test set.}

\item{...}{(The same parameters as above).}
}
\value{
A list of folds. In each fold there are indices observations.
        The structure of outpus is the same as if it was created by
        \code{\link[caret]{createFolds}}.
}
\description{
Function randomly devides observations into folds that are used for k-fold
cross-validation. In these folds observations are:
\enumerate{
     \item \bold{blocked} by values in variable \code{ID} (i.e. observations
       with the same "ID" are treated as one unit and are always in the same fold);
     \item \bold{stratified} by levels of factor variable \code{gr} (the proportions of
     these grouped units of observations per each group (level) are kept aproximately
     constant throughout all folds).
 }
}
\note{
If \code{k} is such big, that some folds have no observations of a certain group
      (i.e. level in \code{gr}), an error is returned. In that case smaller value of
       \code{k} is recommended. \cr \cr
        \code{createFolds2} is a wrapper of \code{stratifiedFolds}.
}
\examples{

# Load data
     data("DataSet1")

# Explore data
     str(DataSet1)
     table(DataSet1[,c("gr","ID")])
     summary(DataSet1)

# Explore functions
     nFolds = 5

# If variables of data frame are provided:
     Folds1_a <- stratifiedFolds(data = DataSet1, gr = gr, ID = ID, nFolds, returnTrain=FALSE)
     # str(Folds1_a)
     foldsTets(Folds1_a, DataSet1)

# If "free" variables are provided:
     Folds1_b <- stratifiedFolds(gr = DataSet1$gr, ID = DataSet1$ID, k=nFolds, returnTrain=FALSE)
     # str(Folds1_b)
     foldsTets(Folds1_b, DataSet1)

# Not blocked but stratified
     Folds1_c <- stratifiedFolds(gr = DataSet1$gr, k=nFolds, returnTrain=FALSE)
     # str(Folds1_c)
     foldsTets(Folds1_c, DataSet1)


# Blocked but not stratified
     Folds1_d <- stratifiedFolds(ID = DataSet1$ID, k=nFolds, returnTrain=FALSE)
     # str(Folds1_d)
     foldsTets(Folds1_d, DataSet1)


}
\seealso{
\code{\link[caret]{createFolds}}

Other \code{spHelper} fold creation functions: \code{\link{foldsTets}}

Other \code{spHelper} fold creation functions: \code{\link{foldsTets}}
}

