% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedFolds.R
\name{stratifiedFolds}
\alias{createFolds_strat}
\alias{createFolds_stratified}
\alias{stratifiedFolds}
\title{[~] Block observations and Create stratified folds for k-fold cross-validation.}
\usage{
stratifiedFolds(data = NULL, gr = NULL, ID = NULL, k = 5,
  returnTrain = TRUE)

createFolds_stratified(ID = NULL, gr = NULL, k = 5, returnTrain = TRUE)

createFolds_strat(gr = NULL, ID = NULL, k = 5, returnTrain = TRUE)
}
\arguments{
\item{data}{A data frame, that contains variables denoted by \code{ID} and by \code{gr}.}

\item{gr}{A vector or a name of factor variable in {data}, which levels
will be used for stratification. E.g., vector with medical groups.}

\item{ID}{A vector or a name of variable in {data}, that contains
identification codes/numbers (ID).}

\item{k}{A number of folds, default k = 5.}

\item{returnTrain}{Logical. If \code{TRUE}, returns indices of variables in training
set. If \code{FALSE}, returns indices of variables in
test set.}
}
\value{
A list of folds. In each fold indices observations.
        The structure of outpus is the same as if it was created by
        \code{\link[caret]{createFolds}}.
}
\description{
Function devides observations into folds that are used for k-fold cross-validation. In
these folds observations are:
\enumerate{
     \item \bold{blocked} by values in variable \code{ID} (i.e. observations
       with the same "ID" are treated as one unit and are always in the same fold)
     \item \bold{stratified} by levels of factor variable \code{gr} (the proportions of
     these grouped units of observations per each group (level) are kept aproximately
     constant throughout all folds).
 }
}
\note{
If \code{k} is such big, that some folds have no observations of a certain group
      (i.e. level in \code{gr}), an error is returned. In that case smaller value of
       \code{k} is recommended.
}
\examples{

TestFolds <- function(Folds)lapply(Folds, function(x)table(df[x,"gr"]))


# Make data with 20 different ID's and 4 different groups:
   df <- data.frame(gr = gl(n = 4, labels = LETTERS[1:4], k = 10),
                    ID = gl(n = 20, k = 2))

   table(df)
   summary(df)

   nFolds = 4

# If variables of data frame are provided:
   Folds1_a <- stratifiedFolds(df, gr, ID, nFolds, returnTrain=FALSE)
   str(Folds1_a)
   TestFolds(Folds1_a)
   as.data.frame(TestFolds(Folds1_a)) # <------- this line needs further developement

# If "free" variables are provided:
   Folds1_b <- stratifiedFolds(gr = df$gr, ID = df$ID, k=nFolds, returnTrain=FALSE)
   str(Folds1_b)
   TestFolds(Folds1_b)

# Blocked but not stratified
   Folds1_c <- stratifiedFolds(ID = df$ID, k=nFolds, returnTrain=FALSE)
   str(Folds1_c)
   TestFolds(Folds1_c)

# Not blocked but stratified  ---- ERROR
   Folds1_d <- stratifiedFolds(gr = df$gr, k=nFolds, returnTrain=FALSE)
   str(Folds1_d)
   TestFolds(Folds1_d)

 #  ---- ERROR
   Folds2 <- createFolds_stratified(df$ID, df$gr, nFolds)

 #  ---- ERROR
   Folds3 <- createFolds_strat(df$gr, df$ID, nFolds)

}

