% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_DEPRECATED.R
\name{test_folds_BS}
\alias{test_folds_BS}
\alias{foldTests}
\title{DEPRECATED!}
\usage{
test_folds_BS(
  obj,
  data = NULL,
  stratify_by = NULL,
  block_by = NULL,
  nColShow = 10
)

foldTests(obj, ...)
}
\arguments{
\item{obj}{A list with validation/test set indices in folds.
\bold{Note:} If indices are from training set, the result will be
incorrect.}

\item{data}{A data frame, for wich \code{obj} was created.}

\item{stratify_by}{A name of variable used for stratification.}

\item{block_by}{A name of variable used for blocking.}

\item{nColShow}{Number of blocking variable cross-tabulation's columns to
be shown. Default is 10.}
}
\value{
Print tables that help to evaluate if data is \strong{(a)} stratified,
 \strong{(b)} blocked.
}
\description{
DEPRECATED! Use \code{\link{cvo_test_bs}} instead.\cr
Test if data in folds is stratified and blocked.
This function run tests, that help to evaluate if data in folds is
 (a) stratified and (b) blocked.
}
