% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFolds.R
\name{TestFolds}
\alias{TestFolds}
\title{[+] Evaluate if data in folds is stratified and blocked}
\usage{
TestFolds(FoldsOBJ, DATA, stratBy = "gr", blockBy = "ID")
}
\arguments{
\item{FoldsOBJ}{A list with validation/test set indices in folds.}

\item{DATA}{A data set, for wich \code{FoldsOBJ} was created.}

\item{stratBy}{A name of variable used for stratification.}

\item{blockBy}{A name of variable used for blocking.}
}
\value{
Print tables that help to evaluate if data is \strong{(A)} stratified,
 \strong{(B)} blocked.
}
\description{
Evaluate if data in folds is stratified and blocked.
}
\examples{
data(DataSet1)
FoldsObj <- stratifiedFolds(data = DataSet1, gr = gr, ID = ID, returnTrain=FALSE)
TestFolds(FoldsObj, DataSet1)

}
\seealso{
\code{\link{stratifiedFolds}}
}

