% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Fun_Component_Analysis.R
\name{whichOutlier}
\alias{whichOutlier}
\title{[!] Find row numbers in scores matrix with outliers}
\usage{
whichOutlier(scores, zScore = 2)
}
\arguments{
\item{scores}{??? a matrix or a hyperSpec object of scores (component amplitudes)
after decompositions to components.}

\item{zScore}{... Default zScore = 2.}
}
\value{
Vector of indices indicating rows which contain outliers.
}
\description{
A row is treated as having an oultier if any
\href{autoscaled}{http://wiki.eigenvector.com/index.php?title=Advanced_Preprocessing:_Variable_Scaling#Autoscale}
value of that row is outside ±zScore (**grey** lines in figures below);
}

