% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichOutlier.R
\name{whichOutlier}
\alias{whichOutlier}
\title{[!] Find rows with outlier scores.}
\usage{
whichOutlier(scores, zScore = 2)
}
\arguments{
\item{scores}{Components' amplitudes (scores) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{zScore}{A threshold for standardized values to be treated as outlier.
If \code{-zScore < scale(scores)} or \code{scale(scores) >zScore}
it is treated as an outlier.\cr
Default \code{zScore = 2}.}
}
\value{
Vector of indices indicating rows containing outliers.
}
\description{
Find row numbers in scores matrix that contain outliers \cr\cr
}
\details{
A row is treated as having an oultier if any
\href{autoscaled}{http://wiki.eigenvector.com/index.php?title=Advanced_Preprocessing:_Variable_Scaling#Autoscale}
value of that row is outside ±zScore (**grey** lines in figures below);
}
\examples{

}

