% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichOutlier.R
\name{whichOutlier}
\alias{whichOutlier}
\title{[+] Find indices of rows that contain outlier scores}
\usage{
whichOutlier(scores, zLimit = 2)
}
\arguments{
\item{scores}{Components' amplitudes (scores) as either a matrix or a
\code{\link[=hyperSpec-class]{hyperSpec}} object.}

\item{zLimit}{A threshold for standardized (autoscaled) values
(i.e., \href{https://en.wikipedia.org/wiki/Standard_score}{z-scores})
 to be treated as an outlier. If \code{(-zLimit) < scale(scores)} or
       \code{scale(scores) > (+zLimit)} it is treated as an outlier.\cr
      Default \code{zLimit = 2}.}
}
\value{
Vector of indices that indicate rows containing outliers.
}
\description{
Return indices of rows in scores matrix that contain outliers. A row is
treated as having an oultier if any
\href{http://wiki.eigenvector.com/index.php?title=Advanced_Preprocessing:_Variable_Scaling#Autoscale}{autoscaled}
score in that row is not between \code{±zLimit}.
}
\examples{

whichOutlier(Scores)
whichOutlier(Scores,3)

}

